+++ benches/Cargo.toml
@@ -11,8 +11,8 @@ async-trait = { workspace = true }
 clap = { workspace = true, features = ["derive"] }
 criterion = { version = "0.5", features = ["html_reports", "async", "async_tokio"] }
 ctrlc = "3.2.3"
-digest = "0.10"
-ed25519-dalek = { version = "2.0", features = ["rand_core"] }
+ed25519-dalek = "1.0" # TODO: upgrade to 2.0 when it's released, and remove rand below
+ed25519-dalek_old_rand = { package = "rand", version = "0.7.3" }
 ethnum = "1.3"
 fuel-core = { path = "../crates/fuel-core", default-features = false, features = ["rocksdb-production"] }
 fuel-core-services = { path = "./../crates/services" }
@@ -20,21 +20,14 @@ fuel-core-storage = { path = "./../crates/storage" }
 fuel-core-sync = { path = "./../crates/services/sync", features = ["benchmarking"] }
 fuel-core-types = { path = "./../crates/types", features = ["test-helpers"] }
 futures = "0.3"
-nalgebra = "0.32"
 p256 =  { version = "0.13", default-features = false, features = ["digest", "ecdsa"] }
 rand = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 serde_yaml = "0.9.13"
-sha2 = { version = "0.10" }
-test-case = { workspace = true }
 test-helpers = { path = "../tests/test-helpers" }
 tikv-jemallocator = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
-
-[[bench]]
-harness = false
-name = "generate_test_data"
 
 [[bench]]
 harness = false
