--- benches/src/collect/state.rs
@@ -1,755 +1,0 @@
-use serde_json::Value;
-use std::{
-    collections::{
-        HashMap,
-        HashSet,
-    },
-    fmt::Display,
-    time::Duration,
-};
-
-use crate::{
-    collect::costs::{
-        dependent_cost,
-        Cost,
-        Costs,
-    },
-    GasCostsValues,
-};
-
-fn map_to_ratio(baseline: Duration, mean: Duration) -> u64 {
-    assert!(!baseline.is_zero());
-    let mean = mean.as_nanos();
-    let baseline = baseline.as_nanos();
-    let ratio = mean / baseline;
-    ratio as u64
-}
-
-const TEMPLATE: [&str; 5] = [
-    r##"use super::*;
-"##,
-    r#"
-pub const GIT: &str = ""#,
-    r#"";"#,
-    r##"
-pub fn default_gas_costs() -> GasCostsValues {
-    GasCostsValues {"##,
-    r##"    }
-}
-"##,
-];
-
-#[derive(Debug, Clone)]
-/// The state of the collector.
-pub struct State {
-    /// The baseline to compare against.
-    pub baseline: String,
-    /// Should all measurements be included.
-    pub all: bool,
-    /// Map of ids to their mean times.
-    pub ids: HashMap<String, Duration>,
-    /// Map of ids to their throughput.
-    pub throughput: HashMap<String, u64>,
-    /// Map of groups to their ids.
-    pub groups: HashMap<String, Vec<String>>,
-}
-
-impl Default for State {
-    fn default() -> Self {
-        State {
-            all: false,
-            ids: HashMap::new(),
-            throughput: HashMap::new(),
-            groups: HashMap::new(),
-            baseline: "noop/noop".to_string(),
-        }
-    }
-}
-
-impl Display for State {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let iter = self.ids.iter();
-        match self.ids.get(&self.baseline) {
-            Some(noop) => {
-                for (id, mean) in iter {
-                    writeln!(
-                        f,
-                        "ID: {}, Mean: {:?}, Diff {:?}",
-                        id,
-                        mean,
-                        mean.saturating_sub(*noop)
-                    )?;
-                }
-            }
-            None => {
-                for (id, mean) in iter {
-                    writeln!(f, "ID: {id}, Mean: {mean:?}")?;
-                }
-            }
-        }
-        Ok(())
-    }
-}
-
-impl State {
-    pub fn to_rust_code(&self, yaml: &mut serde_yaml::Value) -> String {
-        let remaps = [
-            ("mod", "mod_op"),
-            ("move", "move_op"),
-            ("ret_contract", "ret"),
-            ("rvrt_contract", "rvrt"),
-            ("retd_contract", "retd"),
-        ];
-        let removes = [("ret_script"), ("rvrt_script"), ("retd_script")]
-            .into_iter()
-            .collect::<HashSet<_>>();
-        let gas_costs = self.to_gas_costs();
-        let mut value = serde_yaml::to_value(gas_costs).unwrap();
-
-        let update_values = |values: &mut serde_yaml::Value| {
-            let m = values.as_mapping_mut().unwrap();
-            for k in &removes {
-                m.remove(k);
-            }
-            for (old, new) in remaps {
-                let v = m.remove(old);
-                if let Some(v) = v {
-                    m.insert(serde_yaml::Value::String(new.to_string()), v);
-                }
-            }
-        };
-
-        update_values(&mut value);
-        update_values(yaml);
-
-        let have = yaml
-            .as_mapping()
-            .unwrap()
-            .keys()
-            .map(|k| k.as_str().unwrap())
-            .collect::<HashSet<_>>();
-        let all_keys = value
-            .as_mapping()
-            .unwrap()
-            .keys()
-            .map(|k| k.as_str().unwrap())
-            .collect::<HashSet<_>>();
-
-        let mut diff = have.symmetric_difference(&all_keys).collect::<Vec<_>>();
-        diff.sort_unstable();
-
-        if !diff.is_empty() {
-            eprintln!("Warning the following keys were not set by this bench:\n{diff:?}\nWas this intentional?");
-        }
-
-        let map = value.as_mapping().unwrap();
-        let indent = "        ";
-        let inner = map
-            .into_iter()
-            .map(|(name, value)| {
-                let name = name.as_str().unwrap();
-                match value {
-                    serde_yaml::Value::Number(i) => {
-                        format!("\n{}{}: {},", indent, name, i.as_u64().unwrap())
-                    }
-                    serde_yaml::Value::Tagged(m) => {
-                        let tag = &m.tag;
-                        let m = &m.value;
-                        if tag.to_string().contains("HeavyOperation") {
-                            let output = format!(
-                                r"
-        {}: DependentCost::HeavyOperation {{
-            base: {},
-            gas_per_unit: {},
-        }},",
-                                name,
-                                m.get("base").unwrap().as_u64().unwrap(),
-                                m.get("gas_per_unit").unwrap().as_u64().unwrap(),
-                            );
-                            output
-                        } else {
-                            let output = format!(
-                                r"
-        {}: DependentCost::LightOperation {{
-            base: {},
-            units_per_gas: {},
-        }},",
-                                name,
-                                m.get("base").unwrap().as_u64().unwrap(),
-                                m.get("units_per_gas").unwrap().as_u64().unwrap(),
-                            );
-                            output
-                        }
-                    }
-                    _ => unreachable!(),
-                }
-            })
-            .collect::<String>();
-        let git_output = std::process::Command::new("git")
-            .arg("rev-parse")
-            .arg("HEAD")
-            .output()
-            .unwrap();
-        let git_string = String::from_utf8_lossy(&git_output.stdout);
-        let generated_by = format!(
-            "/// File generated by fuel-core: {}:{}. With the following git hash",
-            file!(),
-            line!()
-        );
-        format!(
-            "{}{}{}{}{}{}{}\n{}",
-            TEMPLATE[0],
-            generated_by,
-            TEMPLATE[1],
-            git_string.trim(),
-            TEMPLATE[2],
-            TEMPLATE[3],
-            inner,
-            TEMPLATE[4]
-        )
-    }
-
-    pub fn to_gas_costs(&self) -> GasCostsValues {
-        let yaml = self.to_yaml();
-        serde_yaml::from_value(yaml).unwrap()
-    }
-
-    pub fn to_yaml(&self) -> serde_yaml::Value {
-        let v = serde_yaml::to_value(self.clone().into_costs()).unwrap();
-        let v = match v {
-            serde_yaml::Value::Mapping(m) => m,
-            _ => unreachable!(),
-        };
-        let mut v = v.into_iter().collect::<Vec<_>>();
-        v.sort_unstable_by_key(|(k, _)| k.as_str().unwrap().to_string());
-        v.into_iter().collect::<serde_yaml::Mapping>().into()
-    }
-
-    pub fn to_json(&self) -> serde_json::Value {
-        serde_json::to_value(self.clone().into_costs()).unwrap()
-    }
-
-    pub fn into_costs(self) -> Costs {
-        let (state, costs) = self.into_dependent_costs();
-        state.into_relative_costs(costs)
-    }
-
-    pub fn baseline_duration(&self) -> Duration {
-        let nanos = self
-            .ids
-            .get(&self.baseline)
-            .copied()
-            .unwrap_or_else(|| {
-                panic!(
-                    "Could not produce output as baseline {} was not found in recording",
-                    self.baseline
-                )
-            })
-            .as_nanos();
-        let seconds = (nanos / 1_000_000_000) as u64;
-        let nanos = (nanos % 1_000_000_000) as u32;
-        Duration::new(seconds, nanos)
-    }
-
-    pub fn into_dependent_costs(self) -> (Self, Costs) {
-        let baseline = self.baseline_duration();
-        let State {
-            all,
-            mut ids,
-            mut groups,
-            throughput,
-            baseline: baseline_name,
-        } = self;
-
-        let mut costs = Costs::with_capacity(groups.len());
-        let dependent_groups = groups
-            .iter()
-            .filter(|(_, samples)| {
-                samples.iter().any(|sample| throughput.contains_key(sample))
-            })
-            .map(|(name, samples)| {
-                println!("COSTING FOR {}", name);
-                let mut samples = samples
-                    .iter()
-                    .filter_map(|sample| {
-                        let throughput = *throughput.get(sample)?;
-                        let mean = ids.remove(sample)?;
-                        Some((throughput, mean))
-                    })
-                    .map(|(throughput, mean)| {
-                        let ratio = map_to_ratio(baseline, mean);
-                        (throughput as f64, ratio as f64)
-                    })
-                    .collect::<Vec<_>>();
-                samples.sort_unstable_by_key(|(throughput, _)| (*throughput) as u64);
-                dbg!(&samples);
-                (name.clone(), samples)
-            })
-            .collect::<Vec<_>>();
-        let iter = dependent_groups.into_iter().map(|(name, points)| {
-            groups.remove(&name);
-            let dependent_cost = dependent_cost(&name, points)
-                .expect("Unable to calculate dependent cost");
-            (name, Cost::Dependent(dependent_cost))
-        });
-        costs.0.extend(iter);
-
-        let state = Self {
-            all,
-            baseline: baseline_name,
-            ids,
-            throughput,
-            groups,
-        };
-        (state, costs)
-    }
-
-    fn into_relative_costs(self, mut costs: Costs) -> Costs {
-        let baseline = self.baseline_duration();
-        let State {
-            mut ids, groups, ..
-        } = self;
-
-        let relative = groups.into_iter().filter_map(|(name, samples)| {
-            let relative = samples
-                .into_iter()
-                .find(|sample| sample.starts_with(&name) && ids.contains_key(sample))
-                .and_then(|sample| ids.remove(&sample))
-                .map(|mean| Cost::Relative(map_to_ratio(baseline, mean)))?;
-
-            Some((name, relative))
-        });
-        costs.0.extend(relative);
-        costs
-    }
-}
-
-#[derive(Debug)]
-/// Output from a given measurement.
-pub enum Output {
-    /// The mean time of an op measurement.
-    Mean(Mean),
-    /// The ids linked to a group.
-    Group(Group),
-}
-
-#[derive(Debug)]
-/// The mean time of an op measurement.
-pub struct Mean {
-    /// The id of the measurement.
-    id: String,
-    /// The mean time of the measurement.
-    mean: Duration,
-    /// The throughput of the measurement
-    /// if this measurement has one.
-    throughput: Option<u64>,
-}
-
-#[derive(Debug)]
-/// The ids linked to a group.
-pub struct Group {
-    /// The id of the group.
-    id: String,
-    /// The ids of the measurements in the group.
-    benches: Vec<String>,
-}
-
-pub fn extract_state(line: &str, state: &mut State, debug: bool) {
-    match decode_input(line) {
-        Some(Output::Mean(Mean {
-            id,
-            mean,
-            throughput,
-        })) => {
-            if debug {
-                eprintln!("id: {id}, mean: {mean:?}");
-            }
-            state.ids.insert(id.clone(), mean);
-            if let Some(throughput) = throughput {
-                if debug {
-                    eprintln!("throughput: {throughput}");
-                }
-                state.throughput.insert(id, throughput);
-            }
-        }
-        Some(Output::Group(Group { id, benches })) => {
-            state.groups.entry(id.clone()).or_default().extend(benches);
-            eprintln!("{id} complete");
-        }
-        _ => (),
-    }
-}
-
-fn decode_input(line: &str) -> Option<Output> {
-    let val: Value = serde_json::from_str(line).ok()?;
-    match val.get("reason")?.as_str()? {
-        "benchmark-complete" => {
-            let id = match val.get("id") {
-                Some(Value::String(val)) => val.clone(),
-                _ => return None,
-            };
-            let mean = val.get("mean")?;
-            let mean = match mean.get("estimate") {
-                Some(Value::Number(val)) => match mean.get("unit") {
-                    Some(Value::String(unit)) => {
-                        let val = val.as_f64().unwrap();
-                        match unit.as_str() {
-                            "ns" => std::time::Duration::from_nanos(val as u64),
-                            "us" => std::time::Duration::from_micros(val as u64),
-                            "ms" => std::time::Duration::from_millis(val as u64),
-                            "s" => std::time::Duration::from_secs(val as u64),
-                            _ => return None,
-                        }
-                    }
-                    _ => return None,
-                },
-                _ => return None,
-            };
-            let throughput = if let Some(t) = val.get("throughput")?.as_array()?.get(0) {
-                Some(t.as_object()?.get("per_iteration")?.as_u64()?)
-            } else {
-                None
-            };
-            Some(Output::Mean(Mean {
-                id,
-                mean,
-                throughput,
-            }))
-        }
-        "group-complete" => {
-            let id = val.get("group_name")?.as_str()?.to_owned();
-            let benches = val
-                .get("benchmarks")?
-                .as_array()?
-                .iter()
-                .filter_map(|v| Some(v.as_str()?.to_string()))
-                .collect();
-            Some(Output::Group(Group { id, benches }))
-        }
-        _ => None,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::collect::{
-        costs::{
-            Cost,
-            Costs,
-            DependentCost,
-        },
-        state::{
-            extract_state,
-            State,
-        },
-    };
-    use std::path::PathBuf;
-
-    #[test]
-    fn state_into_relative_costs_matches_samples_to_groups() {
-        let input = r#"
-        {"reason":"benchmark-complete","id":"noop/noop","report_directory":"","iteration_count":[20579,41158,61737,82316,102895,123474,144053,164632,185211,205790,226369,246948,267527,288106,308685,329264,349843,370422,391001,411580,432159,452738,473317,493896,514475,535054,555633,576212,596791,617370,637949,658528,679107,699686,720265,740844,761423,782002,802581,823160,843739,864318,884897,905476,926055,946634,967213,987792,1008371,1028950,1049529,1070108,1090687,1111266,1131845,1152424,1173003,1193582,1214161,1234740,1255319,1275898,1296477,1317056,1337635,1358214,1378793,1399372,1419951,1440530,1461109,1481688,1502267,1522846,1543425,1564004,1584583,1605162,1625741,1646320,1666899,1687478,1708057,1728636,1749215,1769794,1790373,1810952,1831531,1852110,1872689,1893268,1913847,1934426,1955005,1975584,1996163,2016742,2037321,2057900],"measured_values":[389282.0,770797.0,1188427.0,1550880.0,1941738.0,2338649.0,2689515.0,3102895.0,3493736.0,3843974.0,4345608.0,4790617.0,5032116.0,5425160.0,5782422.0,6153689.0,6527003.0,6929596.0,7311469.0,7752961.0,8096879.0,8440853.0,8842395.0,9257219.0,10058182.0,10224299.0,10949755.0,11828899.0,12883105.0,13683938.0,16413343.0,13627041.0,13087408.0,13304575.0,13535697.0,13883527.0,15506137.0,14755226.0,15066220.0,15405773.0,15805531.0,16349976.0,16704086.0,16977259.0,17408869.0,17807203.0,18342606.0,18449738.0,18872388.0,19338438.0,19700298.0,20244544.0,20463792.0,21100051.0,21257107.0,21702514.0,22155040.0,22306333.0,22956387.0,23142080.0,23621411.0,23980328.0,24399532.0,24637273.0,25046541.0,25815672.0,25814418.0,26145303.0,26599206.0,26974627.0,28274114.0,27701356.0,28189035.0,28535485.0,28911095.0,29433053.0,29861502.0,30290108.0,30611717.0,30947687.0,31334219.0,31575650.0,32031711.0,32432008.0,32871005.0,33160229.0,33498200.0,34082473.0,34816058.0,34897933.0,35246469.0,35780041.0,36011687.0,36196850.0,36776318.0,36994151.0,37516038.0,37769430.0,38243417.0,38656250.0],"unit":"ns","throughput":[],"typical":{"estimate":18.857469852901115,"lower_bound":18.80991755561779,"upper_bound":18.923223760418672,"unit":"ns"},"mean":{"estimate":19.018693172864804,"lower_bound":18.874689771701124,"upper_bound":19.208667956348382,"unit":"ns"},"median":{"estimate":18.7980051843273,"lower_bound":18.77140984193862,"upper_bound":18.820593660892023,"unit":"ns"},"median_abs_dev":{"estimate":0.10730922897391375,"lower_bound":0.0812325582132701,"upper_bound":0.14041547076097582,"unit":"ns"},"slope":{"estimate":18.857469852901115,"lower_bound":18.80991755561779,"upper_bound":18.923223760418672,"unit":"ns"},"change":{"mean":{"estimate":-0.022826223844162996,"lower_bound":-0.04175747264497742,"upper_bound":-0.005980222776105934,"unit":"%"},"median":{"estimate":-0.009987775466800741,"lower_bound":-0.01278779831658372,"upper_bound":-0.007773882142105615,"unit":"%"},"change":"NoChange"}}
-        {"reason":"group-complete","group_name":"noop","benchmarks":["noop/noop"],"report_directory":""}
-        {"reason":"benchmark-complete","id":"contract_root/test_b","report_directory":"","iteration_count":[1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1],"measured_values":[55969125.0,56143708.0,56070875.0,56446083.0,55284458.0,55575958.0,55562292.0,54611625.0,55169583.0,55756625.0,55473416.0,55187959.0,54961625.0,55105083.0,55144916.0,55879209.0,54749250.0,54745459.0,54731458.0,54612375.0,55069250.0,55551750.0,54944375.0,55457792.0,55276000.0,55580375.0,54754208.0,55046875.0,55313875.0,54756083.0,55104791.0,54734125.0,54853125.0,55141584.0,55071209.0,54997375.0,54770375.0,55104833.0,54911917.0,54924125.0,54923625.0,54617167.0,55281834.0,54639000.0,55284417.0,56533209.0,56687417.0,55747375.0,54843125.0,54792791.0,54881208.0,54618250.0,55585291.0,54643708.0,55165458.0,55089000.0,55331458.0,55029542.0,54938083.0,54980625.0,55133833.0,55131542.0,55607208.0,54999250.0,55162292.0,55075375.0,55621375.0,54975500.0,54729125.0,55682291.0,55181375.0,55166666.0,54747334.0,54719250.0,55561667.0,54923250.0,55619666.0,55118875.0,55347000.0,55339917.0,54804708.0,55304083.0,54754625.0,55773583.0,55411666.0,55134166.0,54958375.0,54873583.0,55044291.0,55179500.0,55161417.0,55459583.0,54935708.0,55161416.0,54972083.0,55446542.0,54918250.0,54687041.0,55506125.0,54776708.0],"unit":"ns","throughput":[],"typical":{"estimate":55182639.51,"lower_bound":55102229.70249999,"upper_bound":55267996.49725,"unit":"ns"},"mean":{"estimate":55182639.51,"lower_bound":55102229.70249999,"upper_bound":55267996.49725,"unit":"ns"},"median":{"estimate":55111979.0,"lower_bound":55022125.0,"upper_bound":55165458.0,"unit":"ns"},"median_abs_dev":{"estimate":350944.01586949825,"lower_bound":269307.6135188341,"upper_bound":506401.31334207935,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"contract_root/test_c","report_directory":"","iteration_count":[1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1],"measured_values":[54881541.0,55167084.0,55069750.0,55097041.0,55038333.0,55082500.0,54983958.0,54867250.0,55324833.0,54954416.0,55007291.0,55436500.0,54699584.0,55743417.0,54767292.0,55069209.0,54756042.0,54812709.0,55476000.0,54692625.0,55201208.0,55067791.0,54872000.0,55866125.0,54966417.0,54780667.0,54879917.0,55085041.0,54680958.0,54851583.0,54747666.0,54879250.0,55303542.0,54986042.0,54796667.0,54376459.0,54536625.0,54535958.0,56011875.0,54531959.0,54355875.0,55293542.0,54396125.0,54930666.0,54472875.0,55089583.0,54641125.0,54417334.0,54606917.0,54407667.0,54895959.0,55105709.0,54491625.0,54695458.0,54567083.0,54873125.0,54544542.0,64412958.0,56095209.0,54300250.0,55058000.0,54617417.0,55095542.0,55600000.0,56537375.0,71663708.0,59887709.0,58745792.0,62623500.0,60810750.0,66510625.0,61822250.0,57435917.0,57221042.0,56591250.0,56781291.0,56302583.0,55933875.0,55793875.0,55166250.0,55924958.0,55800042.0,56612125.0,59934583.0,57355042.0,56426750.0,55144458.0,55987041.0,54776708.0,54834458.0,55490625.0,55277583.0,55151208.0,55024000.0,54754500.0,55358791.0,54915375.0,55410334.0,54941166.0,55122375.0],"unit":"ns","throughput":[],"typical":{"estimate":55889196.25,"lower_bound":55433887.1785,"upper_bound":56437883.57525,"unit":"ns"},"mean":{"estimate":55889196.25,"lower_bound":55433887.1785,"upper_bound":56437883.57525,"unit":"ns"},"median":{"estimate":55069479.5,"lower_bound":54953791.5,"upper_bound":55151208.0,"unit":"ns"},"median_abs_dev":{"estimate":546275.8211016655,"lower_bound":363514.981046319,"upper_bound":765701.3585060835,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"group-complete","group_name":"contract_root","benchmarks":["contract_root/test_a","contract_root/test_b","contract_root/test_c"],"report_directory":""}
-        "#;
-
-        let mut state = State::default();
-        for line in input.lines() {
-            extract_state(line, &mut state, false);
-        }
-        let costs = Costs::default();
-        let costs = state.into_relative_costs(costs);
-
-        assert!(costs.0.contains_key("noop"));
-        assert!(costs.0.contains_key("contract_root"));
-    }
-
-    #[test]
-    fn handles_groups() {
-        let input = r#"
-        {"reason":"group-complete","group_name":"mcli","benchmarks":["mcli/10000","mcli/100000"],"report_directory":""}
-        {"reason":"benchmark-complete","id":"mcp/10000","report_directory":"","iteration_count":[],"measured_values":[],"unit":"ns","throughput":[{"per_iteration":10000,"unit":"bytes"}], "mean":{"estimate":141.6387085050968,"lower_bound":141.05228097712418,"upper_bound":142.32943553585415,"unit":"ns"},"median":{"estimate":140.51177523784355,"lower_bound":140.39754464285716,"upper_bound":140.73739964179842,"unit":"ns"}}
-        {"reason":"benchmark-complete","id":"mcp/100000","report_directory":"","iteration_count":[],"measured_values":[],"unit":"ns","throughput":[{"per_iteration":10000,"unit":"bytes"}], "mean":{"estimate":141.6387085050968,"lower_bound":141.05228097712418,"upper_bound":142.32943553585415,"unit":"ns"},"median":{"estimate":140.51177523784355,"lower_bound":140.39754464285716,"upper_bound":140.73739964179842,"unit":"ns"}}
-        {"reason":"group-complete","group_name":"mcp","benchmarks":["mcp/10000","mcp/100000"],"report_directory":""}
-        {"reason":"benchmark-complete","id":"mcpi/10000","report_directory":"","iteration_count":[],"measured_values":[],"unit":"ns","throughput":[{"per_iteration":10000,"unit":"bytes"}], "mean":{"estimate":141.6387085050968,"lower_bound":141.05228097712418,"upper_bound":142.32943553585415,"unit":"ns"},"median":{"estimate":140.51177523784355,"lower_bound":140.39754464285716,"upper_bound":140.73739964179842,"unit":"ns"}}
-        "#;
-
-        let mut state = State {
-            all: true,
-            baseline: "".into(),
-            ids: Default::default(),
-            throughput: Default::default(),
-            groups: Default::default(),
-        };
-        for line in input.lines() {
-            extract_state(line, &mut state, false);
-        }
-
-        assert_eq!(state.groups.len(), 2);
-        assert_eq!(
-            state.groups.get("mcli").unwrap(),
-            &["mcli/10000", "mcli/100000"]
-        );
-        assert_eq!(
-            state.groups.get("mcp").unwrap(),
-            &["mcp/10000", "mcp/100000"]
-        );
-        assert_eq!(state.ids.len(), 3);
-        assert!(state.ids.contains_key("mcp/10000"));
-        assert!(state.ids.contains_key("mcp/100000"));
-        assert!(state.ids.contains_key("mcpi/10000"));
-    }
-
-    #[test]
-    fn serialize_gas_costs() {
-        let input = r#"
-        {"reason":"benchmark-complete","id":"noop","report_directory":"","iteration_count":[],"measured_values":[],"unit":"ns","throughput":[],"mean":{"estimate":14.6387085050968,"lower_bound":141.05228097712418,"upper_bound":142.32943553585415,"unit":"ns"},"median":{"estimate":140.51177523784355,"lower_bound":140.39754464285716,"upper_bound":140.73739964179842,"unit":"ns"}}
-        {"reason":"group-complete","group_name":"noop","benchmarks":["noop"],"report_directory":""}
-        {"reason":"benchmark-complete","id":"add","report_directory":"","iteration_count":[],"measured_values":[],"unit":"ns","throughput":[],"mean":{"estimate":35.6387085050968,"lower_bound":141.05228097712418,"upper_bound":142.32943553585415,"unit":"ns"},"median":{"estimate":140.51177523784355,"lower_bound":140.39754464285716,"upper_bound":140.73739964179842,"unit":"ns"}}
-        {"reason":"group-complete","group_name":"add","benchmarks":["add"],"report_directory":""}
-        {"reason":"group-complete","group_name":"mcli","benchmarks":["mcli/10000","mcli/100000"],"report_directory":""}
-        {"reason":"benchmark-complete","id":"mcp/10000","report_directory":"","iteration_count":[],"measured_values":[],"unit":"ns","throughput":[{"per_iteration":10000,"unit":"bytes"}], "mean":{"estimate":141.6387085050968,"lower_bound":141.05228097712418,"upper_bound":142.32943553585415,"unit":"ns"},"median":{"estimate":140.51177523784355,"lower_bound":140.39754464285716,"upper_bound":140.73739964179842,"unit":"ns"}}
-        {"reason":"benchmark-complete","id":"mcp/100000","report_directory":"","iteration_count":[],"measured_values":[],"unit":"ns","throughput":[{"per_iteration":100000,"unit":"bytes"}], "mean":{"estimate":1.6387085050968,"lower_bound":141.05228097712418,"upper_bound":142.32943553585415,"unit":"us"},"median":{"estimate":140.51177523784355,"lower_bound":140.39754464285716,"upper_bound":140.73739964179842,"unit":"ns"}}
-        {"reason":"group-complete","group_name":"mcp","benchmarks":["mcp/10000","mcp/100000"],"report_directory":""}
-        {"reason":"benchmark-complete","id":"mcpi/10000","report_directory":"","iteration_count":[],"measured_values":[],"unit":"ns","throughput":[{"per_iteration":10000,"unit":"bytes"}], "mean":{"estimate":141.6387085050968,"lower_bound":141.05228097712418,"upper_bound":142.32943553585415,"unit":"ns"},"median":{"estimate":140.51177523784355,"lower_bound":140.39754464285716,"upper_bound":140.73739964179842,"unit":"ns"}}
-        "#;
-
-        let mut state = State {
-            all: false,
-            baseline: "noop".into(),
-            ids: Default::default(),
-            throughput: Default::default(),
-            groups: Default::default(),
-        };
-        for line in input.lines() {
-            extract_state(line, &mut state, false);
-        }
-
-        let costs = state.to_gas_costs();
-
-        let ser = serde_yaml::to_string(&costs).unwrap();
-        eprintln!("{ser}");
-        eprintln!("{}", state.to_rust_code());
-        let path = PathBuf::from(concat!(
-            env!("CARGO_MANIFEST_DIR"),
-            "/benches-outputs/src/test_gas_costs_output.rs"
-        ));
-        let manifest_path = PathBuf::from(concat!(
-            env!("CARGO_MANIFEST_DIR"),
-            "/benches-outputs/Cargo.toml"
-        ));
-        std::fs::write(path, state.to_rust_code()).unwrap();
-        let output = std::process::Command::new("cargo")
-            .arg("check")
-            .arg("--manifest-path")
-            .arg(&manifest_path)
-            .output()
-            .unwrap();
-        println!("{}", String::from_utf8(output.stderr).unwrap());
-
-        assert!(output.status.success());
-    }
-
-    #[test]
-    fn categorizes_dependent_cost_from_basic_linear_function() {
-        // f(x) = x
-        let input = r#"
-        {"reason":"benchmark-complete","id":"baseline/baseline","report_directory":"","iteration_count":[400,800,1200,1600,2000,2400,2800,3200,3600,4000,4400,4800,5200,5600,6000,6400,6800,7200,7600,8000,8400,8800,9200,9600,10000,10400,10800,11200,11600,12000,12400,12800,13200,13600,14000,14400,14800,15200,15600,16000,16400,16800,17200,17600,18000,18400,18800,19200,19600,20000,20400,20800,21200,21600,22000,22400,22800,23200,23600,24000,24400,24800,25200,25600,26000,26400,26800,27200,27600,28000,28400,28800,29200,29600,30000,30400,30800,31200,31600,32000,32400,32800,33200,33600,34000,34400,34800,35200,35600,36000,36400,36800,37200,37600,38000,38400,38800,39200,39600,40000],"measured_values":[980375.0,2042583.0,2925708.0,3958209.0,4882042.0,6005750.0,7051542.0,8004542.0,8863250.0,9869000.0,10931500.0,11793625.0,12876458.0,13708167.0,14900000.0,15812292.0,16836750.0,17853833.0,18615583.0,19836500.0,20811334.0,21645541.0,22595958.0,23913208.0,24400125.0,25580500.0,26418375.0,27358458.0,28630917.0,29607541.0,30691208.0,31750417.0,32395208.0,33475125.0,34481000.0,35254041.0,36413917.0,37827125.0,38315125.0,39194958.0,40272959.0,41583625.0,42374875.0,43696542.0,44411167.0,45749958.0,46469667.0,46900250.0,47975417.0,48853791.0,50214458.0,50994375.0,51728708.0,52615792.0,54259250.0,55057542.0,56167875.0,57079125.0,58734875.0,59167417.0,59923167.0,61056083.0,62263458.0,63549000.0,63386084.0,64466000.0,65242375.0,66604208.0,67392667.0,68778375.0,69675459.0,70643459.0,71502000.0,72831917.0,73311750.0,74390875.0,75552708.0,76335333.0,77090958.0,78726875.0,78991000.0,80502875.0,81527375.0,82604250.0,83732334.0,83937500.0,85044208.0,85874250.0,86862375.0,87486417.0,88731000.0,90355041.0,91253709.0,96148875.0,93421584.0,96341167.0,99255916.0,96891042.0,97329833.0,98138667.0],"unit":"ns","throughput":[],"typical":{"estimate":2460.573495847495,"lower_bound":2453.0142813510597,"upper_bound":2469.7349639561435,"unit":"ns"},"mean":{"estimate":2462.9766883009474,"lower_bound":2458.5303726606207,"upper_bound":2467.8023819017535,"unit":"ns"},"median":{"estimate":2458.1929017857146,"lower_bound":2454.9998852115777,"upper_bound":2461.493039215686,"unit":"ns"},"median_abs_dev":{"estimate":17.49073729043944,"lower_bound":12.455655670850211,"upper_bound":23.15857114984513,"unit":"ns"},"slope":{"estimate":2460.573495847495,"lower_bound":2453.0142813510597,"upper_bound":2469.7349639561435,"unit":"ns"},"change":{"mean":{"estimate":0.00034906776017207264,"lower_bound":-0.0035474363915021664,"upper_bound":0.0035938220287743405,"unit":"%"},"median":{"estimate":-0.0003874381483204292,"lower_bound":-0.002575878137824894,"upper_bound":0.0028194699120156486,"unit":"%"},"change":"NoChange"}}
-        {"reason":"group-complete","group_name":"baseline","benchmarks":["baseline/baseline"],"report_directory":""}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_1","report_directory":"","iteration_count":[404,808,1212,1616,2020,2424,2828,3232,3636,4040,4444,4848,5252,5656,6060,6464,6868,7272,7676,8080,8484,8888,9292,9696,10100,10504,10908,11312,11716,12120,12524,12928,13332,13736,14140,14544,14948,15352,15756,16160,16564,16968,17372,17776,18180,18584,18988,19392,19796,20200,20604,21008,21412,21816,22220,22624,23028,23432,23836,24240,24644,25048,25452,25856,26260,26664,27068,27472,27876,28280,28684,29088,29492,29896,30300,30704,31108,31512,31916,32320,32724,33128,33532,33936,34340,34744,35148,35552,35956,36360,36764,37168,37572,37976,38380,38784,39188,39592,39996,40400],"measured_values":[982625.0,1962750.0,3056750.0,3878500.0,4900833.0,5906375.0,7068084.0,7778333.0,8855916.0,9864542.0,10830416.0,11778291.0,12684208.0,13682792.0,14723625.0,15647250.0,16620208.0,17687208.0,18565250.0,19800458.0,20556625.0,21749500.0,22460958.0,23540125.0,24667208.0,25730667.0,26610833.0,27684917.0,28866750.0,29816625.0,30520292.0,31353917.0,32345375.0,33598959.0,34388750.0,35356292.0,36269334.0,37132625.0,38577250.0,39511958.0,41328375.0,42390166.0,42448625.0,43253083.0,44498500.0,44968708.0,46254834.0,47286334.0,48481542.0,49672583.0,50394583.0,51322709.0,52898833.0,53938917.0,54724875.0,55479458.0,56585375.0,57159333.0,59187209.0,59871208.0,60114375.0,61188583.0,62551833.0,63204084.0,64119458.0,65315708.0,66523917.0,67395417.0,68422333.0,69732625.0,69798084.0,70829125.0,72133500.0,73564750.0,74269417.0,75034375.0,76336042.0,76994083.0,77928750.0,79524750.0,80204416.0,80494375.0,81473958.0,82774375.0,84358875.0,85192250.0,86225459.0,86753542.0,87881375.0,88969625.0,90067959.0,93574708.0,93004500.0,93534750.0,94269417.0,95573667.0,96114625.0,103981667.0,99499959.0,99921250.0],"unit":"ns","throughput":[{"per_iteration":1,"unit":"elements"}],"typical":{"estimate":2458.6583068604627,"lower_bound":2450.1344637939987,"upper_bound":2470.406228677048,"unit":"ns"},"mean":{"estimate":2448.5780889488697,"lower_bound":2443.5590106287455,"upper_bound":2454.3517952404422,"unit":"ns"},"median":{"estimate":2444.7558941831685,"lower_bound":2441.7005859130218,"upper_bound":2449.057486360881,"unit":"ns"},"median_abs_dev":{"estimate":18.01958546524158,"lower_bound":12.77885577602741,"upper_bound":21.761071360887026,"unit":"ns"},"slope":{"estimate":2458.6583068604627,"lower_bound":2450.1344637939987,"upper_bound":2470.406228677048,"unit":"ns"},"change":{"mean":{"estimate":-0.02178546830367689,"lower_bound":-0.035545066025520496,"upper_bound":-0.011166490229747089,"unit":"%"},"median":{"estimate":-0.008637896692810232,"lower_bound":-0.011519817537387511,"upper_bound":-0.004972836632816083,"unit":"%"},"change":"Improved"}}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_6","report_directory":"","iteration_count":[68,136,204,272,340,408,476,544,612,680,748,816,884,952,1020,1088,1156,1224,1292,1360,1428,1496,1564,1632,1700,1768,1836,1904,1972,2040,2108,2176,2244,2312,2380,2448,2516,2584,2652,2720,2788,2856,2924,2992,3060,3128,3196,3264,3332,3400,3468,3536,3604,3672,3740,3808,3876,3944,4012,4080,4148,4216,4284,4352,4420,4488,4556,4624,4692,4760,4828,4896,4964,5032,5100,5168,5236,5304,5372,5440,5508,5576,5644,5712,5780,5848,5916,5984,6052,6120,6188,6256,6324,6392,6460,6528,6596,6664,6732,6800],"measured_values":[1000958.0,2017500.0,2940916.0,4036708.0,4998125.0,6052666.0,6958666.0,7974250.0,8927500.0,9989458.0,10906958.0,11915000.0,12938666.0,13870042.0,14877625.0,15787000.0,16875000.0,17961500.0,18761167.0,19801834.0,20734167.0,22045333.0,22703792.0,23882792.0,24857625.0,25907334.0,26780917.0,27816417.0,28617625.0,30014709.0,31384916.0,34229584.0,33465458.0,34212333.0,35105458.0,36032541.0,37121791.0,38105250.0,38742334.0,40320417.0,40830000.0,41791250.0,43322958.0,43732250.0,45081167.0,45803708.0,46701209.0,47932500.0,48739250.0,49461834.0,50807542.0,51965208.0,52589250.0,53638083.0,54881500.0,55580125.0,56591958.0,57563042.0,58873000.0,59528916.0,60535084.0,61957084.0,62490583.0,63519917.0,64792542.0,65438458.0,66447250.0,68259625.0,68451625.0,69985542.0,70804334.0,71597792.0,72686416.0,73606708.0,74453791.0,75512125.0,76621959.0,77600084.0,78662000.0,79534708.0,80793750.0,82017667.0,82492083.0,84439875.0,84872792.0,85935292.0,87143334.0,88118125.0,89013750.0,90187792.0,90973834.0,91724417.0,93465416.0,94475625.0,94862792.0,96068667.0,96915750.0,98172291.0,99418167.0,100188417.0],"unit":"ns","throughput":[{"per_iteration":6,"unit":"elements"}],"typical":{"estimate":14681.622039134554,"lower_bound":14664.568844979916,"upper_bound":14698.195943048446,"unit":"ns"},"mean":{"estimate":14673.648895409504,"lower_bound":14649.958968953435,"upper_bound":14703.229717621001,"unit":"ns"},"median":{"estimate":14651.928211782451,"lower_bound":14630.483408748114,"upper_bound":14683.874048442907,"unit":"ns"},"median_abs_dev":{"estimate":83.91180410708868,"lower_bound":63.96102974537345,"upper_bound":101.59600177097177,"unit":"ns"},"slope":{"estimate":14681.622039134554,"lower_bound":14664.568844979916,"upper_bound":14698.195943048446,"unit":"ns"},"change":{"mean":{"estimate":-0.006601601614287023,"lower_bound":-0.009034825478360061,"upper_bound":-0.0038404134126512816,"unit":"%"},"median":{"estimate":-0.007390747824189692,"lower_bound":-0.009193153965933343,"upper_bound":-0.0044563887625256315,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_11","report_directory":"","iteration_count":[37,74,111,148,185,222,259,296,333,370,407,444,481,518,555,592,629,666,703,740,777,814,851,888,925,962,999,1036,1073,1110,1147,1184,1221,1258,1295,1332,1369,1406,1443,1480,1517,1554,1591,1628,1665,1702,1739,1776,1813,1850,1887,1924,1961,1998,2035,2072,2109,2146,2183,2220,2257,2294,2331,2368,2405,2442,2479,2516,2553,2590,2627,2664,2701,2738,2775,2812,2849,2886,2923,2960,2997,3034,3071,3108,3145,3182,3219,3256,3293,3330,3367,3404,3441,3478,3515,3552,3589,3626,3663,3700],"measured_values":[993875.0,1956708.0,2961334.0,3997875.0,5002292.0,5971042.0,6908667.0,8060458.0,9069000.0,9995792.0,11036167.0,11919500.0,12829250.0,13782833.0,14796334.0,15964542.0,16793166.0,17789166.0,18706250.0,19979292.0,20800458.0,21837792.0,22964250.0,23703542.0,24636500.0,25772917.0,27121375.0,27873917.0,28780833.0,29929208.0,30755166.0,31948459.0,32823208.0,33533916.0,34841500.0,35847542.0,36890250.0,37859041.0,38679083.0,39615625.0,40921416.0,41462708.0,42787750.0,43729375.0,44556792.0,45693166.0,46271416.0,47425542.0,49754708.0,51087709.0,50369542.0,52128000.0,52625833.0,53411167.0,55088666.0,55585708.0,56659958.0,57449208.0,58316625.0,59220208.0,60471959.0,61330958.0,62983250.0,63373583.0,64369500.0,65872916.0,66382375.0,67377292.0,68602250.0,69241042.0,76156750.0,71927458.0,72905833.0,73687209.0,74673834.0,75248125.0,76762583.0,77490458.0,78236625.0,79297000.0,80325375.0,81349875.0,82175958.0,83137250.0,84723875.0,85457292.0,86059292.0,87290416.0,88587875.0,89244792.0,90284792.0,91185458.0,92731500.0,93291583.0,94308541.0,95623083.0,96674000.0,98079291.0,98932083.0,100667916.0],"unit":"ns","throughput":[{"per_iteration":11,"unit":"elements"}],"typical":{"estimate":26903.57039480148,"lower_bound":26846.48862865717,"upper_bound":26984.552739873652,"unit":"ns"},"mean":{"estimate":26883.968449103042,"lower_bound":26837.361467984007,"upper_bound":26943.786470920066,"unit":"ns"},"median":{"estimate":26833.266614777636,"lower_bound":26804.631323631325,"upper_bound":26882.425145064175,"unit":"ns"},"median_abs_dev":{"estimate":146.591516794857,"lower_bound":112.43663406243479,"upper_bound":184.16756130329512,"unit":"ns"},"slope":{"estimate":26903.57039480148,"lower_bound":26846.48862865717,"upper_bound":26984.552739873652,"unit":"ns"},"change":{"mean":{"estimate":-0.01043167955751656,"lower_bound":-0.015827534215718912,"upper_bound":-0.006230312060284117,"unit":"%"},"median":{"estimate":-0.008333270347179456,"lower_bound":-0.010936588983882039,"upper_bound":-0.005942598018336298,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_16","report_directory":"","iteration_count":[26,52,78,104,130,156,182,208,234,260,286,312,338,364,390,416,442,468,494,520,546,572,598,624,650,676,702,728,754,780,806,832,858,884,910,936,962,988,1014,1040,1066,1092,1118,1144,1170,1196,1222,1248,1274,1300,1326,1352,1378,1404,1430,1456,1482,1508,1534,1560,1586,1612,1638,1664,1690,1716,1742,1768,1794,1820,1846,1872,1898,1924,1950,1976,2002,2028,2054,2080,2106,2132,2158,2184,2210,2236,2262,2288,2314,2340,2366,2392,2418,2444,2470,2496,2522,2548,2574,2600],"measured_values":[1012000.0,2090166.0,3019083.0,4080417.0,5155584.0,6175000.0,7234041.0,8270709.0,9079916.0,10178625.0,11378458.0,12286750.0,13217292.0,14251709.0,15201667.0,16292792.0,17211125.0,18197208.0,19165750.0,20283875.0,21365416.0,22364042.0,23708083.0,24314792.0,25511917.0,26746042.0,27946583.0,28878292.0,29827000.0,30813625.0,31889625.0,33078208.0,34006458.0,34834250.0,35747291.0,36883833.0,38151167.0,38973541.0,39754500.0,40987541.0,41612458.0,42999458.0,43566583.0,44786750.0,47014584.0,46833500.0,48242042.0,49865000.0,50468458.0,51101042.0,52870583.0,53390458.0,54509584.0,55208917.0,56154041.0,56861541.0,58006750.0,59209541.0,60082000.0,61055417.0,62520333.0,63783917.0,65996292.0,66034334.0,66232375.0,67603750.0,69145584.0,69946917.0,69959792.0,71535875.0,72735750.0,73075041.0,74272667.0,75337750.0,76302291.0,77178459.0,78699875.0,79200125.0,80308959.0,81562250.0,85588083.0,84420417.0,85076417.0,85701875.0,86933834.0,88264917.0,88284875.0,89607000.0,92013000.0,92534000.0,93726667.0,93881625.0,95623459.0,96334083.0,96519792.0,98056250.0,100008917.0,100712542.0,101072667.0,101815125.0],"unit":"ns","throughput":[{"per_iteration":16,"unit":"elements"}],"typical":{"estimate":39383.62799604415,"lower_bound":39310.751892976776,"upper_bound":39466.10353636629,"unit":"ns"},"mean":{"estimate":39371.331123081705,"lower_bound":39307.611655205576,"upper_bound":39437.252420010926,"unit":"ns"},"median":{"estimate":39316.60866448366,"lower_bound":39244.94416666667,"upper_bound":39416.56423895254,"unit":"ns"},"median_abs_dev":{"estimate":330.9674044971172,"lower_bound":242.1621734729725,"upper_bound":382.51107803962697,"unit":"ns"},"slope":{"estimate":39383.62799604415,"lower_bound":39310.751892976776,"upper_bound":39466.10353636629,"unit":"ns"},"change":{"mean":{"estimate":-0.0011140057329956932,"lower_bound":-0.003826786619819923,"upper_bound":0.0015496929845379172,"unit":"%"},"median":{"estimate":-0.000031113958156581134,"lower_bound":-0.0021916039825998768,"upper_bound":0.0032602832662347314,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_21","report_directory":"","iteration_count":[20,40,60,80,100,120,140,160,180,200,220,240,260,280,300,320,340,360,380,400,420,440,460,480,500,520,540,560,580,600,620,640,660,680,700,720,740,760,780,800,820,840,860,880,900,920,940,960,980,1000,1020,1040,1060,1080,1100,1120,1140,1160,1180,1200,1220,1240,1260,1280,1300,1320,1340,1360,1380,1400,1420,1440,1460,1480,1500,1520,1540,1560,1580,1600,1620,1640,1660,1680,1700,1720,1740,1760,1780,1800,1820,1840,1860,1880,1900,1920,1940,1960,1980,2000],"measured_values":[1022250.0,2034750.0,3050875.0,4052541.0,5089375.0,6839834.0,7608667.0,9363959.0,10991708.0,10563667.0,11439917.0,12253375.0,13364833.0,14430083.0,15568458.0,16370625.0,17572417.0,18676333.0,19715416.0,20421667.0,21501958.0,22796541.0,23681958.0,25506875.0,25535125.0,27236375.0,28777583.0,29048167.0,29903166.0,31001417.0,31932958.0,32838459.0,33625125.0,35473875.0,36917333.0,38318083.0,39094709.0,39330167.0,40800542.0,42858625.0,44139625.0,43938292.0,45038334.0,46317458.0,47031083.0,49248834.0,51852125.0,52236417.0,67018334.0,61653959.0,63073500.0,80988583.0,61553375.0,58744125.0,63273000.0,63890000.0,62701917.0,71223709.0,106560959.0,88261583.0,63972333.0,65300667.0,66263875.0,67220917.0,68605750.0,68834709.0,70658917.0,71254208.0,72590500.0,73033959.0,75753834.0,77588458.0,76571583.0,77241625.0,78542375.0,80058000.0,80786833.0,82018875.0,82825334.0,84601458.0,84099709.0,84333542.0,84956125.0,86390708.0,87163083.0,89186750.0,89462083.0,90743542.0,91780666.0,93409667.0,94531458.0,94712291.0,95856542.0,96012208.0,98458500.0,98875167.0,100031583.0,100541458.0,101672250.0,103323958.0],"unit":"ns","throughput":[{"per_iteration":21,"unit":"elements"}],"typical":{"estimate":53415.2253831831,"lower_bound":52500.35535895742,"upper_bound":54672.9493775342,"unit":"ns"},"mean":{"estimate":53819.85149171375,"lower_bound":52841.69187486958,"upper_bound":55031.42643933504,"unit":"ns"},"median":{"estimate":52223.52302552552,"lower_bound":51882.673684210524,"upper_bound":52447.66049606132,"unit":"ns"},"median_abs_dev":{"estimate":1071.0244913363103,"lower_bound":789.740565205767,"upper_bound":1445.887912204526,"unit":"ns"},"slope":{"estimate":53415.2253831831,"lower_bound":52500.35535895742,"upper_bound":54672.9493775342,"unit":"ns"},"change":{"mean":{"estimate":0.04279892169019672,"lower_bound":0.02220457945132674,"upper_bound":0.06589055615571045,"unit":"%"},"median":{"estimate":0.013212237125545201,"lower_bound":0.005991824054896178,"upper_bound":0.018018301130485748,"unit":"%"},"change":"Regressed"}}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_26","report_directory":"","iteration_count":[16,32,48,64,80,96,112,128,144,160,176,192,208,224,240,256,272,288,304,320,336,352,368,384,400,416,432,448,464,480,496,512,528,544,560,576,592,608,624,640,656,672,688,704,720,736,752,768,784,800,816,832,848,864,880,896,912,928,944,960,976,992,1008,1024,1040,1056,1072,1088,1104,1120,1136,1152,1168,1184,1200,1216,1232,1248,1264,1280,1296,1312,1328,1344,1360,1376,1392,1408,1424,1440,1456,1472,1488,1504,1520,1536,1552,1568,1584,1600],"measured_values":[1059750.0,2171083.0,2989292.0,4153416.0,5331125.0,6117917.0,7076000.0,8284875.0,9172291.0,10357625.0,11233875.0,12240916.0,13611042.0,14028917.0,15275458.0,16279916.0,17543125.0,18500208.0,19479583.0,20602208.0,21456458.0,22460292.0,23497917.0,24291791.0,25593791.0,26488584.0,27379917.0,28997625.0,29603000.0,30706333.0,31572500.0,32824084.0,33621125.0,34421250.0,35329000.0,36435833.0,37412125.0,38430000.0,39381000.0,40504458.0,41684834.0,42915291.0,43434417.0,44486000.0,45937791.0,46630083.0,47751666.0,49163958.0,50311250.0,50815959.0,52197542.0,53503000.0,54122750.0,55187042.0,56213834.0,56992250.0,58701250.0,59981875.0,59726375.0,61032583.0,62317500.0,62845250.0,64406291.0,65240375.0,65884083.0,67060291.0,68617583.0,68763750.0,69842625.0,70886292.0,72055916.0,73439042.0,74359917.0,75394291.0,76704125.0,77435708.0,78486584.0,79538250.0,80252667.0,80984500.0,82760375.0,83798042.0,84307625.0,85561583.0,85783250.0,86675917.0,88390750.0,89244375.0,90762666.0,91560000.0,92257667.0,93526250.0,94248291.0,95801958.0,96631250.0,97475334.0,98360417.0,99632500.0,100566292.0,102694750.0],"unit":"ns","throughput":[{"per_iteration":26,"unit":"elements"}],"typical":{"estimate":63597.01912978425,"lower_bound":63525.50643683138,"upper_bound":63672.810531690644,"unit":"ns"},"mean":{"estimate":63780.74971139392,"lower_bound":63648.18917580876,"upper_bound":63933.45873064412,"unit":"ns"},"median":{"estimate":63675.42693764569,"lower_bound":63559.78077997967,"upper_bound":63740.8702590812,"unit":"ns"},"median_abs_dev":{"estimate":432.7612345504058,"lower_bound":282.9083440026858,"upper_bound":506.76078743019536,"unit":"ns"},"slope":{"estimate":63597.01912978425,"lower_bound":63525.50643683138,"upper_bound":63672.810531690644,"unit":"ns"},"change":{"mean":{"estimate":0.002010368012481223,"lower_bound":-0.0017587550298478292,"upper_bound":0.005206109241407258,"unit":"%"},"median":{"estimate":0.0016979910049379487,"lower_bound":-0.00044041120688664837,"upper_bound":0.0033534369924781426,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_31","report_directory":"","iteration_count":[14,28,42,56,70,84,98,112,126,140,154,168,182,196,210,224,238,252,266,280,294,308,322,336,350,364,378,392,406,420,434,448,462,476,490,504,518,532,546,560,574,588,602,616,630,644,658,672,686,700,714,728,742,756,770,784,798,812,826,840,854,868,882,896,910,924,938,952,966,980,994,1008,1022,1036,1050,1064,1078,1092,1106,1120,1134,1148,1162,1176,1190,1204,1218,1232,1246,1260,1274,1288,1302,1316,1330,1344,1358,1372,1386,1400],"measured_values":[1078959.0,2195167.0,3150000.0,4195083.0,5315625.0,6412333.0,7632792.0,8504084.0,9735792.0,10643250.0,11542417.0,12624125.0,13987000.0,14901333.0,15854292.0,16806166.0,18024125.0,19100125.0,19982750.0,21152833.0,22447834.0,23349583.0,24309209.0,25458792.0,26448916.0,27489875.0,28857917.0,29717625.0,30760625.0,31811292.0,32793209.0,34279458.0,35026708.0,35920584.0,37079250.0,38141541.0,39027708.0,39978417.0,41051250.0,41976417.0,43163208.0,45500709.0,46540416.0,47131041.0,47787917.0,48671208.0,49615375.0,50768834.0,51558833.0,52956167.0,54188375.0,55115042.0,56204375.0,57904666.0,58284625.0,59013375.0,60429667.0,61387000.0,62117875.0,63556542.0,64537542.0,67608125.0,70014500.0,68239834.0,69698834.0,70161958.0,72660042.0,71895209.0,73158042.0,74203834.0,74991209.0,76296458.0,77231583.0,78949458.0,79626791.0,80790333.0,81556959.0,82612042.0,83619417.0,85115917.0,85988834.0,87529500.0,87780000.0,88908167.0,89810542.0,90887375.0,92219208.0,93570500.0,94267375.0,95954125.0,96587709.0,97839625.0,99877541.0,99183500.0,100906083.0,101755125.0,102419333.0,104152167.0,104867625.0,106844250.0],"unit":"ns","throughput":[{"per_iteration":31,"unit":"elements"}],"typical":{"estimate":75885.54352825688,"lower_bound":75759.95133222101,"upper_bound":76031.85895355363,"unit":"ns"},"mean":{"estimate":75906.07407211555,"lower_bound":75769.99019172697,"upper_bound":76056.45722652884,"unit":"ns"},"median":{"estimate":75722.29885427855,"lower_bound":75662.0670995671,"upper_bound":75810.33441558441,"unit":"ns"},"median_abs_dev":{"estimate":336.14440448678323,"lower_bound":252.07830048325252,"upper_bound":474.3479209457329,"unit":"ns"},"slope":{"estimate":75885.54352825688,"lower_bound":75759.95133222101,"upper_bound":76031.85895355363,"unit":"ns"},"change":{"mean":{"estimate":-0.0059965787301514695,"lower_bound":-0.01195577742915325,"upper_bound":-0.0018202149447040814,"unit":"%"},"median":{"estimate":-0.004476114761747474,"lower_bound":-0.006142467883468616,"upper_bound":-0.002860231716547701,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_36","report_directory":"","iteration_count":[12,24,36,48,60,72,84,96,108,120,132,144,156,168,180,192,204,216,228,240,252,264,276,288,300,312,324,336,348,360,372,384,396,408,420,432,444,456,468,480,492,504,516,528,540,552,564,576,588,600,612,624,636,648,660,672,684,696,708,720,732,744,756,768,780,792,804,816,828,840,852,864,876,888,900,912,924,936,948,960,972,984,996,1008,1020,1032,1044,1056,1068,1080,1092,1104,1116,1128,1140,1152,1164,1176,1188,1200],"measured_values":[1059250.0,2087583.0,3183833.0,4189875.0,5303000.0,6399125.0,7382333.0,8527500.0,9431916.0,10475916.0,11538084.0,12537250.0,13869042.0,14942500.0,15727958.0,16705084.0,17782792.0,18907042.0,19889667.0,21012250.0,22067333.0,23097625.0,24091042.0,25181541.0,26152416.0,27093916.0,28652334.0,29601125.0,30679417.0,31561792.0,32470625.0,34254625.0,34982417.0,36039250.0,37048417.0,38040667.0,39072625.0,40023542.0,42417583.0,43118417.0,43051375.0,44231833.0,45178458.0,46986041.0,47975583.0,48423292.0,49540125.0,50393666.0,51234667.0,52208500.0,53629125.0,54720292.0,55345500.0,56788583.0,58165750.0,58704834.0,59716041.0,60967500.0,61845834.0,62985833.0,64157084.0,65390291.0,67393791.0,68170167.0,68803458.0,69908417.0,71680750.0,71921083.0,73882000.0,75039791.0,75631500.0,75683625.0,76598292.0,77846083.0,78611166.0,80047959.0,80815833.0,82331208.0,82625708.0,83919375.0,84804625.0,85957958.0,87436584.0,87967208.0,89162500.0,90472291.0,91205250.0,92697000.0,93393833.0,95138875.0,96506167.0,96907916.0,99141125.0,99880083.0,99930084.0,101558625.0,102519875.0,104239500.0,105701958.0,106484708.0],"unit":"ns","throughput":[{"per_iteration":36,"unit":"elements"}],"typical":{"estimate":87980.03231885128,"lower_bound":87822.01413484663,"upper_bound":88139.39299520472,"unit":"ns"},"mean":{"estimate":87922.29013430641,"lower_bound":87788.60593104259,"upper_bound":88062.12326945101,"unit":"ns"},"median":{"estimate":87708.06535674472,"lower_bound":87593.87191358025,"upper_bound":87898.89668576105,"unit":"ns"},"median_abs_dev":{"estimate":613.6245657656989,"lower_bound":437.53433031700143,"upper_bound":767.4555410798895,"unit":"ns"},"slope":{"estimate":87980.03231885128,"lower_bound":87822.01413484663,"upper_bound":88139.39299520472,"unit":"ns"},"change":{"mean":{"estimate":-0.025155063352442664,"lower_bound":-0.038992962995150926,"upper_bound":-0.015025549776904786,"unit":"%"},"median":{"estimate":-0.014163610426288886,"lower_bound":-0.01772211545638025,"upper_bound":-0.011232665765548822,"unit":"%"},"change":"Improved"}}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_41","report_directory":"","iteration_count":[10,20,30,40,50,60,70,80,90,100,110,120,130,140,150,160,170,180,190,200,210,220,230,240,250,260,270,280,290,300,310,320,330,340,350,360,370,380,390,400,410,420,430,440,450,460,470,480,490,500,510,520,530,540,550,560,570,580,590,600,610,620,630,640,650,660,670,680,690,700,710,720,730,740,750,760,770,780,790,800,810,820,830,840,850,860,870,880,890,900,910,920,930,940,950,960,970,980,990,1000],"measured_values":[999833.0,1977834.0,3010792.0,4067459.0,4973333.0,6280208.0,7091917.0,8043750.0,8945000.0,10294333.0,11123792.0,12150125.0,13058334.0,14076292.0,15242417.0,16167167.0,17096542.0,18156334.0,19238917.0,20079917.0,21181042.0,22361417.0,23780958.0,24314708.0,25697042.0,26636167.0,27734625.0,28765417.0,29463625.0,30388167.0,31705584.0,32497459.0,33975375.0,34373750.0,35812875.0,36259708.0,37304083.0,38700041.0,39267709.0,39864291.0,41642875.0,42715375.0,43709958.0,43997833.0,45105542.0,46806083.0,47216708.0,48082708.0,49075750.0,50111750.0,51270958.0,51970459.0,52875875.0,54481292.0,55561667.0,56398875.0,57219875.0,58984875.0,59289542.0,60887000.0,61698958.0,62370791.0,63008084.0,64112000.0,64702708.0,66871666.0,68979125.0,68330792.0,69179000.0,70368625.0,70739875.0,72654041.0,73347542.0,74566834.0,76038917.0,76709792.0,78116708.0,78771500.0,79976958.0,79531792.0,81181917.0,81767584.0,82815416.0,84423708.0,85934834.0,86192667.0,88453958.0,89995375.0,92265541.0,92845417.0,93795041.0,94070250.0,95071500.0,93844583.0,96378292.0,96518167.0,102366833.0,101168542.0,99870000.0,101857917.0],"unit":"ns","throughput":[{"per_iteration":41,"unit":"elements"}],"typical":{"estimate":101268.17603546624,"lower_bound":100908.4956809108,"upper_bound":101651.05233546335,"unit":"ns"},"mean":{"estimate":101153.51816464477,"lower_bound":100930.99361985174,"upper_bound":101387.37922124886,"unit":"ns"},"median":{"estimate":100961.51970310391,"lower_bound":100703.92222222223,"upper_bound":101272.46583333334,"unit":"ns"},"median_abs_dev":{"estimate":1016.1964503801805,"lower_bound":761.9403243516217,"upper_bound":1230.69992559606,"unit":"ns"},"slope":{"estimate":101268.17603546624,"lower_bound":100908.4956809108,"upper_bound":101651.05233546335,"unit":"ns"},"change":{"mean":{"estimate":0.0016045228271739376,"lower_bound":-0.003098896747436382,"upper_bound":0.005802786449328634,"unit":"%"},"median":{"estimate":0.0024748748521414843,"lower_bound":-0.0013988259437306416,"upper_bound":0.0062332213101732314,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_46","report_directory":"","iteration_count":[9,18,27,36,45,54,63,72,81,90,99,108,117,126,135,144,153,162,171,180,189,198,207,216,225,234,243,252,261,270,279,288,297,306,315,324,333,342,351,360,369,378,387,396,405,414,423,432,441,450,459,468,477,486,495,504,513,522,531,540,549,558,567,576,585,594,603,612,621,630,639,648,657,666,675,684,693,702,711,720,729,738,747,756,765,774,783,792,801,810,819,828,837,846,855,864,873,882,891,900],"measured_values":[1069417.0,2108000.0,3186750.0,4124917.0,5111583.0,6115459.0,7165834.0,8353208.0,9143542.0,10360750.0,11119000.0,12157083.0,13167167.0,14200000.0,15327334.0,16391583.0,17433959.0,18085125.0,19325792.0,20214666.0,21136708.0,22750667.0,23252334.0,24298250.0,25484833.0,26476459.0,27306417.0,28237250.0,29260459.0,30653833.0,31687916.0,32704708.0,33549833.0,34959375.0,35529208.0,36539792.0,37418167.0,38543917.0,39339416.0,40533208.0,41316583.0,42537875.0,43646375.0,45089167.0,45411542.0,46336708.0,47715042.0,48711959.0,49603500.0,50483500.0,52240208.0,53178833.0,54079333.0,54446333.0,56044834.0,57572333.0,57784875.0,58885625.0,59956541.0,61298000.0,62237041.0,64886959.0,64335917.0,65654500.0,67082625.0,67646166.0,68405833.0,69273167.0,70366333.0,71060667.0,72295667.0,73725167.0,74403334.0,74776167.0,75812167.0,77666875.0,78103125.0,79184583.0,80319750.0,81314625.0,81569250.0,83159500.0,83986000.0,84845750.0,86065958.0,86911833.0,95682958.0,89814083.0,91090708.0,91243666.0,92155542.0,93228542.0,93854209.0,95098666.0,96807875.0,97208333.0,98041000.0,101194292.0,101170667.0,101443709.0],"unit":"ns","throughput":[{"per_iteration":46,"unit":"elements"}],"typical":{"estimate":113188.97710818842,"lower_bound":112841.37067162906,"upper_bound":113699.50985118448,"unit":"ns"},"mean":{"estimate":113296.1034817892,"lower_bound":113026.63749533288,"upper_bound":113611.55207832843,"unit":"ns"},"median":{"estimate":112897.87057961916,"lower_bound":112715.23222222223,"upper_bound":113234.48529126885,"unit":"ns"},"median_abs_dev":{"estimate":854.647918752894,"lower_bound":644.4118748925501,"upper_bound":992.404313974183,"unit":"ns"},"slope":{"estimate":113188.97710818842,"lower_bound":112841.37067162906,"upper_bound":113699.50985118448,"unit":"ns"},"change":{"mean":{"estimate":-0.0017885621260025353,"lower_bound":-0.004848778164318875,"upper_bound":0.0014665711955599236,"unit":"%"},"median":{"estimate":-0.0047156373279959585,"lower_bound":-0.007306569064802992,"upper_bound":-0.0006775154345839418,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_51","report_directory":"","iteration_count":[8,16,24,32,40,48,56,64,72,80,88,96,104,112,120,128,136,144,152,160,168,176,184,192,200,208,216,224,232,240,248,256,264,272,280,288,296,304,312,320,328,336,344,352,360,368,376,384,392,400,408,416,424,432,440,448,456,464,472,480,488,496,504,512,520,528,536,544,552,560,568,576,584,592,600,608,616,624,632,640,648,656,664,672,680,688,696,704,712,720,728,736,744,752,760,768,776,784,792,800],"measured_values":[990250.0,2001750.0,3007458.0,3997458.0,4996459.0,6083125.0,7022709.0,8021792.0,8972584.0,10113750.0,10994583.0,11989208.0,12934458.0,13844166.0,15024542.0,15990042.0,17081042.0,19170834.0,20375709.0,20946166.0,20996292.0,21786792.0,23688334.0,24115833.0,24992000.0,26125708.0,27360334.0,28114625.0,28822459.0,29814042.0,30991250.0,31761625.0,33034041.0,34094625.0,35235209.0,35846083.0,37063125.0,37838334.0,39023708.0,40137708.0,41064083.0,42083541.0,42855959.0,44164541.0,44992584.0,45479584.0,46800250.0,47807000.0,48434917.0,50995584.0,52029125.0,52583250.0,52659834.0,53595042.0,54973375.0,55869250.0,56606708.0,57923917.0,58492000.0,59666792.0,60708209.0,61606583.0,62717167.0,63840583.0,64718458.0,65816167.0,66430750.0,67561500.0,69454417.0,69548708.0,70912000.0,71958625.0,72304875.0,73818209.0,75159750.0,75968708.0,77195000.0,78076875.0,78529084.0,79542833.0,80785709.0,81915708.0,82727708.0,84065208.0,84349417.0,85431459.0,86559250.0,87572417.0,89087708.0,89774000.0,91174916.0,91573209.0,93036875.0,93604167.0,94813958.0,95289333.0,96271250.0,97467958.0,98408334.0,100059833.0],"unit":"ns","throughput":[{"per_iteration":51,"unit":"elements"}],"typical":{"estimate":124701.79332791489,"lower_bound":124571.33233316314,"upper_bound":124845.66974698023,"unit":"ns"},"mean":{"estimate":125108.03875804484,"lower_bound":124825.37675852484,"upper_bound":125444.2271910638,"unit":"ns"},"median":{"estimate":124858.2882600481,"lower_bound":124619.22222222222,"upper_bound":124960.0,"unit":"ns"},"median_abs_dev":{"estimate":684.902371221235,"lower_bound":551.2295136200923,"upper_bound":846.5129078079252,"unit":"ns"},"slope":{"estimate":124701.79332791489,"lower_bound":124571.33233316314,"upper_bound":124845.66974698023,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_56","report_directory":"","iteration_count":[8,16,24,32,40,48,56,64,72,80,88,96,104,112,120,128,136,144,152,160,168,176,184,192,200,208,216,224,232,240,248,256,264,272,280,288,296,304,312,320,328,336,344,352,360,368,376,384,392,400,408,416,424,432,440,448,456,464,472,480,488,496,504,512,520,528,536,544,552,560,568,576,584,592,600,608,616,624,632,640,648,656,664,672,680,688,696,704,712,720,728,736,744,752,760,768,776,784,792,800],"measured_values":[1091417.0,2178875.0,3376666.0,4401834.0,5409208.0,6644167.0,7749750.0,8728750.0,10046000.0,11235000.0,12100875.0,13087875.0,14207000.0,15419291.0,16537125.0,17650583.0,18626125.0,19861709.0,20772083.0,22081459.0,23061875.0,24224834.0,25085333.0,26475958.0,27416416.0,28406625.0,29380667.0,30670958.0,31689208.0,32907875.0,33966375.0,35258792.0,36279875.0,37371750.0,38299875.0,39276042.0,40580750.0,41633542.0,42828250.0,43877583.0,44884458.0,46629583.0,47499125.0,48143041.0,49199000.0,50792833.0,51842292.0,52432209.0,53711417.0,54694792.0,56267417.0,57183792.0,57868250.0,59616084.0,60479458.0,61165458.0,62537833.0,63463541.0,64798958.0,71593750.0,67706958.0,68462792.0,69497500.0,70402542.0,71269667.0,71905083.0,73475792.0,74712791.0,75837500.0,76884625.0,77951083.0,79055792.0,79741500.0,81223833.0,82088709.0,83316958.0,84324042.0,84948125.0,86499334.0,87649792.0,89028250.0,89857875.0,91006500.0,91769083.0,92506708.0,94383667.0,94676917.0,96664000.0,97279417.0,98272125.0,99380458.0,100291542.0,101348916.0,102728166.0,103682917.0,106645417.0,107863167.0,107449792.0,109291375.0,113808875.0],"unit":"ns","throughput":[{"per_iteration":56,"unit":"elements"}],"typical":{"estimate":137367.3885750702,"lower_bound":136987.9645750104,"upper_bound":137843.78438588968,"unit":"ns"},"mean":{"estimate":137390.1818702348,"lower_bound":137122.7332468947,"upper_bound":137730.1266042509,"unit":"ns"},"median":{"estimate":137089.60418918918,"lower_bound":136956.80147058822,"upper_bound":137277.9800086919,"unit":"ns"},"median_abs_dev":{"estimate":754.2594564235093,"lower_bound":559.2645644408261,"upper_bound":966.8028097174194,"unit":"ns"},"slope":{"estimate":137367.3885750702,"lower_bound":136987.9645750104,"upper_bound":137843.78438588968,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_61","report_directory":"","iteration_count":[7,14,21,28,35,42,49,56,63,70,77,84,91,98,105,112,119,126,133,140,147,154,161,168,175,182,189,196,203,210,217,224,231,238,245,252,259,266,273,280,287,294,301,308,315,322,329,336,343,350,357,364,371,378,385,392,399,406,413,420,427,434,441,448,455,462,469,476,483,490,497,504,511,518,525,532,539,546,553,560,567,574,581,588,595,602,609,616,623,630,637,644,651,658,665,672,679,686,693,700],"measured_values":[1095167.0,2085625.0,3164417.0,4303750.0,5297542.0,6179541.0,7425083.0,8306833.0,9285125.0,10464458.0,11575333.0,12711083.0,13648709.0,14570542.0,15579667.0,16797208.0,17827583.0,18897750.0,20037333.0,20870791.0,22067417.0,22888334.0,24397083.0,25096583.0,25985000.0,27195625.0,28314834.0,29249375.0,30378125.0,31341334.0,32444042.0,33760500.0,34315417.0,35470083.0,36758334.0,38558083.0,39633625.0,39804042.0,40749875.0,41735042.0,42707916.0,44106666.0,45061292.0,45746208.0,47073125.0,48307959.0,49021792.0,50291041.0,50986417.0,52023250.0,52949208.0,54617375.0,55354500.0,56227291.0,57490625.0,58573708.0,59574625.0,60467000.0,61572000.0,62609792.0,64357833.0,65547541.0,65677375.0,67095459.0,67939875.0,68868542.0,70236583.0,71507417.0,71840625.0,73123500.0,75819250.0,76832708.0,77021291.0,78294167.0,78653083.0,79259375.0,80537500.0,81585959.0,82403167.0,83137666.0,84508833.0,85492417.0,86882917.0,88486250.0,89600333.0,90728250.0,91950167.0,92760417.0,93113750.0,94408834.0,96288583.0,96851083.0,97106500.0,97876875.0,99706541.0,100912750.0,102768542.0,103587833.0,103803791.0,105111292.0],"unit":"ns","throughput":[{"per_iteration":61,"unit":"elements"}],"typical":{"estimate":149902.2901340539,"lower_bound":149663.10998037385,"upper_bound":150139.16312270396,"unit":"ns"},"mean":{"estimate":149854.01207330797,"lower_bound":149608.29691263215,"upper_bound":150121.31471354055,"unit":"ns"},"median":{"estimate":149589.7801431289,"lower_bound":149373.26066790352,"upper_bound":149834.61428571428,"unit":"ns"},"median_abs_dev":{"estimate":884.6417534167471,"lower_bound":699.0124134521193,"upper_bound":1234.871814279715,"unit":"ns"},"slope":{"estimate":149902.2901340539,"lower_bound":149663.10998037385,"upper_bound":150139.16312270396,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_66","report_directory":"","iteration_count":[7,14,21,28,35,42,49,56,63,70,77,84,91,98,105,112,119,126,133,140,147,154,161,168,175,182,189,196,203,210,217,224,231,238,245,252,259,266,273,280,287,294,301,308,315,322,329,336,343,350,357,364,371,378,385,392,399,406,413,420,427,434,441,448,455,462,469,476,483,490,497,504,511,518,525,532,539,546,553,560,567,574,581,588,595,602,609,616,623,630,637,644,651,658,665,672,679,686,693,700],"measured_values":[1121083.0,2217875.0,3468167.0,4446542.0,5627750.0,6746000.0,8041666.0,9107458.0,10278958.0,11407875.0,12534666.0,13538959.0,14663584.0,15954500.0,16890208.0,18173250.0,19220209.0,20384833.0,21400500.0,22432500.0,23703875.0,24761958.0,26152750.0,27265750.0,28312375.0,29394458.0,30464917.0,31744917.0,32640667.0,34352792.0,35063917.0,36628292.0,37820583.0,38780167.0,39735209.0,40628958.0,41965583.0,43182375.0,44322166.0,45206041.0,46464500.0,47713584.0,48828292.0,49756708.0,51366709.0,51740375.0,53252250.0,54400625.0,55632166.0,56284542.0,57160458.0,58523625.0,60745125.0,61839750.0,62579167.0,63787333.0,64331166.0,66025708.0,66623417.0,68286542.0,69801750.0,70014959.0,71299583.0,72605625.0,75234875.0,78065625.0,77061291.0,77862209.0,79293167.0,80006542.0,80006041.0,81141166.0,82939542.0,84029625.0,85172250.0,86708958.0,90498084.0,90822250.0,92995333.0,93345291.0,92318041.0,96800375.0,96734417.0,96620542.0,97490375.0,97451625.0,98713833.0,100793667.0,100927709.0,103206416.0,103956417.0,104563042.0,105800459.0,107179166.0,108118083.0,108413542.0,110495875.0,111972334.0,112659917.0,114210416.0],"unit":"ns","throughput":[{"per_iteration":66,"unit":"elements"}],"typical":{"estimate":163211.6778521818,"lower_bound":162751.9459629699,"upper_bound":163712.8659358417,"unit":"ns"},"mean":{"estimate":162580.88399384616,"lower_bound":162234.04445924057,"upper_bound":162953.1339448004,"unit":"ns"},"median":{"estimate":162302.21616112196,"lower_bound":162060.47889761682,"upper_bound":162604.24000518536,"unit":"ns"},"median_abs_dev":{"estimate":1347.261816794992,"lower_bound":993.1961503557316,"upper_bound":1685.165137985142,"unit":"ns"},"slope":{"estimate":163211.6778521818,"lower_bound":162751.9459629699,"upper_bound":163712.8659358417,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_71","report_directory":"","iteration_count":[6,12,18,24,30,36,42,48,54,60,66,72,78,84,90,96,102,108,114,120,126,132,138,144,150,156,162,168,174,180,186,192,198,204,210,216,222,228,234,240,246,252,258,264,270,276,282,288,294,300,306,312,318,324,330,336,342,348,354,360,366,372,378,384,390,396,402,408,414,420,426,432,438,444,450,456,462,468,474,480,486,492,498,504,510,516,522,528,534,540,546,552,558,564,570,576,582,588,594,600],"measured_values":[1036000.0,2075084.0,3135125.0,4316959.0,5181084.0,6223250.0,7436333.0,8491250.0,9579416.0,10562667.0,11479417.0,12587292.0,13563167.0,14497375.0,15628583.0,16663708.0,17910500.0,18729000.0,19798750.0,20778208.0,21709208.0,22836583.0,23865583.0,25151583.0,25915875.0,27147583.0,27984000.0,29221917.0,30028583.0,32004041.0,32414458.0,34146584.0,34842292.0,35692709.0,36700541.0,37190250.0,38556375.0,40566958.0,41377875.0,41654542.0,42949250.0,43843042.0,44644750.0,45649375.0,47038791.0,47653917.0,48617333.0,49747458.0,51041166.0,51748709.0,52851959.0,53941792.0,55704833.0,56386042.0,56778917.0,58052708.0,58952333.0,59920083.0,63715334.0,62329333.0,63949083.0,65130167.0,65740291.0,66954208.0,67963584.0,69557416.0,71113459.0,71210750.0,72287167.0,76043291.0,81510542.0,75291958.0,76104125.0,76936666.0,78518500.0,78698666.0,79516459.0,80639250.0,82828333.0,84492792.0,85454166.0,86929583.0,87174333.0,87813625.0,89054792.0,89040000.0,90264000.0,91613583.0,92775500.0,94746625.0,96375458.0,97412250.0,97052334.0,102314791.0,99123000.0,100422792.0,100887417.0,103406792.0,106031792.0,105915250.0],"unit":"ns","throughput":[{"per_iteration":71,"unit":"elements"}],"typical":{"estimate":175050.7075720408,"lower_bound":174368.09501159465,"upper_bound":175831.34565149577,"unit":"ns"},"mean":{"estimate":174650.3692174463,"lower_bound":174184.76171955746,"upper_bound":175191.35495799858,"unit":"ns"},"median":{"estimate":174001.64667277166,"lower_bound":173677.36486486485,"upper_bound":174415.34027777778,"unit":"ns"},"median_abs_dev":{"estimate":1629.848192014272,"lower_bound":1270.9765453202506,"upper_bound":2213.340554038697,"unit":"ns"},"slope":{"estimate":175050.7075720408,"lower_bound":174368.09501159465,"upper_bound":175831.34565149577,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_76","report_directory":"","iteration_count":[6,12,18,24,30,36,42,48,54,60,66,72,78,84,90,96,102,108,114,120,126,132,138,144,150,156,162,168,174,180,186,192,198,204,210,216,222,228,234,240,246,252,258,264,270,276,282,288,294,300,306,312,318,324,330,336,342,348,354,360,366,372,378,384,390,396,402,408,414,420,426,432,438,444,450,456,462,468,474,480,486,492,498,504,510,516,522,528,534,540,546,552,558,564,570,576,582,588,594,600],"measured_values":[1111250.0,2238333.0,3456959.0,4440000.0,5536084.0,6705250.0,7835209.0,8866375.0,9938417.0,11175500.0,12213792.0,13347542.0,14400292.0,15559917.0,16762584.0,17807875.0,19032917.0,19990875.0,21135125.0,22205917.0,23516417.0,24684958.0,25603917.0,26904042.0,27970333.0,28863667.0,30592416.0,31172416.0,32575292.0,33769834.0,34700750.0,36057875.0,37104917.0,38227583.0,39473959.0,40129875.0,41501042.0,42544750.0,43331000.0,44595625.0,45809000.0,46869083.0,48223875.0,49078000.0,50228583.0,51385750.0,52570417.0,53841500.0,54928625.0,55920916.0,57442250.0,58381041.0,59404084.0,60374292.0,61349750.0,62567500.0,63147417.0,64338541.0,65356417.0,67017041.0,67905917.0,69253417.0,70437417.0,71591000.0,72469708.0,73298417.0,74898834.0,75604792.0,77240916.0,78517417.0,79801833.0,81138541.0,82606542.0,84406000.0,84696666.0,85581166.0,85634375.0,87445375.0,87786250.0,89375708.0,90850584.0,91611708.0,92658375.0,93994708.0,95784334.0,99386042.0,99714750.0,99255750.0,99143333.0,100380750.0,100878458.0,102677792.0,103734166.0,104688625.0,105655084.0,106768166.0,108026875.0,109414334.0,111179917.0,112144459.0],"unit":"ns","throughput":[{"per_iteration":76,"unit":"elements"}],"typical":{"estimate":186616.82833357961,"lower_bound":186229.72984303057,"upper_bound":187069.55217696467,"unit":"ns"},"mean":{"estimate":186431.0296149564,"lower_bound":186166.50878481203,"upper_bound":186721.17736805245,"unit":"ns"},"median":{"estimate":186233.19024390244,"lower_bound":186031.7888888889,"upper_bound":186468.88666666666,"unit":"ns"},"median_abs_dev":{"estimate":1045.6838160158172,"lower_bound":814.7924819008226,"upper_bound":1333.0815116281224,"unit":"ns"},"slope":{"estimate":186616.82833357961,"lower_bound":186229.72984303057,"upper_bound":187069.55217696467,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_81","report_directory":"","iteration_count":[5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80,85,90,95,100,105,110,115,120,125,130,135,140,145,150,155,160,165,170,175,180,185,190,195,200,205,210,215,220,225,230,235,240,245,250,255,260,265,270,275,280,285,290,295,300,305,310,315,320,325,330,335,340,345,350,355,360,365,370,375,380,385,390,395,400,405,410,415,420,425,430,435,440,445,450,455,460,465,470,475,480,485,490,495,500],"measured_values":[988000.0,1963458.0,2989916.0,3921708.0,4981417.0,5847417.0,6979709.0,7985250.0,8807458.0,9916375.0,10868584.0,11864750.0,12830917.0,13746417.0,14774625.0,15759292.0,16697625.0,17686333.0,18661334.0,19834750.0,20746375.0,21772708.0,22699709.0,23871500.0,24755833.0,25621750.0,26560708.0,27543333.0,28659708.0,29716625.0,30732875.0,31596458.0,32515667.0,33638000.0,34566833.0,35996667.0,36655416.0,37691834.0,38407042.0,39177500.0,40598625.0,41478834.0,42420083.0,45671542.0,44416708.0,45408292.0,46078542.0,47422209.0,48220042.0,49155125.0,50488875.0,51297291.0,52422667.0,53207833.0,54581292.0,55585791.0,56377875.0,57972333.0,58528667.0,59380791.0,60345000.0,61371792.0,62236333.0,63032667.0,64049125.0,65008750.0,65925500.0,67250792.0,67963583.0,69031250.0,70092791.0,71090875.0,72244875.0,73424042.0,74252959.0,75004416.0,75908209.0,76803167.0,77594333.0,78796875.0,81966416.0,97640292.0,84366458.0,84806458.0,86497583.0,86083125.0,87775542.0,88619958.0,88583417.0,89200666.0,90351875.0,91193583.0,92740792.0,93226334.0,94478125.0,95661375.0,96226625.0,97346500.0,98641875.0,99350708.0],"unit":"ns","throughput":[{"per_iteration":81,"unit":"elements"}],"typical":{"estimate":199450.93541776267,"lower_bound":198305.02031409755,"upper_bound":201290.7248408904,"unit":"ns"},"mean":{"estimate":198501.94314135052,"lower_bound":197851.02104044068,"upper_bound":199505.8643909365,"unit":"ns"},"median":{"estimate":197806.7761609907,"lower_bound":197518.25714285715,"upper_bound":198018.83070301294,"unit":"ns"},"median_abs_dev":{"estimate":1092.137555442505,"lower_bound":799.6420583349748,"upper_bound":1390.4064769754016,"unit":"ns"},"slope":{"estimate":199450.93541776267,"lower_bound":198305.02031409755,"upper_bound":201290.7248408904,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_86","report_directory":"","iteration_count":[5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80,85,90,95,100,105,110,115,120,125,130,135,140,145,150,155,160,165,170,175,180,185,190,195,200,205,210,215,220,225,230,235,240,245,250,255,260,265,270,275,280,285,290,295,300,305,310,315,320,325,330,335,340,345,350,355,360,365,370,375,380,385,390,395,400,405,410,415,420,425,430,435,440,445,450,455,460,465,470,475,480,485,490,495,500],"measured_values":[1052833.0,2153709.0,3177208.0,4187417.0,5360583.0,6227167.0,7383584.0,8422792.0,9394375.0,10718333.0,11642541.0,12633708.0,13860500.0,14892125.0,15845542.0,16727542.0,18035250.0,18935334.0,19827042.0,21082375.0,22136250.0,23068167.0,24317292.0,25400625.0,26125750.0,27398292.0,28414500.0,29495584.0,30551292.0,31390625.0,32692125.0,33881834.0,34960167.0,35833000.0,36887125.0,37715333.0,39209041.0,40269375.0,40850625.0,42205625.0,43032375.0,44564334.0,45611042.0,46126334.0,47594250.0,48922917.0,49664250.0,50686458.0,51848125.0,52707958.0,53978417.0,55347834.0,55960709.0,57190959.0,58087625.0,59099417.0,60167667.0,61971625.0,62126125.0,63379208.0,64421083.0,65373375.0,66282083.0,67630333.0,68374709.0,69721541.0,71070625.0,71694792.0,72851208.0,73866459.0,74875708.0,75895625.0,77341000.0,77709291.0,78811500.0,83659833.0,82208959.0,82683375.0,83253666.0,85169083.0,86242584.0,87164667.0,87451042.0,88656625.0,90288000.0,91658125.0,91894541.0,93156666.0,94505750.0,95520542.0,96625000.0,98117000.0,98718625.0,99735208.0,100834375.0,101090292.0,102091542.0,103907291.0,103952125.0,106277666.0],"unit":"ns","throughput":[{"per_iteration":86,"unit":"elements"}],"typical":{"estimate":211746.59339677848,"lower_bound":211392.98934948927,"upper_bound":212158.3933079904,"unit":"ns"},"mean":{"estimate":211414.47310922403,"lower_bound":211125.01254398466,"upper_bound":211732.4755350951,"unit":"ns"},"median":{"estimate":211222.19642324888,"lower_bound":210959.54285714286,"upper_bound":211625.0,"unit":"ns"},"median_abs_dev":{"estimate":1067.8042784289196,"lower_bound":823.3536587158873,"upper_bound":1372.115902931738,"unit":"ns"},"slope":{"estimate":211746.59339677848,"lower_bound":211392.98934948927,"upper_bound":212158.3933079904,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_91","report_directory":"","iteration_count":[5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80,85,90,95,100,105,110,115,120,125,130,135,140,145,150,155,160,165,170,175,180,185,190,195,200,205,210,215,220,225,230,235,240,245,250,255,260,265,270,275,280,285,290,295,300,305,310,315,320,325,330,335,340,345,350,355,360,365,370,375,380,385,390,395,400,405,410,415,420,425,430,435,440,445,450,455,460,465,470,475,480,485,490,495,500],"measured_values":[1141333.0,2243625.0,3403250.0,4538625.0,5601583.0,6717417.0,7874125.0,9059334.0,10024833.0,11233708.0,12243375.0,13247708.0,14549000.0,15617875.0,16695750.0,17925875.0,18824000.0,19905333.0,21072083.0,22216250.0,23707334.0,24927750.0,28241000.0,26540875.0,28393417.0,28861042.0,30121792.0,31055459.0,32399875.0,33671333.0,34806916.0,35906916.0,36785375.0,37918041.0,39093708.0,39964458.0,41672125.0,42629333.0,43586750.0,44681084.0,45758583.0,46884125.0,47716166.0,49097917.0,50009041.0,51258875.0,52547166.0,53361416.0,54442125.0,55951583.0,57237166.0,58152375.0,58718958.0,60748875.0,62467000.0,62895208.0,63581500.0,66354417.0,66719000.0,66979375.0,67710791.0,69396042.0,70465375.0,71530916.0,72533792.0,73489833.0,74856708.0,75983458.0,77021625.0,77766375.0,79666792.0,80600500.0,81341292.0,82473625.0,83861792.0,86285541.0,86625000.0,87039708.0,88994875.0,89096416.0,90248041.0,91612209.0,92833083.0,93980166.0,95173334.0,97806583.0,98759375.0,98949250.0,99749292.0,100878417.0,101729792.0,102899875.0,104482000.0,104820625.0,106164458.0,110727834.0,108717875.0,111307042.0,110847542.0,111323000.0],"unit":"ns","throughput":[{"per_iteration":91,"unit":"elements"}],"typical":{"estimate":224241.11062686567,"lower_bound":223774.19987162016,"upper_bound":224777.45815161054,"unit":"ns"},"mean":{"estimate":224163.27043389494,"lower_bound":223690.01769723202,"upper_bound":224772.9117623163,"unit":"ns"},"median":{"estimate":223618.2035177305,"lower_bound":223405.42,"upper_bound":223890.27777777778,"unit":"ns"},"median_abs_dev":{"estimate":1216.6900896298594,"lower_bound":904.8179435081975,"upper_bound":1591.149861289872,"unit":"ns"},"slope":{"estimate":224241.11062686567,"lower_bound":223774.19987162016,"upper_bound":224777.45815161054,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"linear_data/linear_data_96","report_directory":"","iteration_count":[5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80,85,90,95,100,105,110,115,120,125,130,135,140,145,150,155,160,165,170,175,180,185,190,195,200,205,210,215,220,225,230,235,240,245,250,255,260,265,270,275,280,285,290,295,300,305,310,315,320,325,330,335,340,345,350,355,360,365,370,375,380,385,390,395,400,405,410,415,420,425,430,435,440,445,450,455,460,465,470,475,480,485,490,495,500],"measured_values":[1148959.0,2418208.0,3503083.0,4726042.0,6036917.0,7104208.0,8286084.0,9539916.0,10694792.0,11762500.0,12920500.0,14222125.0,15761500.0,16555042.0,17548334.0,18814000.0,20071916.0,21329541.0,22466166.0,23763750.0,24829583.0,26177291.0,27263875.0,28169917.0,29473334.0,30949458.0,32144083.0,32918542.0,34559417.0,35572208.0,36574625.0,37757667.0,39150166.0,40127792.0,41319625.0,42178875.0,43627917.0,44598333.0,48058958.0,48048208.0,48593708.0,49620375.0,50670541.0,51821375.0,53647000.0,54017875.0,55392125.0,56619125.0,57845000.0,58853625.0,60124042.0,61637041.0,62883542.0,63690416.0,65032375.0,65816292.0,66772166.0,68379916.0,69564000.0,70358792.0,71553000.0,72908792.0,73969209.0,75132875.0,76727791.0,77366166.0,78756167.0,80555292.0,80785667.0,82330167.0,83753042.0,84633834.0,85866750.0,86925208.0,88011500.0,89482542.0,90490042.0,90939000.0,92631917.0,93498167.0,95176250.0,96108709.0,97632625.0,98624875.0,99709750.0,100812917.0,101958708.0,102807875.0,104451833.0,105304000.0,106526750.0,108639042.0,109235459.0,110360708.0,111681667.0,112645291.0,113450042.0,114907667.0,115903625.0,116984250.0],"unit":"ns","throughput":[{"per_iteration":96,"unit":"elements"}],"typical":{"estimate":235047.28217940003,"lower_bound":234823.66373598424,"upper_bound":235313.19422963035,"unit":"ns"},"mean":{"estimate":235869.29480942682,"lower_bound":235484.4261022497,"upper_bound":236292.37763172665,"unit":"ns"},"median":{"estimate":235447.43684210526,"lower_bound":235093.98333333334,"upper_bound":235890.42962962962,"unit":"ns"},"median_abs_dev":{"estimate":1261.7667075991674,"lower_bound":1003.5466428286445,"upper_bound":1740.2878220914424,"unit":"ns"},"slope":{"estimate":235047.28217940003,"lower_bound":234823.66373598424,"upper_bound":235313.19422963035,"unit":"ns"},"change":null}
-        {"reason":"group-complete","group_name":"linear_data","benchmarks":["linear_data/linear_data_10","linear_data/linear_data_17","linear_data/linear_data_19","linear_data/linear_data_18","linear_data/linear_data_5","linear_data/linear_data_2","linear_data/linear_data_3","linear_data/linear_data_4","linear_data/linear_data_14","linear_data/linear_data_13","linear_data/linear_data_12","linear_data/linear_data_15","linear_data/linear_data_8","linear_data/linear_data_9","linear_data/linear_data_7","linear_data/linear_data_0","linear_data/linear_data_1","linear_data/linear_data_6","linear_data/linear_data_11","linear_data/linear_data_16","linear_data/linear_data_21","linear_data/linear_data_26","linear_data/linear_data_31","linear_data/linear_data_36","linear_data/linear_data_41","linear_data/linear_data_46","linear_data/linear_data_51","linear_data/linear_data_56","linear_data/linear_data_61","linear_data/linear_data_66","linear_data/linear_data_71","linear_data/linear_data_76","linear_data/linear_data_81","linear_data/linear_data_86","linear_data/linear_data_91","linear_data/linear_data_96"],"report_directory":""}
-        "#;
-
-        let mut state = State {
-            all: false,
-            baseline: "baseline/baseline".into(),
-            ids: Default::default(),
-            throughput: Default::default(),
-            groups: Default::default(),
-        };
-        for line in input.lines() {
-            extract_state(line, &mut state, false);
-        }
-
-        let (_, costs) = state.into_dependent_costs();
-        let linear_data_cost = costs
-            .0
-            .get("linear_data")
-            .expect("Expected linear_data cost");
-
-        let expected_cost = DependentCost::HeavyOperation {
-            base: 0,
-            gas_per_unit: 1,
-        };
-        assert!(matches!(linear_data_cost, Cost::Dependent(c) if *c == expected_cost));
-    }
-
-    #[test]
-    fn categorizes_dependent_cost_from_light_linear_function() {
-        // f(x) = (1/5)x + 5
-        let input = r#"
-        {"reason":"benchmark-complete","id":"baseline/baseline","report_directory":"","iteration_count":[41,82,123,164,205,246,287,328,369,410,451,492,533,574,615,656,697,738,779,820,861,902,943,984,1025,1066,1107,1148,1189,1230,1271,1312,1353,1394,1435,1476,1517,1558,1599,1640,1681,1722,1763,1804,1845,1886,1927,1968,2009,2050,2091,2132,2173,2214,2255,2296,2337,2378,2419,2460,2501,2542,2583,2624,2665,2706,2747,2788,2829,2870,2911,2952,2993,3034,3075,3116,3157,3198,3239,3280,3321,3362,3403,3444,3485,3526,3567,3608,3649,3690,3731,3772,3813,3854,3895,3936,3977,4018,4059,4100],"measured_values":[998584.0,1993042.0,2996041.0,3963958.0,4998417.0,5941125.0,6937583.0,7903375.0,8945500.0,10217875.0,11207291.0,12003833.0,12982125.0,13842083.0,14813375.0,15806458.0,16826459.0,17768375.0,18752542.0,19766959.0,20827250.0,21772875.0,22785625.0,23710625.0,24711958.0,25709042.0,26650792.0,27660709.0,28631833.0,29929750.0,31062917.0,31601667.0,32601292.0,33617334.0,34600333.0,35537584.0,36540667.0,37551542.0,38547584.0,39566292.0,40550250.0,41561250.0,42511833.0,43501292.0,44450166.0,45576041.0,46952625.0,47476250.0,48538334.0,49435792.0,50452125.0,51410875.0,52388500.0,54074625.0,54531000.0,55362584.0,56386542.0,57314250.0,58393000.0,59342000.0,60227375.0,61304708.0,62348166.0,63229917.0,64186209.0,66035083.0,66211792.0,67190083.0,68166958.0,69498083.0,70519584.0,71268042.0,72165125.0,73223417.0,74125083.0,75092625.0,76101708.0,77534834.0,78195208.0,79206167.0,80784208.0,81155083.0,82210709.0,83595833.0,83991083.0,85000209.0,86115750.0,87014292.0,87932125.0,88952833.0,89892167.0,90863875.0,92707125.0,92984667.0,94789917.0,94954791.0,95898167.0,96898875.0,97954125.0,99027292.0],"unit":"ns","throughput":[],"typical":{"estimate":24151.94796635033,"lower_bound":24133.120859289535,"upper_bound":24173.17719456765,"unit":"ns"},"mean":{"estimate":24173.277332128688,"lower_bound":24148.82030491546,"upper_bound":24202.309261650902,"unit":"ns"},"median":{"estimate":24122.72952607575,"lower_bound":24113.633562764517,"upper_bound":24136.08614276297,"unit":"ns"},"median_abs_dev":{"estimate":38.939216886584376,"lower_bound":27.52019699617441,"upper_bound":53.46095875411639,"unit":"ns"},"slope":{"estimate":24151.94796635033,"lower_bound":24133.120859289535,"upper_bound":24173.17719456765,"unit":"ns"},"change":{"mean":{"estimate":-0.007781223077718891,"lower_bound":-0.009854181667997675,"upper_bound":-0.005781475911209411,"unit":"%"},"median":{"estimate":-0.007718846536701185,"lower_bound":-0.009635017085901532,"upper_bound":-0.006383464948326423,"unit":"%"},"change":"NoChange"}}
-        {"reason":"group-complete","group_name":"baseline","benchmarks":["baseline/baseline"],"report_directory":""}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_0","report_directory":"","iteration_count":[9,18,27,36,45,54,63,72,81,90,99,108,117,126,135,144,153,162,171,180,189,198,207,216,225,234,243,252,261,270,279,288,297,306,315,324,333,342,351,360,369,378,387,396,405,414,423,432,441,450,459,468,477,486,495,504,513,522,531,540,549,558,567,576,585,594,603,612,621,630,639,648,657,666,675,684,693,702,711,720,729,738,747,756,765,774,783,792,801,810,819,828,837,846,855,864,873,882,891,900],"measured_values":[1101333.0,2166625.0,3244792.0,4326375.0,5414750.0,6502792.0,7598416.0,8660625.0,9768125.0,10831250.0,11959250.0,12992834.0,14126000.0,15420459.0,16477875.0,17613458.0,18560667.0,19577041.0,21067708.0,21692625.0,22863542.0,23955583.0,24905334.0,26009584.0,27093417.0,28161417.0,29286375.0,30367666.0,31450292.0,32564625.0,33628958.0,34705875.0,35819375.0,36952625.0,38058167.0,39109042.0,40102709.0,41235542.0,42282125.0,43346375.0,44483042.0,45528333.0,46662958.0,47781166.0,49272875.0,50291833.0,51060792.0,52540250.0,53171459.0,54300375.0,55387166.0,56367792.0,57454709.0,58529375.0,59642250.0,61105375.0,61874625.0,62891291.0,64072250.0,65123459.0,66208500.0,67478583.0,68874166.0,69426834.0,71037500.0,71523083.0,72694416.0,73863000.0,74794083.0,75952459.0,77021708.0,78090125.0,79177916.0,80391792.0,81347292.0,82956709.0,83623291.0,85139125.0,85780041.0,87277166.0,87889750.0,89005917.0,90053125.0,91096542.0,92302833.0,93342666.0,95155125.0,95648333.0,96496209.0,98471375.0,98730375.0,99909666.0,100753291.0,102879500.0,103055583.0,104135917.0,105866125.0,106464042.0,107447500.0,109113333.0],"unit":"ns","throughput":[{"per_iteration":0,"unit":"elements"}],"typical":{"estimate":120777.14163932811,"lower_bound":120680.69776165739,"upper_bound":120880.7925052394,"unit":"ns"},"mean":{"estimate":120779.25409318134,"lower_bound":120682.15389881228,"upper_bound":120886.99189275739,"unit":"ns"},"median":{"estimate":120595.9477461958,"lower_bound":120550.24932249323,"upper_bound":120657.29803921569,"unit":"ns"},"median_abs_dev":{"estimate":202.48471862739828,"lower_bound":141.18311570216338,"upper_bound":278.41580172380264,"unit":"ns"},"slope":{"estimate":120777.14163932811,"lower_bound":120680.69776165739,"upper_bound":120880.7925052394,"unit":"ns"},"change":{"mean":{"estimate":-0.005026835893504167,"lower_bound":-0.009137736002197454,"upper_bound":-0.0017935573072914153,"unit":"%"},"median":{"estimate":-0.0014807286890632376,"lower_bound":-0.0040016785875601635,"upper_bound":-0.0004575695454654971,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_5","report_directory":"","iteration_count":[7,14,21,28,35,42,49,56,63,70,77,84,91,98,105,112,119,126,133,140,147,154,161,168,175,182,189,196,203,210,217,224,231,238,245,252,259,266,273,280,287,294,301,308,315,322,329,336,343,350,357,364,371,378,385,392,399,406,413,420,427,434,441,448,455,462,469,476,483,490,497,504,511,518,525,532,539,546,553,560,567,574,581,588,595,602,609,616,623,630,637,644,651,658,665,672,679,686,693,700],"measured_values":[1026667.0,2038084.0,3046959.0,4042458.0,5084333.0,6129292.0,7103458.0,8122000.0,9160959.0,10186792.0,11207375.0,12164125.0,13193209.0,14159958.0,15196375.0,16188708.0,17182625.0,18295875.0,19278167.0,20276542.0,21250458.0,22303541.0,23285750.0,24251417.0,25306458.0,26336417.0,27347958.0,28340709.0,29797834.0,30674916.0,31436166.0,32449375.0,33377167.0,34441875.0,35507458.0,36438833.0,37484625.0,38523792.0,39488417.0,40470958.0,41573209.0,42619125.0,43565584.0,44524792.0,46216708.0,46534167.0,47603458.0,48605458.0,50318167.0,50587375.0,51622625.0,52614792.0,54294500.0,54916875.0,55882333.0,56954083.0,57837917.0,58720833.0,59869583.0,60846417.0,61840000.0,62814125.0,63918958.0,65373250.0,65778542.0,66812125.0,70741291.0,77854459.0,71567125.0,71259417.0,72307750.0,72958375.0,73927041.0,75194958.0,75937125.0,76918541.0,79067541.0,81892750.0,82113917.0,82159209.0,82329791.0,85101209.0,84129291.0,85033958.0,86162250.0,87065416.0,88185333.0,89084083.0,90110084.0,91146875.0,93413708.0,93102500.0,94595917.0,95487000.0,96145666.0,97171125.0,98239833.0,99236417.0,100257166.0,101377500.0],"unit":"ns","throughput":[{"per_iteration":5,"unit":"elements"}],"typical":{"estimate":145583.38429225865,"lower_bound":145094.78831599656,"upper_bound":146265.60915773132,"unit":"ns"},"mean":{"estimate":145404.52570182245,"lower_bound":145063.77523998573,"upper_bound":145879.26724024347,"unit":"ns"},"median":{"estimate":144824.67798594848,"lower_bound":144721.02117874176,"upper_bound":144940.947845805,"unit":"ns"},"median_abs_dev":{"estimate":337.5570569267652,"lower_bound":217.53298673035482,"upper_bound":489.0108913183356,"unit":"ns"},"slope":{"estimate":145583.38429225865,"lower_bound":145094.78831599656,"upper_bound":146265.60915773132,"unit":"ns"},"change":{"mean":{"estimate":0.0009225279216704685,"lower_bound":-0.0018496809221638236,"upper_bound":0.004202015789273443,"unit":"%"},"median":{"estimate":-0.0002428463328854047,"lower_bound":-0.002307690452197697,"upper_bound":0.0005940970760707209,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_10","report_directory":"","iteration_count":[6,12,18,24,30,36,42,48,54,60,66,72,78,84,90,96,102,108,114,120,126,132,138,144,150,156,162,168,174,180,186,192,198,204,210,216,222,228,234,240,246,252,258,264,270,276,282,288,294,300,306,312,318,324,330,336,342,348,354,360,366,372,378,384,390,396,402,408,414,420,426,432,438,444,450,456,462,468,474,480,486,492,498,504,510,516,522,528,534,540,546,552,558,564,570,576,582,588,594,600],"measured_values":[1034459.0,2060167.0,3055292.0,4071541.0,5037625.0,6107667.0,7237459.0,8146083.0,9188291.0,10163541.0,11149917.0,12213792.0,13204584.0,14161875.0,15192708.0,16254291.0,17215250.0,18182083.0,19532541.0,21700000.0,21570875.0,22347500.0,23738708.0,25543875.0,26376875.0,27152916.0,28117542.0,29178125.0,29813208.0,31085875.0,32075750.0,32901833.0,34197792.0,35243625.0,36150500.0,37256458.0,38064625.0,38679958.0,39579666.0,40538625.0,41495584.0,42488417.0,43555167.0,44564083.0,45526209.0,46530209.0,48594958.0,50015833.0,50668500.0,51384167.0,52576750.0,54120500.0,59130542.0,62695208.0,62785917.0,62537542.0,61400750.0,59529750.0,60153208.0,60937875.0,62366125.0,62777291.0,63733875.0,64874334.0,65841042.0,66920750.0,68736833.0,69421333.0,69980125.0,71454333.0,71923959.0,72917208.0,73916708.0,74874334.0,75910875.0,77006667.0,77948208.0,78998333.0,79953458.0,81602167.0,82515583.0,83117208.0,84763625.0,85138334.0,86068542.0,87032875.0,88078750.0,89945708.0,90180125.0,91200000.0,92314833.0,94227167.0,94200416.0,95361083.0,96773458.0,97236125.0,98345334.0,99189083.0,100276666.0,101305458.0],"unit":"ns","throughput":[{"per_iteration":10,"unit":"elements"}],"typical":{"estimate":170281.3246037141,"lower_bound":169624.26908754747,"upper_bound":171146.75291814984,"unit":"ns"},"mean":{"estimate":171106.60226090372,"lower_bound":170331.11577633812,"upper_bound":172028.90304272223,"unit":"ns"},"median":{"estimate":169641.7708333333,"lower_bound":169034.11835748793,"upper_bound":170129.36428571428,"unit":"ns"},"median_abs_dev":{"estimate":1330.0000478606792,"lower_bound":589.1359397647208,"upper_bound":1966.509277104527,"unit":"ns"},"slope":{"estimate":170281.3246037141,"lower_bound":169624.26908754747,"upper_bound":171146.75291814984,"unit":"ns"},"change":{"mean":{"estimate":-0.087343734845722,"lower_bound":-0.11550642707942371,"upper_bound":-0.06121228142764275,"unit":"%"},"median":{"estimate":-0.043596714837708994,"lower_bound":-0.05488027509821902,"upper_bound":-0.03802814178377589,"unit":"%"},"change":"Improved"}}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_15","report_directory":"","iteration_count":[6,12,18,24,30,36,42,48,54,60,66,72,78,84,90,96,102,108,114,120,126,132,138,144,150,156,162,168,174,180,186,192,198,204,210,216,222,228,234,240,246,252,258,264,270,276,282,288,294,300,306,312,318,324,330,336,342,348,354,360,366,372,378,384,390,396,402,408,414,420,426,432,438,444,450,456,462,468,474,480,486,492,498,504,510,516,522,528,534,540,546,552,558,564,570,576,582,588,594,600],"measured_values":[1164625.0,2304583.0,3474666.0,4626667.0,5769917.0,6972708.0,8105167.0,9260250.0,10405666.0,11786833.0,13003750.0,13856416.0,15071125.0,16193958.0,17383375.0,18503708.0,19669750.0,20850458.0,22044917.0,23138125.0,24286958.0,25512542.0,26615792.0,27796750.0,28937833.0,30135667.0,31282542.0,32367333.0,33583292.0,34819167.0,36094541.0,37850833.0,38848875.0,39345250.0,40561250.0,41621750.0,42865125.0,44002292.0,45098542.0,46366750.0,47463042.0,48622625.0,49705792.0,51694000.0,52118709.0,53239417.0,54384750.0,55643750.0,56732875.0,57934125.0,59021291.0,60356958.0,62060500.0,62550042.0,63619583.0,64849125.0,65956292.0,67106834.0,68238083.0,69452208.0,71247916.0,71773750.0,72893333.0,74174750.0,75266708.0,76537208.0,78284875.0,84262500.0,87422292.0,81392166.0,82269708.0,83416250.0,84424792.0,86184875.0,87141541.0,87998375.0,89116000.0,90325209.0,92375542.0,92489209.0,93823417.0,95112458.0,96255416.0,97262750.0,98399625.0,100351958.0,100853916.0,101844875.0,103801750.0,104219542.0,105564875.0,106411250.0,107657417.0,108936667.0,109959792.0,111639708.0,112296875.0,114099417.0,114858709.0,117450250.0],"unit":"ns","throughput":[{"per_iteration":15,"unit":"elements"}],"typical":{"estimate":193829.08066696223,"lower_bound":193304.10569653957,"upper_bound":194549.92096762,"unit":"ns"},"mean":{"estimate":193668.42611830097,"lower_bound":193270.62672830338,"upper_bound":194197.17732963018,"unit":"ns"},"median":{"estimate":193032.62083333335,"lower_bound":192974.56462585035,"upper_bound":193131.18055555556,"unit":"ns"},"median_abs_dev":{"estimate":289.1118063519352,"lower_bound":220.00915669505002,"upper_bound":424.55335651399736,"unit":"ns"},"slope":{"estimate":193829.08066696223,"lower_bound":193304.10569653957,"upper_bound":194549.92096762,"unit":"ns"},"change":{"mean":{"estimate":-0.004958255386169563,"lower_bound":-0.00723700826142473,"upper_bound":-0.0018778129582542155,"unit":"%"},"median":{"estimate":-0.0069146406320230724,"lower_bound":-0.008933829284363481,"upper_bound":-0.005797595399657629,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_20","report_directory":"","iteration_count":[5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80,85,90,95,100,105,110,115,120,125,130,135,140,145,150,155,160,165,170,175,180,185,190,195,200,205,210,215,220,225,230,235,240,245,250,255,260,265,270,275,280,285,290,295,300,305,310,315,320,325,330,335,340,345,350,355,360,365,370,375,380,385,390,395,400,405,410,415,420,425,430,435,440,445,450,455,460,465,470,475,480,485,490,495,500],"measured_values":[1097667.0,2181625.0,3267666.0,4331792.0,5434958.0,6533500.0,7613209.0,8660875.0,9777083.0,10824375.0,11957833.0,12995416.0,14085041.0,15159458.0,16303917.0,17338959.0,18444500.0,19575500.0,20699250.0,22048625.0,23387042.0,24108792.0,25060625.0,26089125.0,27158500.0,28189334.0,29405125.0,30453667.0,31535500.0,32572458.0,33648542.0,34708709.0,35776250.0,36862000.0,37907125.0,39188083.0,40914167.0,41940084.0,42375250.0,43415625.0,44515792.0,45684959.0,46711083.0,47725125.0,48836541.0,49935708.0,51063125.0,52117000.0,53218041.0,54296458.0,55906583.0,56999375.0,58889500.0,59883667.0,61506083.0,62485084.0,63174708.0,64253416.0,65520541.0,66597292.0,67711417.0,68909167.0,70644083.0,71402292.0,71391417.0,72343292.0,73171125.0,75422208.0,76323000.0,77425292.0,78918125.0,79657458.0,80511875.0,82072875.0,81986917.0,84107084.0,85596250.0,86558833.0,86652125.0,87369583.0,88337292.0,90556416.0,91987625.0,92940292.0,93874167.0,94328834.0,95187291.0,97702417.0,98307750.0,100117958.0,100536958.0,102207709.0,102907458.0,104257459.0,105129917.0,107419875.0,108646917.0,108517625.0,110832625.0,111498250.0],"unit":"ns","throughput":[{"per_iteration":20,"unit":"elements"}],"typical":{"estimate":220934.84288044923,"lower_bound":220451.99896925455,"upper_bound":221378.99166706245,"unit":"ns"},"mean":{"estimate":219490.12738849476,"lower_bound":219039.70271926443,"upper_bound":219948.31248054525,"unit":"ns"},"median":{"estimate":219196.46666666667,"lower_bound":217844.4,"upper_bound":220737.28421052633,"unit":"ns"},"median_abs_dev":{"estimate":3083.024208526265,"lower_bound":1638.9691515065438,"upper_bound":3457.774605583651,"unit":"ns"},"slope":{"estimate":220934.84288044923,"lower_bound":220451.99896925455,"upper_bound":221378.99166706245,"unit":"ns"},"change":{"mean":{"estimate":-0.00867978321349061,"lower_bound":-0.011053886284249522,"upper_bound":-0.006265004161814957,"unit":"%"},"median":{"estimate":-0.00972665301978215,"lower_bound":-0.0161115977033659,"upper_bound":-0.0027887876840928527,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_25","report_directory":"","iteration_count":[4,8,12,16,20,24,28,32,36,40,44,48,52,56,60,64,68,72,76,80,84,88,92,96,100,104,108,112,116,120,124,128,132,136,140,144,148,152,156,160,164,168,172,176,180,184,188,192,196,200,204,208,212,216,220,224,228,232,236,240,244,248,252,256,260,264,268,272,276,280,284,288,292,296,300,304,308,312,316,320,324,328,332,336,340,344,348,352,356,360,364,368,372,376,380,384,388,392,396,400],"measured_values":[974542.0,1920958.0,2964709.0,3866333.0,4826875.0,5816208.0,6844292.0,7838208.0,8768792.0,9762541.0,10610333.0,11596875.0,12522292.0,13509166.0,14537958.0,15439000.0,16446125.0,17420583.0,18305917.0,19283333.0,20275875.0,21311334.0,22170250.0,23153209.0,24161334.0,25078000.0,26041125.0,27002000.0,28029292.0,29329375.0,30188208.0,30966625.0,32161458.0,33034208.0,33787125.0,34746000.0,35699667.0,36593417.0,37634750.0,38616458.0,39572625.0,40484625.0,42206500.0,42579042.0,43498083.0,44662584.0,45799500.0,46911459.0,47295583.0,48242334.0,49168750.0,50132083.0,51086125.0,52429292.0,53581125.0,54240959.0,55345917.0,56014167.0,56912459.0,57884833.0,58889750.0,59796959.0,61247666.0,62202292.0,63066250.0,63913833.0,64657334.0,65602459.0,66565250.0,67448625.0,69061625.0,69658917.0,70951667.0,71497375.0,72333375.0,73531917.0,74255292.0,75263209.0,76818292.0,77984292.0,78246208.0,79260542.0,79987125.0,81465500.0,82306542.0,83441208.0,84057833.0,84920625.0,85790667.0,86808500.0,87974542.0,89539333.0,90147209.0,90744000.0,91693458.0,93301167.0,93683792.0,95159667.0,95641083.0,96440084.0],"unit":"ns","throughput":[{"per_iteration":25,"unit":"elements"}],"typical":{"estimate":241893.60654499778,"lower_bound":241688.51904173096,"upper_bound":242111.5100477862,"unit":"ns"},"mean":{"estimate":242006.53088705987,"lower_bound":241788.6470109987,"upper_bound":242242.75097713672,"unit":"ns"},"median":{"estimate":241545.34146862378,"lower_bound":241340.4255319149,"upper_bound":241871.23958333334,"unit":"ns"},"median_abs_dev":{"estimate":676.6488675505103,"lower_bound":425.0192769456497,"upper_bound":1080.9438315118518,"unit":"ns"},"slope":{"estimate":241893.60654499778,"lower_bound":241688.51904173096,"upper_bound":242111.5100477862,"unit":"ns"},"change":{"mean":{"estimate":-0.003881236282179068,"lower_bound":-0.0054498590147517845,"upper_bound":-0.0023578814665513176,"unit":"%"},"median":{"estimate":-0.0046877896498657146,"lower_bound":-0.00606849426264533,"upper_bound":-0.003236551783506772,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_30","report_directory":"","iteration_count":[4,8,12,16,20,24,28,32,36,40,44,48,52,56,60,64,68,72,76,80,84,88,92,96,100,104,108,112,116,120,124,128,132,136,140,144,148,152,156,160,164,168,172,176,180,184,188,192,196,200,204,208,212,216,220,224,228,232,236,240,244,248,252,256,260,264,268,272,276,280,284,288,292,296,300,304,308,312,316,320,324,328,332,336,340,344,348,352,356,360,364,368,372,376,380,384,388,392,396,400],"measured_values":[1069458.0,2120458.0,3190167.0,4268667.0,5441750.0,6381125.0,7498000.0,8620334.0,9690625.0,10771958.0,11761667.0,12808334.0,13775250.0,14851750.0,15972375.0,16995750.0,18009750.0,19104084.0,20194208.0,21301708.0,22348750.0,23339208.0,24428334.0,25478334.0,26501417.0,27581334.0,28634459.0,29646208.0,30714792.0,31912375.0,33066209.0,34418333.0,35117708.0,36093083.0,37083458.0,38221583.0,39905375.0,40528791.0,41725583.0,42725875.0,43488375.0,44578541.0,45653708.0,46757750.0,48271500.0,48910666.0,49840250.0,50874125.0,51950916.0,52968250.0,54126958.0,55163875.0,56599000.0,57758917.0,58573084.0,59431458.0,60518500.0,62190666.0,63151916.0,63728375.0,64686542.0,65923458.0,67404625.0,67958375.0,69108834.0,70045292.0,71078500.0,72078166.0,73170792.0,74849625.0,75452417.0,76294000.0,78271291.0,78562083.0,80704958.0,81260041.0,82037417.0,82663000.0,83789917.0,84819958.0,85975000.0,87562458.0,88419416.0,89265125.0,91128708.0,91626000.0,92386292.0,94030166.0,94399875.0,95908667.0,96676250.0,97693375.0,99209292.0,100001375.0,101299958.0,102574333.0,102870250.0,104373958.0,105421666.0,106151041.0],"unit":"ns","throughput":[{"per_iteration":30,"unit":"elements"}],"typical":{"estimate":266090.9201655091,"lower_bound":265856.4727364091,"upper_bound":266337.81944724306,"unit":"ns"},"mean":{"estimate":266194.29619126423,"lower_bound":265944.70603047387,"upper_bound":266463.7823566406,"unit":"ns"},"median":{"estimate":265695.393902891,"lower_bound":265451.57263062545,"upper_bound":266124.7462121212,"unit":"ns"},"median_abs_dev":{"estimate":870.5058697640995,"lower_bound":601.024567293608,"upper_bound":1311.1975050347867,"unit":"ns"},"slope":{"estimate":266090.9201655091,"lower_bound":265856.4727364091,"upper_bound":266337.81944724306,"unit":"ns"},"change":{"mean":{"estimate":-0.0007781778448620047,"lower_bound":-0.002246131698382483,"upper_bound":0.000498159665871511,"unit":"%"},"median":{"estimate":-0.0010432311226999946,"lower_bound":-0.0035035601186417242,"upper_bound":0.0006664386549208867,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_35","report_directory":"","iteration_count":[4,8,12,16,20,24,28,32,36,40,44,48,52,56,60,64,68,72,76,80,84,88,92,96,100,104,108,112,116,120,124,128,132,136,140,144,148,152,156,160,164,168,172,176,180,184,188,192,196,200,204,208,212,216,220,224,228,232,236,240,244,248,252,256,260,264,268,272,276,280,284,288,292,296,300,304,308,312,316,320,324,328,332,336,340,344,348,352,356,360,364,368,372,376,380,384,388,392,396,400],"measured_values":[1171625.0,2310959.0,3466500.0,4635291.0,5781458.0,7127292.0,8152667.0,9396584.0,10566500.0,11758625.0,12836542.0,13931708.0,15009125.0,16197084.0,17376458.0,18565500.0,19661792.0,20823542.0,22005958.0,23198375.0,24356500.0,25448500.0,26753958.0,28080166.0,29164000.0,30115959.0,31329500.0,32402250.0,33597542.0,35044917.0,36190625.0,37099375.0,38413125.0,39321000.0,41146459.0,41689250.0,42856125.0,43986875.0,45167708.0,46360500.0,47478833.0,48873166.0,50088916.0,50964541.0,52050333.0,53214083.0,54377834.0,56333167.0,56825792.0,57904792.0,59115125.0,60715166.0,61413792.0,62501459.0,64423375.0,64869833.0,65915584.0,67114541.0,68335500.0,70136375.0,70868625.0,71861166.0,73373000.0,74583375.0,75356041.0,76390917.0,78289500.0,78836916.0,79886833.0,81643417.0,82432250.0,83713333.0,85088875.0,85717666.0,86856667.0,88767250.0,89226458.0,90146625.0,92117916.0,92656791.0,93774250.0,95649625.0,96030583.0,97815458.0,98781459.0,99829250.0,101356291.0,101821875.0,103496625.0,104534667.0,105271291.0,107384458.0,107701041.0,109439500.0,110244667.0,111797875.0,112399208.0,113435000.0,115243375.0,115912958.0],"unit":"ns","throughput":[{"per_iteration":35,"unit":"elements"}],"typical":{"estimate":290328.74127826217,"lower_bound":290091.23536442284,"upper_bound":290573.79902574344,"unit":"ns"},"mean":{"estimate":290438.6736695659,"lower_bound":290172.6158745346,"upper_bound":290728.3617041126,"unit":"ns"},"median":{"estimate":289834.8970552884,"lower_bound":289633.9827586207,"upper_bound":290243.9166666667,"unit":"ns"},"median_abs_dev":{"estimate":911.2973573184938,"lower_bound":606.9443853452726,"upper_bound":1369.3981094525345,"unit":"ns"},"slope":{"estimate":290328.74127826217,"lower_bound":290091.23536442284,"upper_bound":290573.79902574344,"unit":"ns"},"change":{"mean":{"estimate":-0.0015577953190506477,"lower_bound":-0.0036369335070613597,"upper_bound":0.00021211193989265613,"unit":"%"},"median":{"estimate":-0.0008268406603817446,"lower_bound":-0.002646847224856308,"upper_bound":0.0007765761619880074,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_40","report_directory":"","iteration_count":[4,8,12,16,20,24,28,32,36,40,44,48,52,56,60,64,68,72,76,80,84,88,92,96,100,104,108,112,116,120,124,128,132,136,140,144,148,152,156,160,164,168,172,176,180,184,188,192,196,200,204,208,212,216,220,224,228,232,236,240,244,248,252,256,260,264,268,272,276,280,284,288,292,296,300,304,308,312,316,320,324,328,332,336,340,344,348,352,356,360,364,368,372,376,380,384,388,392,396,400],"measured_values":[1264375.0,2587916.0,3783000.0,5099334.0,6319375.0,7624125.0,8855708.0,10196250.0,11436625.0,12641458.0,13770416.0,15029625.0,16305333.0,17581667.0,18802292.0,20056250.0,21288666.0,22605458.0,23852541.0,25065209.0,26358792.0,27611000.0,28804333.0,30037125.0,31343750.0,32647166.0,33888334.0,35619000.0,36876708.0,37670459.0,38976625.0,40150417.0,41329750.0,42579000.0,43898959.0,45159959.0,46999084.0,47716625.0,49592167.0,50497959.0,51773834.0,52680292.0,53881250.0,55282542.0,56434750.0,57646667.0,58872125.0,60155250.0,61701208.0,63165625.0,64034417.0,65267458.0,66409916.0,67802209.0,69590833.0,71153625.0,72083417.0,72752375.0,73917000.0,75308459.0,76489167.0,77699458.0,78958500.0,81102208.0,81636792.0,82740791.0,84180625.0,85462416.0,87431209.0,88383625.0,89181042.0,90332125.0,91669500.0,92908958.0,94113209.0,96293542.0,96695000.0,97847333.0,100932583.0,101150791.0,102100750.0,103060542.0,103959000.0,107755417.0,106856000.0,108561958.0,109123042.0,110378250.0,112422416.0,113546959.0,114482916.0,115328666.0,116604208.0,117919167.0,119918167.0,120347042.0,122406458.0,123372417.0,125230375.0,125304250.0],"unit":"ns","throughput":[{"per_iteration":40,"unit":"elements"}],"typical":{"estimate":314685.491789567,"lower_bound":314295.1655739863,"upper_bound":315123.1673988458,"unit":"ns"},"mean":{"estimate":314764.4344200046,"lower_bound":314411.07824964845,"upper_bound":315148.88429881074,"unit":"ns"},"median":{"estimate":313922.8280701754,"lower_bound":313773.6097027972,"upper_bound":314200.0588235294,"unit":"ns"},"median_abs_dev":{"estimate":922.3897517321835,"lower_bound":633.7171976356342,"upper_bound":1456.9928048844024,"unit":"ns"},"slope":{"estimate":314685.491789567,"lower_bound":314295.1655739863,"upper_bound":315123.1673988458,"unit":"ns"},"change":{"mean":{"estimate":-0.005793782012290749,"lower_bound":-0.011862173575281148,"upper_bound":-0.001652401762977414,"unit":"%"},"median":{"estimate":-0.0041960455817844355,"lower_bound":-0.005759508187828111,"upper_bound":-0.001589119871464817,"unit":"%"},"change":"NoChange"}}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_45","report_directory":"","iteration_count":[3,6,9,12,15,18,21,24,27,30,33,36,39,42,45,48,51,54,57,60,63,66,69,72,75,78,81,84,87,90,93,96,99,102,105,108,111,114,117,120,123,126,129,132,135,138,141,144,147,150,153,156,159,162,165,168,171,174,177,180,183,186,189,192,195,198,201,204,207,210,213,216,219,222,225,228,231,234,237,240,243,246,249,252,255,258,261,264,267,270,273,276,279,282,285,288,291,294,297,300],"measured_values":[1028875.0,2048125.0,3035416.0,4041542.0,5127833.0,6174083.0,7128250.0,8223042.0,9245333.0,10253750.0,11265208.0,12209417.0,13204166.0,14146042.0,15529125.0,16451167.0,17227167.0,18239125.0,19271542.0,20224709.0,21268875.0,22273375.0,23283166.0,24300625.0,25385625.0,26334000.0,27381958.0,28313917.0,29425917.0,30433417.0,31404917.0,32936125.0,33676875.0,34457875.0,35574625.0,36490042.0,37920500.0,38851292.0,39573708.0,40501750.0,41486334.0,42517458.0,43650458.0,44638708.0,46027417.0,46941208.0,47658125.0,49208750.0,49627042.0,50714583.0,51731000.0,52724458.0,53754375.0,54749083.0,56334500.0,56857083.0,57755375.0,59407000.0,59953500.0,60719041.0,61885292.0,62804042.0,63875250.0,65152333.0,65739958.0,67645917.0,68025708.0,70877459.0,71969000.0,70928792.0,72406791.0,72902625.0,74363667.0,74862084.0,75919500.0,76959875.0,77924833.0,79319333.0,79977625.0,81496125.0,82073583.0,83362500.0,84086250.0,85418666.0,86199042.0,87369583.0,88094875.0,89073167.0,90012584.0,91563583.0,92079625.0,93794291.0,94560833.0,95299750.0,96546834.0,97295208.0,98453958.0,99125166.0,100150208.0,101626416.0],"unit":"ns","throughput":[{"per_iteration":45,"unit":"elements"}],"typical":{"estimate":338608.18201566424,"lower_bound":338219.4019124888,"upper_bound":339091.3464960211,"unit":"ns"},"mean":{"estimate":338990.9167015526,"lower_bound":338594.4127265365,"upper_bound":339428.96690065577,"unit":"ns"},"median":{"estimate":338130.09444444446,"lower_bound":337953.33080040524,"upper_bound":338502.35859165713,"unit":"ns"},"median_abs_dev":{"estimate":1040.0442396514782,"lower_bound":782.8969630239086,"upper_bound":1519.735501561859,"unit":"ns"},"slope":{"estimate":338608.18201566424,"lower_bound":338219.4019124888,"upper_bound":339091.3464960211,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_50","report_directory":"","iteration_count":[3,6,9,12,15,18,21,24,27,30,33,36,39,42,45,48,51,54,57,60,63,66,69,72,75,78,81,84,87,90,93,96,99,102,105,108,111,114,117,120,123,126,129,132,135,138,141,144,147,150,153,156,159,162,165,168,171,174,177,180,183,186,189,192,195,198,201,204,207,210,213,216,219,222,225,228,231,234,237,240,243,246,249,252,255,258,261,264,267,270,273,276,279,282,285,288,291,294,297,300],"measured_values":[1091167.0,2172375.0,3257166.0,4346791.0,5562375.0,6599417.0,7638000.0,8874958.0,9807125.0,11187542.0,12027125.0,13082709.0,14100542.0,15206084.0,16298917.0,17386125.0,18482375.0,20002167.0,21611500.0,22735334.0,24098833.0,24060875.0,25162125.0,26297167.0,27812958.0,28640541.0,30344208.0,31684666.0,33484250.0,33925166.0,34422500.0,35504625.0,36489458.0,37711792.0,38693917.0,39731500.0,41053791.0,42207083.0,43136791.0,44488250.0,44951042.0,46283209.0,47790333.0,48694542.0,49243667.0,50122459.0,51715625.0,52579042.0,53451208.0,55286042.0,56189916.0,56576083.0,57722917.0,58665958.0,60879458.0,60914708.0,61852167.0,63132125.0,64254000.0,65248292.0,66452916.0,67424875.0,68336208.0,70196834.0,70847625.0,72413125.0,73113625.0,74363958.0,75654584.0,78039875.0,78071958.0,78652500.0,80243833.0,81383417.0,82265917.0,83603875.0,85290334.0,86459042.0,86824416.0,87038583.0,88054833.0,89623500.0,90407500.0,92116792.0,93968666.0,94339084.0,95496458.0,96734625.0,98208042.0,98655000.0,99805542.0,100668833.0,101185709.0,102247167.0,103363125.0,104341250.0,106168541.0,107676250.0,108630041.0,109226167.0],"unit":"ns","throughput":[{"per_iteration":50,"unit":"elements"}],"typical":{"estimate":365434.6093561893,"lower_bound":364882.6389623674,"upper_bound":366033.586842262,"unit":"ns"},"mean":{"estimate":366549.4763283235,"lower_bound":365692.3817733334,"upper_bound":367468.55913968105,"unit":"ns"},"median":{"estimate":365565.44566544564,"lower_bound":364650.4303359684,"upper_bound":366410.196347032,"unit":"ns"},"median_abs_dev":{"estimate":4169.779181811816,"lower_bound":2856.1500991141147,"upper_bound":4695.175015658,"unit":"ns"},"slope":{"estimate":365434.6093561893,"lower_bound":364882.6389623674,"upper_bound":366033.586842262,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_55","report_directory":"","iteration_count":[3,6,9,12,15,18,21,24,27,30,33,36,39,42,45,48,51,54,57,60,63,66,69,72,75,78,81,84,87,90,93,96,99,102,105,108,111,114,117,120,123,126,129,132,135,138,141,144,147,150,153,156,159,162,165,168,171,174,177,180,183,186,189,192,195,198,201,204,207,210,213,216,219,222,225,228,231,234,237,240,243,246,249,252,255,258,261,264,267,270,273,276,279,282,285,288,291,294,297,300],"measured_values":[1212125.0,2358625.0,3476416.0,4710625.0,5863792.0,7025708.0,8230666.0,9449666.0,10492958.0,11808833.0,12833583.0,14024666.0,15037666.0,16348958.0,17511250.0,18606000.0,19888417.0,20978292.0,22300250.0,23359250.0,24608291.0,25557833.0,26813834.0,27936417.0,29277458.0,30403083.0,31540084.0,32712458.0,33768333.0,34988375.0,36006166.0,37328166.0,38865791.0,39891875.0,40845000.0,42704958.0,44062959.0,44904084.0,45932375.0,46813000.0,48066875.0,49008500.0,50071375.0,51218291.0,52314167.0,53583125.0,54865791.0,55623833.0,57507791.0,58221250.0,59418167.0,60637500.0,61806500.0,63127541.0,64333542.0,65900708.0,66417458.0,67857333.0,69022083.0,69894750.0,71105416.0,72569834.0,73590458.0,74969375.0,75778416.0,76905041.0,78145333.0,79599834.0,80313000.0,81550000.0,82917959.0,84304792.0,85206375.0,86085958.0,87509375.0,89119375.0,89611417.0,91008291.0,92031041.0,93565250.0,94715875.0,95262833.0,97110583.0,97985875.0,99240208.0,100261250.0,102003041.0,102587375.0,103683708.0,104811167.0,106341250.0,107095208.0,108574916.0,109743125.0,111060833.0,112536375.0,112997583.0,114349708.0,115277334.0,116835209.0],"unit":"ns","throughput":[{"per_iteration":55,"unit":"elements"}],"typical":{"estimate":389179.819612827,"lower_bound":388926.06864004605,"upper_bound":389447.9860272532,"unit":"ns"},"mean":{"estimate":389657.5840513791,"lower_bound":389234.43329048174,"upper_bound":390143.42943651683,"unit":"ns"},"median":{"estimate":389148.146953405,"lower_bound":388835.0625,"upper_bound":389481.20604395604,"unit":"ns"},"median_abs_dev":{"estimate":1248.5581349725308,"lower_bound":961.4717843051,"upper_bound":1622.9478383895437,"unit":"ns"},"slope":{"estimate":389179.819612827,"lower_bound":388926.06864004605,"upper_bound":389447.9860272532,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_60","report_directory":"","iteration_count":[3,6,9,12,15,18,21,24,27,30,33,36,39,42,45,48,51,54,57,60,63,66,69,72,75,78,81,84,87,90,93,96,99,102,105,108,111,114,117,120,123,126,129,132,135,138,141,144,147,150,153,156,159,162,165,168,171,174,177,180,183,186,189,192,195,198,201,204,207,210,213,216,219,222,225,228,231,234,237,240,243,246,249,252,255,258,261,264,267,270,273,276,279,282,285,288,291,294,297,300],"measured_values":[1238000.0,2451208.0,3767625.0,4902583.0,6388625.0,7460792.0,8779791.0,9963667.0,11312000.0,12430041.0,13638917.0,14920667.0,16156041.0,17411208.0,18605333.0,19801291.0,21074583.0,22417709.0,23518708.0,24793667.0,25935167.0,27316833.0,28539875.0,29698042.0,30872083.0,32121709.0,33482833.0,34779167.0,35918541.0,37052042.0,38766167.0,39574625.0,41004292.0,42349250.0,43341167.0,44614375.0,46116375.0,47489917.0,48357667.0,49538666.0,50734583.0,52072125.0,53301583.0,54592500.0,56043541.0,57202250.0,58246334.0,59451458.0,61183375.0,62075292.0,63375291.0,64527708.0,65631333.0,66868834.0,68632167.0,69602291.0,70711958.0,71866583.0,73029667.0,74374416.0,75633584.0,76591334.0,77734792.0,79697250.0,80186792.0,83023417.0,83073625.0,84730709.0,85476958.0,86763416.0,87838667.0,89269042.0,90540500.0,92012917.0,93306291.0,94336083.0,95696334.0,97380125.0,98674083.0,99168958.0,100236667.0,101685750.0,102539917.0,104265375.0,105211667.0,107690834.0,109501041.0,109967125.0,110755083.0,111636792.0,112651708.0,113940458.0,115029750.0,116805125.0,117622833.0,118953084.0,121143125.0,121606917.0,122691167.0,124451667.0],"unit":"ns","throughput":[{"per_iteration":60,"unit":"elements"}],"typical":{"estimate":413974.173382592,"lower_bound":413536.006626821,"upper_bound":414457.1098704824,"unit":"ns"},"mean":{"estimate":413958.5087765462,"lower_bound":413544.8956447477,"upper_bound":414419.79365861166,"unit":"ns"},"median":{"estimate":413397.62464062235,"lower_bound":413215.88749999995,"upper_bound":413793.38404761907,"unit":"ns"},"median_abs_dev":{"estimate":1282.8683344361036,"lower_bound":924.9867426691272,"upper_bound":1622.2097913106709,"unit":"ns"},"slope":{"estimate":413974.173382592,"lower_bound":413536.006626821,"upper_bound":414457.1098704824,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_65","report_directory":"","iteration_count":[3,6,9,12,15,18,21,24,27,30,33,36,39,42,45,48,51,54,57,60,63,66,69,72,75,78,81,84,87,90,93,96,99,102,105,108,111,114,117,120,123,126,129,132,135,138,141,144,147,150,153,156,159,162,165,168,171,174,177,180,183,186,189,192,195,198,201,204,207,210,213,216,219,222,225,228,231,234,237,240,243,246,249,252,255,258,261,264,267,270,273,276,279,282,285,288,291,294,297,300],"measured_values":[1347583.0,2639917.0,3968042.0,5204000.0,6566459.0,7953333.0,9204959.0,10629750.0,11952458.0,13210250.0,14514500.0,15844208.0,16998083.0,18353542.0,19671417.0,21021958.0,22295333.0,23748916.0,24956917.0,26117125.0,27530667.0,28859084.0,30262584.0,31540625.0,32863750.0,34069959.0,35666250.0,37130958.0,38045625.0,39332500.0,40601667.0,41885292.0,43488916.0,45318208.0,46411792.0,47503458.0,48429875.0,49948750.0,51368750.0,52483333.0,54961417.0,56218917.0,57197833.0,58144208.0,59297208.0,60380500.0,61532792.0,62595792.0,63845375.0,65316292.0,66604541.0,68843417.0,69541334.0,71755667.0,73145625.0,73379875.0,74818084.0,76031459.0,77175667.0,79089667.0,80146417.0,81120667.0,82738500.0,84032292.0,85743333.0,86624875.0,87666375.0,89108667.0,90501458.0,91758958.0,93362083.0,94821625.0,96015500.0,97040125.0,98291583.0,100195750.0,100834041.0,101743000.0,103784292.0,104279292.0,105569625.0,109634792.0,110055166.0,109564166.0,111332208.0,112540334.0,113422125.0,115129125.0,116863458.0,117462958.0,118545250.0,119916208.0,121286333.0,123286208.0,123744000.0,126733375.0,127022542.0,127656875.0,128937167.0,130344542.0],"unit":"ns","throughput":[{"per_iteration":65,"unit":"elements"}],"typical":{"estimate":437074.99229496083,"lower_bound":436437.9709461356,"upper_bound":437809.2018118485,"unit":"ns"},"mean":{"estimate":438128.0110258863,"lower_bound":437554.22552669863,"upper_bound":438731.99512390274,"unit":"ns"},"median":{"estimate":437433.1245950067,"lower_bound":437027.77777777775,"upper_bound":437957.4583333333,"unit":"ns"},"median_abs_dev":{"estimate":2372.7998835154085,"lower_bound":1545.4367010340497,"upper_bound":3346.4914630865856,"unit":"ns"},"slope":{"estimate":437074.99229496083,"lower_bound":436437.9709461356,"upper_bound":437809.2018118485,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_70","report_directory":"","iteration_count":[3,6,9,12,15,18,21,24,27,30,33,36,39,42,45,48,51,54,57,60,63,66,69,72,75,78,81,84,87,90,93,96,99,102,105,108,111,114,117,120,123,126,129,132,135,138,141,144,147,150,153,156,159,162,165,168,171,174,177,180,183,186,189,192,195,198,201,204,207,210,213,216,219,222,225,228,231,234,237,240,243,246,249,252,255,258,261,264,267,270,273,276,279,282,285,288,291,294,297,300],"measured_values":[1399334.0,2761875.0,4207667.0,5601709.0,6937000.0,8422292.0,9721125.0,11139250.0,12506959.0,13750834.0,15172542.0,16471334.0,17934667.0,19457458.0,20877792.0,22420708.0,23576583.0,24950875.0,26367833.0,27784209.0,29038334.0,30479834.0,31830042.0,33123625.0,34666625.0,35952750.0,37347042.0,39005792.0,40258583.0,41678500.0,44133709.0,46273834.0,46950792.0,47515875.0,49411167.0,50509292.0,52064375.0,53288292.0,54091416.0,55569584.0,57562541.0,58778416.0,60307792.0,62587166.0,63060375.0,65355458.0,66224250.0,67584167.0,68543833.0,69439959.0,70620083.0,72679791.0,73360625.0,74770875.0,77076375.0,79008334.0,80376709.0,81315167.0,82721459.0,83555667.0,84597333.0,85918625.0,88340250.0,89354666.0,90825292.0,92279250.0,93932500.0,94747000.0,96378958.0,97983959.0,99201125.0,102112458.0,101274167.0,102527125.0,103859125.0,105595834.0,107313542.0,107967334.0,109449166.0,110386083.0,112267792.0,113249584.0,115603917.0,116704333.0,118993625.0,120078458.0,121566042.0,121806500.0,123251583.0,124335375.0,126030458.0,127727959.0,129062750.0,131096667.0,131607250.0,133086666.0,134306167.0,135532542.0,137102417.0,138489250.0],"unit":"ns","throughput":[{"per_iteration":70,"unit":"elements"}],"typical":{"estimate":463936.4485946505,"lower_bound":463285.24926728575,"upper_bound":464692.97260963003,"unit":"ns"},"mean":{"estimate":464593.8055685307,"lower_bound":463845.71697222994,"upper_bound":465401.5682779132,"unit":"ns"},"median":{"estimate":463180.1632553606,"lower_bound":462593.56140350876,"upper_bound":464974.8366477273,"unit":"ns"},"median_abs_dev":{"estimate":3297.9982374959436,"lower_bound":2198.999020050892,"upper_bound":4162.803550916836,"unit":"ns"},"slope":{"estimate":463936.4485946505,"lower_bound":463285.24926728575,"upper_bound":464692.97260963003,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_75","report_directory":"","iteration_count":[3,6,9,12,15,18,21,24,27,30,33,36,39,42,45,48,51,54,57,60,63,66,69,72,75,78,81,84,87,90,93,96,99,102,105,108,111,114,117,120,123,126,129,132,135,138,141,144,147,150,153,156,159,162,165,168,171,174,177,180,183,186,189,192,195,198,201,204,207,210,213,216,219,222,225,228,231,234,237,240,243,246,249,252,255,258,261,264,267,270,273,276,279,282,285,288,291,294,297,300],"measured_values":[1455417.0,2906167.0,4430417.0,5815958.0,7314167.0,8723125.0,10472542.0,11890666.0,13212791.0,14615916.0,15993583.0,17467500.0,18835292.0,20341875.0,21736417.0,23252083.0,24748791.0,26305416.0,27657041.0,29240167.0,30642500.0,31971166.0,33381750.0,34958291.0,36607916.0,37846875.0,39409208.0,40788333.0,42320666.0,43803750.0,45324834.0,46671834.0,48002667.0,50295791.0,52538792.0,54202583.0,53965042.0,55301750.0,57439875.0,58276916.0,59712792.0,60954667.0,62757834.0,64247291.0,65712541.0,67447000.0,68517000.0,69809375.0,71440584.0,73072583.0,74632292.0,75637458.0,77150167.0,79306208.0,80882917.0,81879333.0,83343209.0,84880750.0,86270167.0,87706291.0,89301292.0,90604666.0,92151958.0,93447000.0,94758375.0,96653583.0,97785125.0,99591500.0,101162292.0,102000666.0,103675292.0,105019750.0,107071375.0,107846250.0,111036875.0,111751750.0,112968666.0,113678958.0,115324916.0,116556000.0,118667500.0,120128334.0,121010791.0,122447292.0,124366208.0,125218708.0,127367792.0,128224333.0,129680917.0,131764750.0,132426417.0,134601334.0,148488167.0,138417166.0,139816542.0,140390208.0,142623708.0,145121166.0,144268625.0,145633083.0],"unit":"ns","throughput":[{"per_iteration":75,"unit":"elements"}],"typical":{"estimate":488758.8463070785,"lower_bound":487165.71928171144,"upper_bound":491402.0770777346,"unit":"ns"},"mean":{"estimate":487846.92375878466,"lower_bound":486930.53594555764,"upper_bound":489056.2417513877,"unit":"ns"},"median":{"estimate":486705.7291666666,"lower_bound":486096.45682161755,"upper_bound":487336.11666666664,"unit":"ns"},"median_abs_dev":{"estimate":1931.3365242847851,"lower_bound":1500.5488951004459,"upper_bound":2478.9724145460127,"unit":"ns"},"slope":{"estimate":488758.8463070785,"lower_bound":487165.71928171144,"upper_bound":491402.0770777346,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_80","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1026541.0,2026083.0,3041583.0,4061000.0,5162125.0,6171334.0,7217125.0,8193459.0,9265083.0,10250666.0,11350333.0,12357375.0,13339875.0,14330750.0,15271500.0,16233584.0,17244333.0,18295584.0,19344917.0,20353500.0,21344250.0,22348375.0,23337875.0,24389125.0,25415917.0,26464375.0,27466041.0,28425959.0,29434125.0,30540417.0,31487458.0,32667000.0,33487500.0,34563542.0,35574959.0,36555958.0,37707667.0,38603959.0,39651833.0,41022792.0,41718458.0,42773500.0,43687541.0,44882208.0,45742416.0,47191666.0,47882500.0,48792417.0,49726250.0,50745333.0,52096875.0,52958542.0,53924625.0,54986750.0,56141541.0,57203000.0,57963917.0,59223750.0,60028291.0,61294750.0,62492375.0,62994208.0,64287333.0,65776416.0,66502750.0,67151292.0,68359792.0,69143042.0,70237917.0,71245917.0,72369417.0,73316125.0,74152417.0,75337375.0,76242458.0,77813833.0,78516875.0,79849125.0,80464000.0,81271417.0,82562833.0,83609167.0,84343750.0,85573333.0,86509375.0,87506542.0,88498250.0,90062708.0,90676667.0,91615458.0,92512584.0,93421291.0,94584708.0,96142625.0,96626667.0,97743416.0,98670250.0,100007250.0,100917583.0,102325917.0],"unit":"ns","throughput":[{"per_iteration":80,"unit":"elements"}],"typical":{"estimate":509458.8335761785,"lower_bound":509129.93748634856,"upper_bound":509804.60326085164,"unit":"ns"},"mean":{"estimate":509652.35787699313,"lower_bound":509253.50315761054,"upper_bound":510082.19557120145,"unit":"ns"},"median":{"estimate":508990.85833333334,"lower_bound":508718.60393348255,"upper_bound":509290.9136904762,"unit":"ns"},"median_abs_dev":{"estimate":1459.803089758503,"lower_bound":1014.0650546438069,"upper_bound":1877.2519055770765,"unit":"ns"},"slope":{"estimate":509458.8335761785,"lower_bound":509129.93748634856,"upper_bound":509804.60326085164,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_85","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1067875.0,2136792.0,3198000.0,4257084.0,5357125.0,6445375.0,7454917.0,8599375.0,9720000.0,10696125.0,11761042.0,12768667.0,13895167.0,14922750.0,16088458.0,17044583.0,18101916.0,19160916.0,20196750.0,21341833.0,22348042.0,23482250.0,24577000.0,25785875.0,27016875.0,27717458.0,28773333.0,29824375.0,30929834.0,32117750.0,33329958.0,34084416.0,35162333.0,36242625.0,37247708.0,38345292.0,39463084.0,40572375.0,41637250.0,42644792.0,43603041.0,44872041.0,45859709.0,46839791.0,47910542.0,49095750.0,49987292.0,51138334.0,52385750.0,53752250.0,54285625.0,55493167.0,56396042.0,58036958.0,58701584.0,59748334.0,60850917.0,61729917.0,62903542.0,64043500.0,65087667.0,66096333.0,67216791.0,68177417.0,69275416.0,71063333.0,71403166.0,72454416.0,73494500.0,74874792.0,75683916.0,76621916.0,77924667.0,78913084.0,79910000.0,81032500.0,82093667.0,83153375.0,84654625.0,85162542.0,86478875.0,87217292.0,89002917.0,89582208.0,90551417.0,91980833.0,92918333.0,93910375.0,94815208.0,96827833.0,96856958.0,97941542.0,99103042.0,100872708.0,101365209.0,102410917.0,103343667.0,104378500.0,105576250.0,106911417.0],"unit":"ns","throughput":[{"per_iteration":85,"unit":"elements"}],"typical":{"estimate":533595.9385458844,"lower_bound":533229.8557439339,"upper_bound":534001.3617036165,"unit":"ns"},"mean":{"estimate":533750.2282900958,"lower_bound":533412.342860269,"upper_bound":534111.1747550188,"unit":"ns"},"median":{"estimate":533204.9486759487,"lower_bound":532985.323943662,"upper_bound":533506.5464285715,"unit":"ns"},"median_abs_dev":{"estimate":947.5288232808921,"lower_bound":765.8112978326777,"upper_bound":1428.5380246385037,"unit":"ns"},"slope":{"estimate":533595.9385458844,"lower_bound":533229.8557439339,"upper_bound":534001.3617036165,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_90","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1126083.0,2221417.0,3326250.0,4469125.0,5643250.0,6696541.0,7837417.0,8989583.0,10123208.0,11231750.0,12453750.0,13476709.0,14466625.0,15730542.0,16894542.0,18037167.0,18932709.0,20266792.0,21152875.0,22327417.0,23448500.0,24538417.0,25627042.0,26720417.0,27796583.0,28927500.0,30226875.0,31235542.0,32249875.0,33461125.0,34567709.0,35798667.0,36989875.0,37848625.0,38997417.0,40129458.0,41225417.0,42685000.0,43504542.0,44666000.0,45801042.0,46775875.0,47822667.0,49296584.0,50562208.0,51218542.0,52282917.0,53436625.0,54522542.0,55717500.0,56759875.0,57861417.0,59065459.0,60155750.0,61310333.0,62437083.0,63646542.0,65153875.0,65793334.0,66827750.0,67980375.0,69483792.0,70129250.0,71280541.0,74197833.0,73894917.0,74981292.0,75824917.0,76875292.0,77960000.0,79065500.0,80635333.0,81262041.0,82542500.0,83878833.0,84711084.0,85738125.0,86879625.0,87992625.0,89148209.0,91014500.0,92954709.0,92896791.0,94175500.0,95678667.0,97444583.0,98121000.0,99020834.0,99921416.0,100988250.0,101965417.0,103512458.0,105135166.0,105109625.0,106979375.0,108991750.0,113629791.0,119219292.0,117653125.0,114903000.0],"unit":"ns","throughput":[{"per_iteration":90,"unit":"elements"}],"typical":{"estimate":563567.5826791783,"lower_bound":560396.4762950426,"upper_bound":567330.8337794039,"unit":"ns"},"mean":{"estimate":560517.0306519145,"lower_bound":559243.0683570124,"upper_bound":562091.1672339702,"unit":"ns"},"median":{"estimate":558241.5220238096,"lower_bound":557508.824452765,"upper_bound":559659.4125207297,"unit":"ns"},"median_abs_dev":{"estimate":2443.5439288002526,"lower_bound":1458.8364732874788,"upper_bound":3874.9883522392297,"unit":"ns"},"slope":{"estimate":563567.5826791783,"lower_bound":560396.4762950426,"upper_bound":567330.8337794039,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_95","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1211834.0,2366417.0,3492667.0,4713583.0,5826667.0,7065417.0,8165167.0,9478458.0,10527333.0,11798416.0,12856292.0,13930583.0,15201666.0,16333583.0,17434166.0,18550208.0,19754709.0,20919791.0,22056167.0,23226584.0,24455958.0,25513583.0,26749458.0,27873292.0,28999084.0,30481667.0,31516250.0,32486083.0,33589458.0,35079250.0,36453334.0,37216167.0,38555000.0,39494875.0,40647125.0,41831875.0,43005125.0,44551166.0,45528417.0,46556917.0,47641750.0,48813917.0,49956958.0,51216042.0,52309750.0,53534125.0,54990709.0,55764125.0,57546541.0,58450625.0,59990041.0,62123000.0,61741459.0,63034125.0,63800958.0,65041833.0,66190750.0,67253875.0,68476083.0,69721292.0,70701250.0,71902917.0,73269875.0,74726708.0,75865375.0,77072667.0,78683500.0,79853084.0,81110125.0,81631334.0,86140916.0,88854500.0,89825708.0,88805292.0,87533000.0,89200959.0,89743500.0,90641417.0,92056500.0,93882500.0,100565750.0,101133375.0,101552167.0,99065125.0,99895666.0,100856792.0,101895875.0,103036750.0,103778292.0,105499541.0,105939958.0,107894959.0,108752084.0,110296500.0,111024500.0,112895791.0,113178416.0,114623083.0,116047834.0,116586625.0],"unit":"ns","throughput":[{"per_iteration":95,"unit":"elements"}],"typical":{"estimate":587840.3599778336,"lower_bound":585460.0230130876,"upper_bound":590563.7565775753,"unit":"ns"},"mean":{"estimate":585968.3485338399,"lower_bound":584428.7125919074,"upper_bound":587735.8486895236,"unit":"ns"},"median":{"estimate":583566.5705128205,"lower_bound":582566.6471698113,"upper_bound":584515.5378787878,"unit":"ns"},"median_abs_dev":{"estimate":3846.570513839232,"lower_bound":2790.8877056799597,"upper_bound":4745.219879966042,"unit":"ns"},"slope":{"estimate":587840.3599778336,"lower_bound":585460.0230130876,"upper_bound":590563.7565775753,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_100","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1255167.0,2539625.0,3748584.0,4853917.0,6154209.0,7346833.0,8668958.0,9939792.0,11087000.0,12512250.0,13535000.0,15524334.0,16190875.0,17512959.0,18549875.0,20018334.0,21051958.0,22446541.0,23660541.0,24623416.0,25881875.0,27003500.0,28159875.0,29322875.0,30327250.0,31449458.0,32640625.0,34004167.0,35216542.0,36352125.0,37535084.0,40171792.0,41201000.0,42549250.0,44147166.0,45039833.0,47523709.0,48082792.0,49289291.0,49439584.0,51364750.0,52396292.0,53019209.0,54388000.0,56084750.0,57007875.0,58261625.0,59062208.0,60869000.0,61710084.0,64293625.0,64584458.0,66148958.0,66464959.0,67808333.0,69314000.0,70283416.0,70711167.0,71869958.0,73375375.0,74486125.0,75752792.0,76982000.0,78080292.0,79272250.0,80128500.0,81495834.0,82704500.0,85032292.0,85684375.0,87429291.0,88493791.0,89684208.0,89996208.0,91062292.0,92048709.0,94244334.0,95071083.0,96211459.0,98515291.0,99391334.0,99841542.0,101807000.0,102817208.0,103872417.0,104792917.0,108720750.0,147929375.0,114420083.0,116302750.0,117707917.0,119988000.0,120288083.0,120709875.0,120648167.0,122121083.0,121359625.0,120946917.0,120935917.0,123274959.0],"unit":"ns","throughput":[{"per_iteration":100,"unit":"elements"}],"typical":{"estimate":625503.6121117186,"lower_bound":617590.305261073,"upper_bound":637464.9026971174,"unit":"ns"},"mean":{"estimate":621311.5067816136,"lower_bound":617502.7076021748,"upper_bound":626866.9140366529,"unit":"ns"},"median":{"estimate":616407.0929545455,"lower_bound":615231.3333333334,"upper_bound":619175.2352941176,"unit":"ns"},"median_abs_dev":{"estimate":10233.750705402803,"lower_bound":7667.837182618707,"upper_bound":11910.370197244238,"unit":"ns"},"slope":{"estimate":625503.6121117186,"lower_bound":617590.305261073,"upper_bound":637464.9026971174,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_105","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1266667.0,2559125.0,3819125.0,5187583.0,6326833.0,7619708.0,8987916.0,10119042.0,11511000.0,12718292.0,13956584.0,15151500.0,16491834.0,17821917.0,18896667.0,20270666.0,21513208.0,22722416.0,24211250.0,25278875.0,26506583.0,27767167.0,29138334.0,30267417.0,31558208.0,32774042.0,34877000.0,35555750.0,36704708.0,37938208.0,38856416.0,40093667.0,42055000.0,43104208.0,44260458.0,45522958.0,46740417.0,47973916.0,49734458.0,50897375.0,51622625.0,53049000.0,54096417.0,55763833.0,57110250.0,58924208.0,59518125.0,61356833.0,62246667.0,63282750.0,64346792.0,66325083.0,66829792.0,67829833.0,69552541.0,74495625.0,77120250.0,78080042.0,79969792.0,80996375.0,82166334.0,83250375.0,84412875.0,86434834.0,87097583.0,88351500.0,89080167.0,90052541.0,91062917.0,89576667.0,93136958.0,92691625.0,97197375.0,98871542.0,99988084.0,102248208.0,104720750.0,105706792.0,108859833.0,108037584.0,108699500.0,107631834.0,110268750.0,109785166.0,109592333.0,109521708.0,116312583.0,118410917.0,121027667.0,121652375.0,122607875.0,123158917.0,123881584.0,123157125.0,124649458.0,125766583.0,125167167.0,125581667.0,125862917.0,127987417.0],"unit":"ns","throughput":[{"per_iteration":105,"unit":"elements"}],"typical":{"estimate":657196.3865095315,"lower_bound":653069.0770763678,"upper_bound":661103.610793991,"unit":"ns"},"mean":{"estimate":647442.3738747012,"lower_bound":644047.633765447,"upper_bound":650945.4196411265,"unit":"ns"},"median":{"estimate":638437.4298878205,"lower_bound":635542.3357142857,"upper_bound":644660.7823529412,"unit":"ns"},"median_abs_dev":{"estimate":10892.486656029632,"lower_bound":6763.886364765423,"upper_bound":19799.72075861401,"unit":"ns"},"slope":{"estimate":657196.3865095315,"lower_bound":653069.0770763678,"upper_bound":661103.610793991,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_110","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1317833.0,2611791.0,3919333.0,5234750.0,6598667.0,7995500.0,9297542.0,10541417.0,11987625.0,13199167.0,14418958.0,15754333.0,17043542.0,18417333.0,19700791.0,20984875.0,22365459.0,23663167.0,25246083.0,26424041.0,27550666.0,28880208.0,30237583.0,31412542.0,32761167.0,34209334.0,35401875.0,36699250.0,38087375.0,39394542.0,40772875.0,42083834.0,43405542.0,44707375.0,45939834.0,47237292.0,48772542.0,50121083.0,51769750.0,52528542.0,53633958.0,55176250.0,56843458.0,57673167.0,59002500.0,60365250.0,61592000.0,62869625.0,64290166.0,65569917.0,67133834.0,68158375.0,69748459.0,70947083.0,72253500.0,73858959.0,74626375.0,76055292.0,77454417.0,78403458.0,79692208.0,80797750.0,82223459.0,83440334.0,84929500.0,86239792.0,87428875.0,88703209.0,91039958.0,91798625.0,92732792.0,94064791.0,95304042.0,96811166.0,98039792.0,99233000.0,101065958.0,102823541.0,103758750.0,105808542.0,106289125.0,107668750.0,108839250.0,110040541.0,111317250.0,112993125.0,114101667.0,115618834.0,116775083.0,118277750.0,119201500.0,120721875.0,122013375.0,123079167.0,125129459.0,125907917.0,127571167.0,128907459.0,130078000.0,131124792.0],"unit":"ns","throughput":[{"per_iteration":110,"unit":"elements"}],"typical":{"estimate":655909.9228417319,"lower_bound":655415.8545076227,"upper_bound":656403.1336048342,"unit":"ns"},"mean":{"estimate":656545.2740182516,"lower_bound":656006.5297901374,"upper_bound":657113.7051117475,"unit":"ns"},"median":{"estimate":656124.8658078369,"lower_bound":655737.4326636905,"upper_bound":656678.8793103448,"unit":"ns"},"median_abs_dev":{"estimate":2054.5660596560665,"lower_bound":1349.1351171211534,"upper_bound":2585.2016376185734,"unit":"ns"},"slope":{"estimate":655909.9228417319,"lower_bound":655415.8545076227,"upper_bound":656403.1336048342,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_115","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1369584.0,2697875.0,4102959.0,5445333.0,6817750.0,8170250.0,9525333.0,11100083.0,12516875.0,13622250.0,14871167.0,16274459.0,17622041.0,18932625.0,20310167.0,21615250.0,22992333.0,24301791.0,25636458.0,26992875.0,28363083.0,29817250.0,31052083.0,32421875.0,33769125.0,35109000.0,36458958.0,38301083.0,39440750.0,40588250.0,41914833.0,43357583.0,45181917.0,45920417.0,47331875.0,48595417.0,50013000.0,51338083.0,53150791.0,54706583.0,55464667.0,56768083.0,58230417.0,59563791.0,61038625.0,62347042.0,63496000.0,65475250.0,66294000.0,67783375.0,69712958.0,70301083.0,71540458.0,72941709.0,74622125.0,76334500.0,77097417.0,78461875.0,79810667.0,81080583.0,82424875.0,84574125.0,85337791.0,87402875.0,87849125.0,89104666.0,90477708.0,92390834.0,93753958.0,94569542.0,95930250.0,97664125.0,99265458.0,100808750.0,101491083.0,104482792.0,104324208.0,106360208.0,107176333.0,108101416.0,109471000.0,110642791.0,112943834.0,114127166.0,114858500.0,116143833.0,118104166.0,120940542.0,120154291.0,121531667.0,123414167.0,125525208.0,125604375.0,126919666.0,128455417.0,130054250.0,131675875.0,132407000.0,134312625.0,136082667.0],"unit":"ns","throughput":[{"per_iteration":115,"unit":"elements"}],"typical":{"estimate":677940.2232821044,"lower_bound":677188.5723710472,"upper_bound":678767.246996872,"unit":"ns"},"mean":{"estimate":678215.9637792195,"lower_bound":677505.315913426,"upper_bound":678994.9806629132,"unit":"ns"},"median":{"estimate":676734.2406818182,"lower_bound":676125.392481203,"upper_bound":677770.8076923077,"unit":"ns"},"median_abs_dev":{"estimate":2235.9936459064115,"lower_bound":1524.9564631475384,"upper_bound":3527.3665554090794,"unit":"ns"},"slope":{"estimate":677940.2232821044,"lower_bound":677188.5723710472,"upper_bound":678767.246996872,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_120","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1410500.0,2797041.0,4193541.0,5685083.0,7046542.0,8472667.0,9922750.0,11442916.0,12772083.0,14061542.0,15428042.0,16778416.0,18222500.0,19619375.0,21015125.0,22368000.0,23814625.0,25239042.0,26884417.0,28011625.0,29342875.0,30781875.0,32183833.0,33654416.0,34919458.0,36396000.0,38114083.0,39512875.0,40640209.0,41953125.0,43294292.0,44767417.0,46287125.0,48041708.0,49236833.0,52296042.0,52042334.0,53472334.0,55020000.0,56034125.0,57392500.0,58893959.0,60240833.0,61592791.0,62996166.0,64389459.0,66459708.0,67128709.0,68579583.0,70025125.0,71379459.0,74054208.0,74201000.0,76194416.0,77015208.0,78449792.0,79728041.0,81109458.0,82606792.0,84787791.0,85920500.0,87050333.0,88159583.0,89951542.0,91629125.0,92951209.0,93842792.0,95062250.0,96628875.0,97899792.0,99473875.0,101156125.0,101992125.0,103841417.0,106247709.0,106713750.0,108141250.0,109180375.0,110595125.0,111871292.0,113940667.0,114789750.0,116099625.0,118717042.0,119751250.0,120866291.0,121834542.0,123143917.0,124522333.0,126663333.0,127356750.0,129287416.0,130797875.0,132422917.0,132914917.0,134492959.0,135961083.0,137781916.0,139092708.0,140011292.0],"unit":"ns","throughput":[{"per_iteration":120,"unit":"elements"}],"typical":{"estimate":701778.290326585,"lower_bound":701186.8103951322,"upper_bound":702418.7717546432,"unit":"ns"},"mean":{"estimate":702267.909899716,"lower_bound":701536.5920836306,"upper_bound":703106.941267651,"unit":"ns"},"median":{"estimate":700606.1053822937,"lower_bound":700251.25,"upper_bound":701669.4451368524,"unit":"ns"},"median_abs_dev":{"estimate":2056.362640349855,"lower_bound":1299.6451055874763,"upper_bound":3022.3369074223942,"unit":"ns"},"slope":{"estimate":701778.290326585,"lower_bound":701186.8103951322,"upper_bound":702418.7717546432,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_125","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1461709.0,2897083.0,4348250.0,5772625.0,7295791.0,8856166.0,10274541.0,11653708.0,13250375.0,14526792.0,15943125.0,17424416.0,18774708.0,20591917.0,21894000.0,23167625.0,24661958.0,26053250.0,27505166.0,28970167.0,30367916.0,31897416.0,33242875.0,34655250.0,36150000.0,37635334.0,39453792.0,40786875.0,42240625.0,43788792.0,44975625.0,46412584.0,48051958.0,49711917.0,50705958.0,52101209.0,53514292.0,56098250.0,57018750.0,58129208.0,59565667.0,60923709.0,62293458.0,64029125.0,65105167.0,67146459.0,68442125.0,69698208.0,71018083.0,72265083.0,73795833.0,75369209.0,77329292.0,78363583.0,79968583.0,81396375.0,82534833.0,84079625.0,85671791.0,87722000.0,88534625.0,89743292.0,91350750.0,92611209.0,94538792.0,95690208.0,97611708.0,98611584.0,100538541.0,103737000.0,104743667.0,105537959.0,106877916.0,107574167.0,109637500.0,110300458.0,112362459.0,113576333.0,114929792.0,116993417.0,117372709.0,119062250.0,121664541.0,125305083.0,126446375.0,133396916.0,149038500.0,150762291.0,143018334.0,137967167.0,136375917.0,142356166.0,143751500.0,147834833.0,146125250.0,144716250.0,150986542.0,147947334.0,149340875.0,150314875.0],"unit":"ns","throughput":[{"per_iteration":125,"unit":"elements"}],"typical":{"estimate":747960.3129939411,"lower_bound":739432.528959481,"upper_bound":757003.7111052355,"unit":"ns"},"mean":{"estimate":735584.0975243743,"lower_bound":731427.6375276316,"upper_bound":740434.6147352596,"unit":"ns"},"median":{"estimate":727312.61187926,"lower_bound":726006.1646341463,"upper_bound":728983.6023655308,"unit":"ns"},"median_abs_dev":{"estimate":4927.55396513702,"lower_bound":3311.0015921253676,"upper_bound":6706.04225473095,"unit":"ns"},"slope":{"estimate":747960.3129939411,"lower_bound":739432.528959481,"upper_bound":757003.7111052355,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_130","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1503583.0,2984750.0,4535084.0,5967375.0,7539334.0,9032083.0,10536959.0,12087167.0,13474458.0,14935667.0,16477833.0,17903625.0,19453792.0,20957292.0,22418042.0,23955291.0,25460458.0,26902417.0,28445792.0,29920375.0,31883541.0,32938541.0,34386125.0,35943167.0,37362667.0,38845750.0,40420584.0,41981333.0,43567958.0,45054666.0,46668041.0,48734333.0,49660291.0,51016959.0,52877542.0,53865333.0,55282459.0,56872542.0,58421625.0,59936458.0,61745333.0,62908167.0,64592334.0,65917875.0,67431875.0,68821459.0,70421583.0,71904583.0,73552708.0,75503250.0,76598834.0,77724792.0,79308792.0,80838084.0,82210166.0,83755583.0,85343000.0,87246583.0,88318625.0,89925167.0,91274166.0,93692042.0,95234083.0,97291041.0,97997458.0,99089083.0,100349500.0,102048583.0,103522500.0,105483500.0,106862750.0,108930167.0,110044292.0,111631542.0,113235834.0,114418125.0,115838542.0,117628208.0,119245500.0,120734333.0,122737375.0,122965292.0,124374167.0,126737250.0,128290625.0,128923375.0,130361333.0,132698166.0,134526916.0,135941542.0,137083000.0,138784083.0,140050083.0,142385042.0,143094625.0,144627750.0,146863709.0,147062250.0,148331875.0,150326667.0],"unit":"ns","throughput":[{"per_iteration":130,"unit":"elements"}],"typical":{"estimate":752445.7350480272,"lower_bound":751692.2233062021,"upper_bound":753168.1536759011,"unit":"ns"},"mean":{"estimate":751234.9154860547,"lower_bound":750591.0943504511,"upper_bound":751905.9231761371,"unit":"ns"},"median":{"estimate":750337.4149909965,"lower_bound":749205.725,"upper_bound":752125.7155172414,"unit":"ns"},"median_abs_dev":{"estimate":3432.6293015586107,"lower_bound":2255.0824888435873,"upper_bound":4436.0070336315275,"unit":"ns"},"slope":{"estimate":752445.7350480272,"lower_bound":751692.2233062021,"upper_bound":753168.1536759011,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_135","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1556000.0,3091834.0,4633458.0,6282750.0,7694875.0,9281875.0,11063500.0,12414583.0,14088000.0,15461375.0,16999125.0,18543792.0,20123208.0,21620292.0,23197250.0,24752375.0,26382209.0,28545958.0,29595792.0,31078833.0,32684458.0,34318042.0,35816834.0,37495458.0,38828625.0,40432667.0,42306167.0,43638917.0,45152750.0,46728584.0,48030125.0,49914625.0,51659250.0,53189792.0,54259208.0,56188916.0,57379042.0,58946250.0,60336458.0,61871333.0,63985875.0,64857833.0,66466166.0,68080125.0,69568750.0,71045250.0,72673750.0,74479583.0,76470708.0,77839625.0,79073041.0,80927625.0,82917583.0,83748458.0,85734209.0,87017917.0,88219833.0,90537041.0,91808958.0,93215958.0,95541541.0,96282500.0,98099041.0,99372416.0,100946625.0,103284625.0,104447542.0,106073250.0,106600833.0,108267000.0,110491375.0,112126583.0,113650834.0,115094458.0,115917791.0,118122583.0,119380500.0,121590959.0,123068250.0,124726959.0,125748000.0,127443041.0,129457791.0,130863583.0,132275584.0,133668500.0,135572042.0,136785334.0,138430458.0,140029375.0,141430667.0,142692000.0,143820500.0,146612459.0,148044667.0,149493125.0,150467500.0,152632375.0,155080167.0,155381584.0],"unit":"ns","throughput":[{"per_iteration":135,"unit":"elements"}],"typical":{"estimate":777634.9782399882,"lower_bound":777010.9313944975,"upper_bound":778258.5771711847,"unit":"ns"},"mean":{"estimate":777416.6604373278,"lower_bound":776698.5086190387,"upper_bound":778169.109910842,"unit":"ns"},"median":{"estimate":777370.3430944056,"lower_bound":776572.5,"upper_bound":778269.3222554505,"unit":"ns"},"median_abs_dev":{"estimate":3138.2019436273104,"lower_bound":2408.309243804372,"upper_bound":4371.203159049874,"unit":"ns"},"slope":{"estimate":777634.9782399882,"lower_bound":777010.9313944975,"upper_bound":778258.5771711847,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_140","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1601250.0,3234208.0,4784959.0,6462625.0,7983500.0,9622167.0,11261458.0,12837709.0,14476958.0,15941958.0,17557917.0,19152125.0,20770292.0,22277834.0,23974167.0,25882917.0,27204542.0,28712958.0,30242708.0,31823167.0,33527959.0,35106250.0,37414542.0,38294250.0,40046709.0,41922792.0,43691750.0,45735000.0,46805042.0,47842375.0,49422542.0,51063167.0,52767083.0,54171541.0,55808000.0,57416292.0,58985167.0,61036208.0,62471417.0,64023625.0,65796833.0,67313500.0,69512292.0,71240417.0,71964833.0,73398500.0,74985583.0,76560791.0,78223958.0,79779500.0,81327000.0,83177875.0,84347792.0,86539334.0,87980292.0,90334833.0,91529500.0,92922375.0,94719791.0,96193750.0,97818417.0,99584250.0,100609541.0,102703875.0,104323250.0,105742208.0,108520458.0,109770375.0,111194500.0,112571167.0,114207792.0,115545167.0,119355583.0,120215917.0,121941792.0,124079958.0,127583709.0,128571375.0,129110583.0,128762167.0,129500875.0,131986125.0,133414541.0,138055917.0,142016667.0,143842708.0,141581500.0,145041167.0,144011708.0,145498542.0,146405917.0,147882667.0,149396625.0,150771917.0,152767625.0,154581042.0,156730125.0,156940667.0,158412084.0,160566625.0],"unit":"ns","throughput":[{"per_iteration":140,"unit":"elements"}],"typical":{"estimate":807408.9661149697,"lower_bound":805009.1920674731,"upper_bound":810036.1420578865,"unit":"ns"},"mean":{"estimate":804283.1079988442,"lower_bound":802788.9747838969,"upper_bound":805935.1305143335,"unit":"ns"},"median":{"estimate":802365.41796875,"lower_bound":800934.18,"upper_bound":803405.0271084337,"unit":"ns"},"median_abs_dev":{"estimate":6195.307794475618,"lower_bound":4209.667552284758,"upper_bound":7382.463363253219,"unit":"ns"},"slope":{"estimate":807408.9661149697,"lower_bound":805009.1920674731,"upper_bound":810036.1420578865,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_145","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1670000.0,3414750.0,4953334.0,6941375.0,8423917.0,10113375.0,11797833.0,13373584.0,14972583.0,16586375.0,18317917.0,19819375.0,22060583.0,23299167.0,25015584.0,26585291.0,28130958.0,29886958.0,31397125.0,32983667.0,34798583.0,37655917.0,39722250.0,42580916.0,42768291.0,43434041.0,45683542.0,46721208.0,48240250.0,50388958.0,51531042.0,52940458.0,54610333.0,55879125.0,57773708.0,59150916.0,61017000.0,63271708.0,64332750.0,65871041.0,68087417.0,70037292.0,71573583.0,72825167.0,74716959.0,76212958.0,78179041.0,79627667.0,80981584.0,82686292.0,84286750.0,86276791.0,88480583.0,91396542.0,91605500.0,92778417.0,94250625.0,96388250.0,96754500.0,99199625.0,100717291.0,103304708.0,105384417.0,107101208.0,107890584.0,111661208.0,111421750.0,112930292.0,113864667.0,116865709.0,117764958.0,120056542.0,121140542.0,123104792.0,124703625.0,130257916.0,134499750.0,136452917.0,138154416.0,140379917.0,141879250.0,143946375.0,145043041.0,145424083.0,143965292.0,148300791.0,152221500.0,147850708.0,149268083.0,149496417.0,151303334.0,152216833.0,154398625.0,155734625.0,156497000.0,158799792.0,160417750.0,162026792.0,163953542.0,165595125.0],"unit":"ns","throughput":[{"per_iteration":145,"unit":"elements"}],"typical":{"estimate":839698.8667976947,"lower_bound":835086.7872367871,"upper_bound":844684.2669744967,"unit":"ns"},"mean":{"estimate":837606.8204512465,"lower_bound":834647.9932247243,"upper_bound":840791.0174063304,"unit":"ns"},"median":{"estimate":831431.5485074627,"lower_bound":830146.70609319,"upper_bound":832867.5373900293,"unit":"ns"},"median_abs_dev":{"estimate":6749.209065259338,"lower_bound":5003.93788814759,"upper_bound":9401.43060400719,"unit":"ns"},"slope":{"estimate":839698.8667976947,"lower_bound":835086.7872367871,"upper_bound":844684.2669744967,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_150","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1705417.0,3368875.0,5100583.0,6858334.0,8535917.0,10210917.0,12045209.0,13716958.0,15389375.0,16997417.0,18751041.0,20484167.0,22121167.0,23755125.0,25548541.0,27238125.0,29250833.0,30675375.0,32309750.0,33980167.0,35773542.0,37314208.0,39147875.0,40751750.0,42985292.0,44532125.0,45951416.0,47619958.0,49345292.0,51144083.0,52759166.0,54526666.0,56248500.0,57847750.0,59469084.0,61338417.0,62874667.0,64418083.0,66129750.0,68581042.0,69783166.0,71353000.0,73654750.0,74867666.0,76642417.0,77727417.0,79373250.0,80992541.0,83339333.0,85170042.0,86754917.0,88541750.0,90500958.0,91736541.0,94110917.0,95782417.0,96961917.0,97902709.0,100460333.0,102248375.0,103901916.0,105451292.0,107524959.0,108961208.0,111330250.0,112539167.0,114138458.0,116008500.0,117471333.0,119073209.0,121521709.0,122614000.0,123728709.0,125381041.0,128010500.0,129311167.0,131013292.0,132734125.0,134269750.0,136381333.0,138274041.0,139503458.0,141379750.0,142699666.0,144346750.0,145376708.0,146794125.0,149727958.0,151428792.0,153817625.0,154952791.0,156736875.0,158387750.0,159812000.0,160966042.0,162873333.0,164765125.0,167975375.0,168193834.0,173149834.0],"unit":"ns","throughput":[{"per_iteration":150,"unit":"elements"}],"typical":{"estimate":851224.0028387764,"lower_bound":850195.6645284381,"upper_bound":852442.7502990667,"unit":"ns"},"mean":{"estimate":851406.7998004956,"lower_bound":850704.8237192697,"upper_bound":852127.2629590882,"unit":"ns"},"median":{"estimate":851027.4901908801,"lower_bound":850731.3482142857,"upper_bound":851582.5381720429,"unit":"ns"},"median_abs_dev":{"estimate":2033.920468079883,"lower_bound":1381.2963616183856,"upper_bound":2867.7304715876135,"unit":"ns"},"slope":{"estimate":851224.0028387764,"lower_bound":850195.6645284381,"upper_bound":852442.7502990667,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_155","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1758834.0,3469375.0,5300458.0,6965000.0,8874292.0,10511208.0,12353041.0,14125584.0,15883375.0,17530291.0,19257250.0,20969042.0,22625833.0,24778666.0,26462459.0,28105750.0,29788875.0,31538083.0,33310542.0,35022667.0,36840125.0,38396750.0,40235125.0,41917667.0,43599917.0,45452375.0,47277125.0,48945375.0,51456833.0,52722583.0,54213834.0,56039000.0,57784167.0,59495042.0,61278500.0,63002417.0,65427167.0,66507167.0,68296042.0,69993958.0,71328083.0,72858958.0,74667125.0,76464917.0,78966041.0,80144666.0,81828958.0,84082958.0,85866167.0,87827541.0,89305000.0,91058000.0,92796084.0,94492792.0,96602792.0,97960917.0,100210083.0,101570042.0,103331917.0,105417084.0,106845750.0,108468375.0,110542125.0,112118583.0,113795667.0,115071209.0,118229209.0,118958209.0,121066334.0,122657416.0,124352916.0,125897333.0,128061875.0,128678791.0,131277916.0,133021083.0,135098167.0,136610916.0,138408333.0,139861292.0,141588500.0,143784625.0,145596042.0,147398583.0,148863291.0,150648833.0,151253958.0,153260125.0,154478000.0,157546042.0,159663959.0,160921542.0,162638333.0,165199750.0,166388291.0,168323667.0,169545500.0,170070875.0,173093792.0,174979875.0],"unit":"ns","throughput":[{"per_iteration":155,"unit":"elements"}],"typical":{"estimate":874915.5711452637,"lower_bound":874111.8499838482,"upper_bound":875671.1186139599,"unit":"ns"},"mean":{"estimate":875634.5846036086,"lower_bound":874886.4087406697,"upper_bound":876396.9107301264,"unit":"ns"},"median":{"estimate":875562.1836538462,"lower_bound":875186.1066666667,"upper_bound":875795.9966008773,"unit":"ns"},"median_abs_dev":{"estimate":2155.0759590474686,"lower_bound":1378.761989034104,"upper_bound":2640.162613617313,"unit":"ns"},"slope":{"estimate":874915.5711452637,"lower_bound":874111.8499838482,"upper_bound":875671.1186139599,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_160","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1802292.0,3573541.0,5386708.0,7320375.0,8965250.0,10866500.0,12618666.0,14482583.0,16043917.0,17834209.0,19654209.0,21385083.0,23212250.0,24956875.0,26755417.0,28526708.0,30354875.0,32127791.0,33950459.0,35947916.0,37501167.0,39565792.0,41576084.0,42849500.0,44711042.0,46995959.0,48272833.0,49974166.0,51800208.0,53579000.0,55390125.0,57072375.0,58952334.0,60636542.0,62582750.0,64242875.0,66081500.0,67849916.0,69667167.0,71676125.0,73715833.0,74941417.0,77262667.0,78494458.0,80282250.0,82139000.0,83942667.0,85685833.0,87591875.0,89184042.0,91182875.0,93253041.0,95243000.0,96363875.0,98857375.0,99979958.0,101729333.0,103525167.0,105287125.0,107638250.0,108852625.0,111430875.0,112725875.0,114298125.0,116624459.0,117892875.0,119768500.0,121486666.0,123223375.0,125045833.0,127814000.0,128488083.0,130943459.0,132142166.0,134050250.0,135511500.0,137430959.0,140080167.0,141249708.0,142819750.0,145433791.0,146393542.0,148112833.0,150078500.0,152564125.0,153667041.0,155243958.0,157792125.0,158870084.0,161156333.0,162826042.0,164594125.0,166019292.0,168361834.0,169702708.0,171321584.0,173818333.0,175568041.0,176748750.0,179007166.0],"unit":"ns","throughput":[{"per_iteration":160,"unit":"elements"}],"typical":{"estimate":894241.0951426039,"lower_bound":893740.62103421,"upper_bound":894762.0809428452,"unit":"ns"},"mean":{"estimate":894746.1517756243,"lower_bound":894055.124680751,"upper_bound":895516.3049996404,"unit":"ns"},"median":{"estimate":893200.8516233766,"lower_bound":892919.5150669643,"upper_bound":893794.6428571428,"unit":"ns"},"median_abs_dev":{"estimate":1424.104118851078,"lower_bound":1049.858151899733,"upper_bound":2354.449286892875,"unit":"ns"},"slope":{"estimate":894241.0951426039,"lower_bound":893740.62103421,"upper_bound":894762.0809428452,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_165","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1893541.0,3883041.0,5821167.0,7711709.0,9422917.0,11511250.0,13401958.0,15253000.0,17285291.0,19187958.0,21321583.0,22651166.0,24766416.0,27435291.0,28849167.0,31795042.0,32603500.0,33453125.0,35610750.0,37704083.0,39032583.0,41480125.0,43140500.0,44748958.0,46521416.0,47907125.0,50297291.0,51904542.0,53701542.0,55349333.0,57596250.0,58932875.0,60898500.0,62811417.0,65005583.0,67017833.0,68395583.0,70390375.0,72470750.0,73831792.0,75887792.0,77563916.0,79552375.0,81697083.0,83444625.0,85470750.0,86874917.0,88688042.0,91487750.0,92847584.0,95244625.0,96334750.0,97693167.0,99985375.0,102021833.0,103762000.0,105508208.0,107250125.0,109259125.0,110933334.0,113502417.0,114330042.0,117371209.0,120481583.0,121143708.0,123968875.0,134967458.0,126369375.0,128362875.0,129510333.0,131377000.0,134541500.0,137911000.0,137080291.0,139563083.0,141317125.0,142129208.0,143321541.0,146142875.0,148519042.0,150554417.0,151011042.0,154015333.0,156109042.0,158182541.0,159405459.0,160757125.0,162849458.0,165209250.0,166587833.0,168342667.0,170023875.0,172987834.0,174614041.0,178284375.0,177680000.0,180047834.0,181173917.0,183957875.0,185113334.0],"unit":"ns","throughput":[{"per_iteration":165,"unit":"elements"}],"typical":{"estimate":929068.7191325551,"lower_bound":927130.5886695453,"upper_bound":931858.1147475804,"unit":"ns"},"mean":{"estimate":934601.2153016016,"lower_bound":931648.040281006,"upper_bound":937900.3903712982,"unit":"ns"},"median":{"estimate":928884.9323095402,"lower_bound":927162.5,"upper_bound":929717.927631579,"unit":"ns"},"median_abs_dev":{"estimate":5823.097288402841,"lower_bound":4176.033443515548,"upper_bound":7825.768531257104,"unit":"ns"},"slope":{"estimate":929068.7191325551,"lower_bound":927130.5886695453,"upper_bound":931858.1147475804,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_170","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1893625.0,3845333.0,5728833.0,7633541.0,9702834.0,11454125.0,13494458.0,15260750.0,17188959.0,18972959.0,21005458.0,22941125.0,24692875.0,26349000.0,28906083.0,30381834.0,32251584.0,34037500.0,35836417.0,37670625.0,39586459.0,41411958.0,43253583.0,45216667.0,47155000.0,49681875.0,51272000.0,53084000.0,55055542.0,56985292.0,59691958.0,60868792.0,62402875.0,64432916.0,66379417.0,68603625.0,70962834.0,74471416.0,76738875.0,79038125.0,82083917.0,81106542.0,81120458.0,83317000.0,86038084.0,87716125.0,89653750.0,90917750.0,93574833.0,95604834.0,100191625.0,100745209.0,105873458.0,102488833.0,104875667.0,106650125.0,108058208.0,109907291.0,112202083.0,113876708.0,115950708.0,117632791.0,119333791.0,122117084.0,123622083.0,125771708.0,127508542.0,128818333.0,129969708.0,132105750.0,133719334.0,137695375.0,138423000.0,140274333.0,142125541.0,144886541.0,145216166.0,147488375.0,151243500.0,151711041.0,153644459.0,155669500.0,157840333.0,159250291.0,162194416.0,165791000.0,167676458.0,167818250.0,169409166.0,170658666.0,174177791.0,174846000.0,177566417.0,180047542.0,181779167.0,183544833.0,185745083.0,186780500.0,188072791.0,189668334.0],"unit":"ns","throughput":[{"per_iteration":170,"unit":"elements"}],"typical":{"estimate":953081.5336544998,"lower_bound":951425.2188119692,"upper_bound":954901.579773329,"unit":"ns"},"mean":{"estimate":953622.4453699103,"lower_bound":951595.1918813728,"upper_bound":955892.3331681201,"unit":"ns"},"median":{"estimate":950902.1050847457,"lower_bound":949233.4827586206,"upper_bound":953415.1545454545,"unit":"ns"},"median_abs_dev":{"estimate":5671.48973337148,"lower_bound":4599.8520940072995,"upper_bound":7609.310069501134,"unit":"ns"},"slope":{"estimate":953081.5336544998,"lower_bound":951425.2188119692,"upper_bound":954901.579773329,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_175","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[2012959.0,4052917.0,5999084.0,7947875.0,9854625.0,11739917.0,13660292.0,15787958.0,17532250.0,19446667.0,21445833.0,23293500.0,25315250.0,27245166.0,29014375.0,30955500.0,33213666.0,34987042.0,36993917.0,38979250.0,40678875.0,42648459.0,44818458.0,46745333.0,48618958.0,50445791.0,52510458.0,54740334.0,56875292.0,58116875.0,60278208.0,62354041.0,64714750.0,65810708.0,68018500.0,70029584.0,71915083.0,73892666.0,75765083.0,77722875.0,79872000.0,81822958.0,83432958.0,85562125.0,87345666.0,89555666.0,91745417.0,93241042.0,95415792.0,97065875.0,99228709.0,101122917.0,103144708.0,105511833.0,106924000.0,109562125.0,110910000.0,113096791.0,116191750.0,119463125.0,120797750.0,121457416.0,122768000.0,125586875.0,128050708.0,129375333.0,130819542.0,132390875.0,134694458.0,135711125.0,139777875.0,141890709.0,145011417.0,144414792.0,145949167.0,148029250.0,149658916.0,151981625.0,154844458.0,155982709.0,157483708.0,159512791.0,161741250.0,162890084.0,165057792.0,167194708.0,169499875.0,185335084.0,173543750.0,177626666.0,179727250.0,182192958.0,187735750.0,192617625.0,221271625.0,208363209.0,199378792.0,206667000.0,193825750.0,195779916.0],"unit":"ns","throughput":[{"per_iteration":175,"unit":"elements"}],"typical":{"estimate":992474.5361149698,"lower_bound":981617.192039066,"upper_bound":1005351.7916444797,"unit":"ns"},"mean":{"estimate":982563.3393782872,"lower_bound":978072.5531956202,"upper_bound":988160.6063168272,"unit":"ns"},"median":{"estimate":974261.538261218,"lower_bound":973524.1315789474,"upper_bound":975735.1428571428,"unit":"ns"},"median_abs_dev":{"estimate":4180.87341772181,"lower_bound":2921.6180250539105,"upper_bound":6360.518152419116,"unit":"ns"},"slope":{"estimate":992474.5361149698,"lower_bound":981617.192039066,"upper_bound":1005351.7916444797,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_180","report_directory":"","iteration_count":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100],"measured_values":[1003667.0,1979292.0,3181542.0,4003333.0,4982583.0,5990750.0,7122917.0,8001291.0,9078833.0,10109292.0,10949208.0,11914125.0,12881041.0,13876458.0,14893875.0,15800917.0,16799625.0,17783209.0,18756417.0,19742209.0,20784584.0,21739667.0,22903125.0,23768500.0,24721417.0,25681250.0,27773125.0,28067208.0,28979417.0,29879708.0,30734000.0,31602208.0,32653208.0,33595583.0,34611792.0,35591792.0,36618250.0,37744959.0,39477833.0,40377791.0,41522208.0,43502875.0,44157375.0,45712458.0,46273125.0,46898458.0,48424917.0,50134625.0,51715750.0,50815209.0,52530583.0,54560834.0,56177625.0,59706208.0,57131458.0,56372625.0,57843834.0,58771750.0,59678625.0,60765667.0,61794792.0,62675250.0,63700875.0,64035000.0,65502667.0,66173000.0,68220042.0,68419958.0,69167625.0,70366083.0,71245791.0,72296333.0,72748167.0,73904625.0,74380125.0,76143500.0,77056500.0,78048834.0,80141083.0,80561167.0,81054042.0,81997000.0,83376000.0,83109167.0,84163750.0,87443334.0,86524792.0,87401292.0,88709166.0,89271500.0,90002375.0,90980417.0,91828375.0,93397375.0,94449167.0,95005500.0,97143541.0,97002584.0,97855459.0,100609083.0],"unit":"ns","throughput":[{"per_iteration":180,"unit":"elements"}],"typical":{"estimate":1003536.102890498,"lower_bound":1000229.7694261521,"upper_bound":1007534.3920796448,"unit":"ns"},"mean":{"estimate":1005491.4657992343,"lower_bound":1001830.2095101994,"upper_bound":1009564.9206214913,"unit":"ns"},"median":{"estimate":1000646.6310541311,"lower_bound":996270.2497716895,"upper_bound":1003788.7905907668,"unit":"ns"},"median_abs_dev":{"estimate":15796.518343939877,"lower_bound":10726.233713484278,"upper_bound":18085.4252203193,"unit":"ns"},"slope":{"estimate":1003536.102890498,"lower_bound":1000229.7694261521,"upper_bound":1007534.3920796448,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_185","report_directory":"","iteration_count":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100],"measured_values":[1027417.0,2034292.0,3035500.0,4047500.0,5058416.0,6182958.0,7267834.0,8139000.0,9297917.0,10291917.0,11395333.0,12292750.0,13152292.0,14378500.0,15241417.0,16285416.0,17238416.0,18333000.0,19418625.0,22228875.0,21447958.0,22668208.0,23416458.0,24399750.0,25354417.0,26387208.0,27366208.0,28302667.0,29405417.0,30412625.0,31407792.0,32370250.0,33462166.0,34789250.0,35733834.0,36510792.0,37528750.0,38574458.0,39568417.0,40503292.0,41514458.0,42593208.0,43586334.0,44526875.0,45623583.0,47079041.0,47759625.0,48792958.0,50389250.0,50977958.0,51900041.0,52626708.0,53626750.0,54713833.0,55697667.0,57082458.0,57709875.0,58755584.0,59765917.0,60756250.0,61786958.0,62921458.0,63864083.0,65169125.0,66006958.0,68021792.0,67897959.0,69510167.0,69947292.0,70962917.0,71940416.0,73064916.0,73930500.0,75020083.0,75983583.0,77673459.0,78243084.0,79193791.0,80517375.0,81600875.0,82005667.0,83507750.0,84052666.0,85386500.0,86212416.0,87269959.0,88140750.0,89297459.0,90124166.0,91814458.0,93727750.0,95612333.0,94660834.0,95380334.0,96917417.0,97727250.0,98471041.0,99278416.0,100360667.0,101683583.0],"unit":"ns","throughput":[{"per_iteration":185,"unit":"elements"}],"typical":{"estimate":1016898.193155017,"lower_bound":1015521.3959624663,"upper_bound":1018558.8049053497,"unit":"ns"},"mean":{"estimate":1018344.7438001669,"lower_bound":1016497.2270703812,"upper_bound":1020840.2146230334,"unit":"ns"},"median":{"estimate":1014826.3629032258,"lower_bound":1014100.5752941177,"upper_bound":1016505.9523809524,"unit":"ns"},"median_abs_dev":{"estimate":3310.83784154347,"lower_bound":2272.283838218775,"upper_bound":4968.743647271042,"unit":"ns"},"slope":{"estimate":1016898.193155017,"lower_bound":1015521.3959624663,"upper_bound":1018558.8049053497,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_190","report_directory":"","iteration_count":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100],"measured_values":[1050417.0,2072042.0,3113167.0,4175333.0,5252542.0,6297250.0,7569500.0,8286583.0,9589250.0,10622917.0,11484292.0,12468250.0,13479000.0,14514041.0,15605375.0,16641292.0,17632000.0,18659166.0,19696083.0,20734042.0,21775542.0,22879208.0,23860125.0,24877000.0,25923083.0,26962292.0,28079209.0,29054917.0,30227292.0,31235208.0,32157375.0,33250667.0,34516167.0,36044375.0,36281250.0,37382709.0,38383625.0,39417708.0,40453000.0,41499708.0,42550208.0,43604917.0,45036792.0,45612750.0,46956708.0,48037000.0,48703625.0,49814958.0,50868500.0,51946375.0,54143833.0,55040209.0,55196334.0,56327750.0,58010542.0,58223833.0,60026542.0,60182958.0,61218500.0,62494584.0,63541333.0,64726042.0,65346750.0,67054708.0,67411042.0,68449667.0,69561375.0,70878750.0,71922708.0,73010583.0,74095208.0,74729917.0,75655583.0,76810125.0,78223083.0,78988208.0,79855542.0,81996959.0,81955500.0,83143708.0,84144541.0,85055041.0,87285542.0,87249875.0,88243292.0,89131458.0,90692625.0,91574625.0,92258667.0,94206167.0,95853125.0,95716667.0,96636333.0,98762750.0,98669292.0,99577458.0,101078542.0,101831917.0,102748291.0,103826917.0],"unit":"ns","throughput":[{"per_iteration":190,"unit":"elements"}],"typical":{"estimate":1041319.5135510566,"lower_bound":1040077.1925899079,"upper_bound":1042697.5325469293,"unit":"ns"},"mean":{"estimate":1041959.7228621587,"lower_bound":1040584.2580471828,"upper_bound":1043513.3511135223,"unit":"ns"},"median":{"estimate":1038974.1666666667,"lower_bound":1038065.5193163753,"upper_bound":1040164.5185185185,"unit":"ns"},"median_abs_dev":{"estimate":3207.935776893872,"lower_bound":1879.0716570809439,"upper_bound":4485.957204557732,"unit":"ns"},"slope":{"estimate":1041319.5135510566,"lower_bound":1040077.1925899079,"upper_bound":1042697.5325469293,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"light_linear_data/light_linear_data_195","report_directory":"","iteration_count":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100],"measured_values":[1075375.0,2215708.0,3244667.0,4249083.0,5352625.0,6352291.0,7516916.0,8615166.0,9705333.0,10724666.0,11757708.0,12851625.0,13781917.0,14877000.0,15964209.0,16975166.0,18044417.0,19190000.0,20209791.0,21313625.0,22281750.0,23341542.0,24436583.0,25503791.0,26519750.0,27572125.0,28685625.0,29709000.0,30754292.0,31957959.0,33376000.0,34136333.0,35196167.0,36766375.0,37209291.0,38563125.0,39852209.0,40433708.0,41347041.0,42444292.0,43525833.0,44540334.0,45638125.0,47989458.0,49215958.0,48995916.0,49946542.0,51413042.0,52000833.0,53029000.0,54083334.0,55140667.0,56306417.0,57707375.0,58551375.0,59455708.0,61568041.0,61697083.0,62930250.0,63622416.0,64712292.0,66339375.0,66943625.0,67875000.0,68963958.0,70013792.0,71154792.0,72232708.0,73470291.0,74452500.0,75582250.0,77291250.0,77779084.0,78531333.0,79618584.0,81016333.0,81735000.0,82961917.0,83897333.0,84823292.0,86541917.0,87332583.0,88172875.0,89984708.0,90658708.0,91211250.0,92641708.0,93546208.0,94488500.0,95813542.0,96625458.0,97865459.0,99003917.0,100631292.0,100946042.0,102613875.0,103150666.0,104142875.0,105014875.0,106511334.0],"unit":"ns","throughput":[{"per_iteration":195,"unit":"elements"}],"typical":{"estimate":1064597.7931284173,"lower_bound":1063661.5898194865,"upper_bound":1065639.3445314642,"unit":"ns"},"mean":{"estimate":1066065.0748829776,"lower_bound":1064662.8479368957,"upper_bound":1067659.7906183684,"unit":"ns"},"median":{"estimate":1063508.0353460973,"lower_bound":1062460.1304347827,"upper_bound":1064576.5792114695,"unit":"ns"},"median_abs_dev":{"estimate":3721.6316076323424,"lower_bound":2580.163829807544,"upper_bound":5258.735467851016,"unit":"ns"},"slope":{"estimate":1064597.7931284173,"lower_bound":1063661.5898194865,"upper_bound":1065639.3445314642,"unit":"ns"},"change":null}
-        {"reason":"group-complete","group_name":"light_linear_data","benchmarks":["light_linear_data/light_linear_data_4","light_linear_data/light_linear_data_3","light_linear_data/light_linear_data_14","light_linear_data/light_linear_data_13","light_linear_data/light_linear_data_22","light_linear_data/light_linear_data_2","light_linear_data/light_linear_data_23","light_linear_data/light_linear_data_24","light_linear_data/light_linear_data_12","light_linear_data/light_linear_data_37","light_linear_data/light_linear_data_39","light_linear_data/light_linear_data_38","light_linear_data/light_linear_data_31","light_linear_data/light_linear_data_36","light_linear_data/light_linear_data_28","light_linear_data/light_linear_data_17","light_linear_data/light_linear_data_21","light_linear_data/light_linear_data_19","light_linear_data/light_linear_data_26","light_linear_data/light_linear_data_7","light_linear_data/light_linear_data_9","light_linear_data/light_linear_data_18","light_linear_data/light_linear_data_27","light_linear_data/light_linear_data_29","light_linear_data/light_linear_data_16","light_linear_data/light_linear_data_11","light_linear_data/light_linear_data_8","light_linear_data/light_linear_data_6","light_linear_data/light_linear_data_1","light_linear_data/light_linear_data_33","light_linear_data/light_linear_data_34","light_linear_data/light_linear_data_32","light_linear_data/light_linear_data_0","light_linear_data/light_linear_data_5","light_linear_data/light_linear_data_10","light_linear_data/light_linear_data_15","light_linear_data/light_linear_data_20","light_linear_data/light_linear_data_25","light_linear_data/light_linear_data_30","light_linear_data/light_linear_data_35","light_linear_data/light_linear_data_40","light_linear_data/light_linear_data_45","light_linear_data/light_linear_data_50","light_linear_data/light_linear_data_55","light_linear_data/light_linear_data_60","light_linear_data/light_linear_data_65","light_linear_data/light_linear_data_70","light_linear_data/light_linear_data_75","light_linear_data/light_linear_data_80","light_linear_data/light_linear_data_85","light_linear_data/light_linear_data_90","light_linear_data/light_linear_data_95","light_linear_data/light_linear_data_100","light_linear_data/light_linear_data_105","light_linear_data/light_linear_data_110","light_linear_data/light_linear_data_115","light_linear_data/light_linear_data_120","light_linear_data/light_linear_data_125","light_linear_data/light_linear_data_130","light_linear_data/light_linear_data_135","light_linear_data/light_linear_data_140","light_linear_data/light_linear_data_145","light_linear_data/light_linear_data_150","light_linear_data/light_linear_data_155","light_linear_data/light_linear_data_160","light_linear_data/light_linear_data_165","light_linear_data/light_linear_data_170","light_linear_data/light_linear_data_175","light_linear_data/light_linear_data_180","light_linear_data/light_linear_data_185","light_linear_data/light_linear_data_190","light_linear_data/light_linear_data_195"],"report_directory":""}
-        "#;
-
-        let mut state = State {
-            all: false,
-            baseline: "baseline/baseline".into(),
-            ids: Default::default(),
-            throughput: Default::default(),
-            groups: Default::default(),
-        };
-        for line in input.lines() {
-            extract_state(line, &mut state, false);
-        }
-
-        let (_, costs) = state.into_dependent_costs();
-        let linear_data_cost = costs
-            .0
-            .get("light_linear_data")
-            .expect("Expected light_linear_data cost");
-
-        // f(x) = (1/5)x + 5
-        let expected_cost = DependentCost::LightOperation {
-            base: 5,
-            units_per_gas: 5,
-        };
-        assert!(matches!(linear_data_cost, Cost::Dependent(c) if *c == expected_cost));
-    }
-
-    #[test]
-    fn categorizes_dependent_cost_from_heavy_linear_function() {
-        // f(x) = 2x + 10
-        let input = r#"
-        {"reason":"benchmark-complete","id":"baseline/baseline","report_directory":"","iteration_count":[41,82,123,164,205,246,287,328,369,410,451,492,533,574,615,656,697,738,779,820,861,902,943,984,1025,1066,1107,1148,1189,1230,1271,1312,1353,1394,1435,1476,1517,1558,1599,1640,1681,1722,1763,1804,1845,1886,1927,1968,2009,2050,2091,2132,2173,2214,2255,2296,2337,2378,2419,2460,2501,2542,2583,2624,2665,2706,2747,2788,2829,2870,2911,2952,2993,3034,3075,3116,3157,3198,3239,3280,3321,3362,3403,3444,3485,3526,3567,3608,3649,3690,3731,3772,3813,3854,3895,3936,3977,4018,4059,4100],"measured_values":[1024750.0,2034125.0,3050375.0,4073625.0,5117084.0,6007375.0,6944125.0,7953708.0,8976708.0,10026750.0,11039416.0,11974083.0,12832125.0,13910833.0,14982667.0,15998416.0,16948083.0,17965375.0,18906458.0,19935500.0,20881791.0,21798042.0,22722792.0,23772500.0,24899583.0,25781042.0,26793708.0,27746833.0,29052041.0,29859459.0,30968042.0,32008541.0,33069167.0,33995791.0,34683375.0,35840458.0,36740500.0,37554625.0,38616334.0,39846542.0,40798917.0,41886375.0,42554791.0,43666500.0,44857334.0,46191083.0,47023583.0,47765125.0,48668708.0,49641541.0,50547083.0,51671208.0,52758708.0,56393042.0,55771541.0,56209000.0,56916208.0,58265584.0,59250958.0,59928167.0,60456458.0,62097125.0,62621625.0,63759208.0,65108625.0,65371083.0,66127750.0,67234875.0,68585458.0,69966709.0,71205666.0,71766583.0,72900833.0,73424458.0,74742459.0,75535042.0,76276708.0,77651875.0,78934625.0,79662208.0,80375416.0,81877166.0,83869334.0,84756708.0,85806209.0,87086542.0,88260083.0,88333750.0,88901416.0,89340666.0,90566250.0,92137625.0,92691250.0,94276125.0,95762666.0,96414625.0,96679208.0,97547292.0,98526542.0,99481875.0],"unit":"ns","throughput":[],"typical":{"estimate":24368.771698090084,"lower_bound":24324.60001542914,"upper_bound":24415.222083827928,"unit":"ns"},"mean":{"estimate":24362.850103593777,"lower_bound":24322.611557310047,"upper_bound":24408.090478951748,"unit":"ns"},"median":{"estimate":24310.377600019558,"lower_bound":24279.202945237,"upper_bound":24355.417937176833,"unit":"ns"},"median_abs_dev":{"estimate":151.08854480064943,"lower_bound":106.55709469359827,"upper_bound":185.53076162879583,"unit":"ns"},"slope":{"estimate":24368.771698090084,"lower_bound":24324.60001542914,"upper_bound":24415.222083827928,"unit":"ns"},"change":{"mean":{"estimate":0.005186365234391088,"lower_bound":0.003045682373511866,"upper_bound":0.007426882227977042,"unit":"%"},"median":{"estimate":0.0052777020670709,"lower_bound":0.003296357893202595,"upper_bound":0.007667895973157757,"unit":"%"},"change":"NoChange"}}
-        {"reason":"group-complete","group_name":"baseline","benchmarks":["baseline/baseline"],""}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_0","report_directory":"","iteration_count":[5,10,15,20,25,30,35,40,45,50,55,60,65,70,75,80,85,90,95,100,105,110,115,120,125,130,135,140,145,150,155,160,165,170,175,180,185,190,195,200,205,210,215,220,225,230,235,240,245,250,255,260,265,270,275,280,285,290,295,300,305,310,315,320,325,330,335,340,345,350,355,360,365,370,375,380,385,390,395,400,405,410,415,420,425,430,435,440,445,450,455,460,465,470,475,480,485,490,495,500],"measured_values":[1262125.0,2479459.0,3726958.0,4966541.0,6191625.0,7233333.0,8525583.0,9720708.0,11092083.0,12145792.0,13336708.0,14551584.0,15769042.0,17155209.0,18222333.0,19414375.0,20468417.0,21700125.0,22874541.0,24073667.0,25332334.0,26518542.0,27860208.0,29086166.0,30361917.0,31524167.0,32551416.0,33772208.0,34972542.0,36227458.0,37388500.0,38606791.0,39822292.0,41126833.0,42398125.0,44226458.0,44828167.0,45844166.0,47028208.0,48182083.0,49602125.0,51460750.0,52442833.0,53791625.0,54280500.0,55368667.0,57017000.0,58200625.0,59088375.0,60276625.0,61490875.0,62641125.0,64084834.0,65194250.0,66253834.0,67513750.0,69065709.0,70390833.0,71894792.0,72520084.0,73545958.0,74703958.0,76227791.0,77083334.0,78439833.0,79524917.0,81053458.0,82004167.0,83265417.0,84378166.0,86180625.0,87814000.0,88023208.0,89105625.0,90730500.0,91519166.0,93176917.0,94305250.0,95579459.0,96557208.0,97726792.0,99668958.0,101081667.0,101789458.0,103288292.0,104543417.0,106194584.0,106909000.0,107449625.0,108481375.0,109980333.0,111876792.0,113121459.0,113896125.0,115551708.0,116284000.0,117548708.0,118704208.0,119792542.0,121940375.0],"unit":"ns","throughput":[{"per_iteration":0,"unit":"elements"}],"typical":{"estimate":242193.72566100192,"lower_bound":241943.25637054347,"upper_bound":242440.87479912373,"unit":"ns"},"mean":{"estimate":242475.55609234053,"lower_bound":242121.3556242586,"upper_bound":242872.9141176644,"unit":"ns"},"median":{"estimate":241983.15061281895,"lower_bound":241516.38666666666,"upper_bound":242344.01661600813,"unit":"ns"},"median_abs_dev":{"estimate":1235.1102854532612,"lower_bound":873.2926804289611,"upper_bound":1593.912878774232,"unit":"ns"},"slope":{"estimate":242193.72566100192,"lower_bound":241943.25637054347,"upper_bound":242440.87479912373,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_5","report_directory":"","iteration_count":[3,6,9,12,15,18,21,24,27,30,33,36,39,42,45,48,51,54,57,60,63,66,69,72,75,78,81,84,87,90,93,96,99,102,105,108,111,114,117,120,123,126,129,132,135,138,141,144,147,150,153,156,159,162,165,168,171,174,177,180,183,186,189,192,195,198,201,204,207,210,213,216,219,222,225,228,231,234,237,240,243,246,249,252,255,258,261,264,267,270,273,276,279,282,285,288,291,294,297,300],"measured_values":[1464750.0,2980333.0,4379083.0,5946292.0,7277667.0,8802083.0,10376792.0,11720292.0,13142333.0,14825208.0,16515583.0,17870291.0,19039083.0,20691667.0,21957666.0,23461208.0,25080916.0,26474875.0,27804292.0,29359125.0,30998459.0,32503416.0,33802250.0,35403166.0,36941708.0,38106250.0,39559167.0,40683834.0,42250666.0,44055792.0,49692292.0,48108500.0,48725875.0,50071000.0,51383833.0,52730417.0,53838125.0,55858583.0,56586791.0,58410333.0,59825542.0,61260334.0,62841125.0,63960667.0,65675834.0,67056209.0,68465583.0,72576875.0,72054709.0,73286125.0,75183250.0,77442125.0,81107500.0,82431959.0,82458750.0,82405333.0,84804666.0,87062667.0,88077958.0,89138875.0,90534583.0,91741167.0,92070875.0,93581083.0,94315542.0,96424291.0,98029833.0,99632000.0,100430584.0,101785000.0,104020875.0,105132250.0,106562083.0,108672500.0,109086250.0,110196375.0,112027291.0,112989833.0,114654583.0,116953459.0,118203125.0,119581583.0,121794292.0,123478375.0,124697500.0,125597000.0,127249959.0,129613750.0,131834416.0,132829500.0,134007209.0,135099959.0,135694875.0,138095625.0,140486125.0,143471875.0,145034000.0,145250541.0,147227958.0,145792625.0],"unit":"ns","throughput":[{"per_iteration":5,"unit":"elements"}],"typical":{"estimate":490226.1836047485,"lower_bound":488984.5864652994,"upper_bound":491530.4670938426,"unit":"ns"},"mean":{"estimate":490767.8296126263,"lower_bound":489536.9594065041,"upper_bound":492204.87726355286,"unit":"ns"},"median":{"estimate":489071.7533664068,"lower_bound":488250.0,"upper_bound":490077.82930540637,"unit":"ns"},"median_abs_dev":{"estimate":4143.525693104397,"lower_bound":3010.5641306125513,"upper_bound":5354.271161853297,"unit":"ns"},"slope":{"estimate":490226.1836047485,"lower_bound":488984.5864652994,"upper_bound":491530.4670938426,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_10","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1582167.0,2883541.0,4359125.0,5842833.0,7371084.0,8790458.0,10200750.0,11716250.0,13310500.0,14638291.0,16236125.0,17609042.0,18924459.0,20522750.0,21839250.0,23498708.0,24737792.0,26317667.0,27800750.0,29062541.0,30672959.0,32695000.0,34063333.0,35515334.0,37121833.0,37957042.0,39473875.0,41164917.0,43400584.0,44612458.0,45889791.0,47272584.0,48244542.0,49558750.0,50919542.0,52609625.0,54156125.0,56364042.0,57902791.0,58729167.0,59611208.0,61200250.0,63206125.0,64282708.0,65439833.0,67062375.0,68334750.0,69843708.0,71207250.0,72520667.0,73970333.0,75427709.0,77475833.0,79060417.0,80109041.0,81795958.0,83008791.0,84113666.0,85537125.0,86956625.0,88928042.0,90550125.0,91501583.0,92813791.0,94958709.0,96322583.0,97329333.0,98810083.0,100690208.0,102006084.0,103801750.0,104912750.0,105916041.0,107741042.0,110701583.0,111571042.0,112361833.0,114769917.0,115268167.0,116840750.0,118570667.0,121482542.0,122271375.0,123165291.0,124350542.0,125864666.0,126592792.0,128345542.0,129945750.0,131990875.0,134398166.0,134394292.0,135532083.0,138201750.0,138699000.0,140557458.0,141331667.0,143123375.0,144795625.0,145310833.0],"unit":"ns","throughput":[{"per_iteration":10,"unit":"elements"}],"typical":{"estimate":730757.714224915,"lower_bound":729850.0549386311,"upper_bound":731717.4444780161,"unit":"ns"},"mean":{"estimate":732001.4861425048,"lower_bound":730658.9150475737,"upper_bound":733681.381385002,"unit":"ns"},"median":{"estimate":730337.5892857143,"lower_bound":729639.1884057971,"upper_bound":731144.6373950775,"unit":"ns"},"median_abs_dev":{"estimate":4110.222652764288,"lower_bound":2783.1410152047715,"upper_bound":5377.341401045584,"unit":"ns"},"slope":{"estimate":730757.714224915,"lower_bound":729850.0549386311,"upper_bound":731717.4444780161,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_15","report_directory":"","iteration_count":[2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38,40,42,44,46,48,50,52,54,56,58,60,62,64,66,68,70,72,74,76,78,80,82,84,86,88,90,92,94,96,98,100,102,104,106,108,110,112,114,116,118,120,122,124,126,128,130,132,134,136,138,140,142,144,146,148,150,152,154,156,158,160,162,164,166,168,170,172,174,176,178,180,182,184,186,188,190,192,194,196,198,200],"measured_values":[1959875.0,3920375.0,6042167.0,7852625.0,9927000.0,11752083.0,13645291.0,15638000.0,17458375.0,19328125.0,21254458.0,23191375.0,25257417.0,27193208.0,29194250.0,30975209.0,33066917.0,34715458.0,37036083.0,39058500.0,40957916.0,42942166.0,44700708.0,46347000.0,48431333.0,50664667.0,52713209.0,54211958.0,56081000.0,58242458.0,60537250.0,61791833.0,63894666.0,65568959.0,67581084.0,69680500.0,72330125.0,74495542.0,76113875.0,77897750.0,79814667.0,82961250.0,83991834.0,85234917.0,87622000.0,88953625.0,90661833.0,92760959.0,95603375.0,97645292.0,98946625.0,100854584.0,103271167.0,104560291.0,106848250.0,108957666.0,110827250.0,112214541.0,115389459.0,116372708.0,118423542.0,120239708.0,122226916.0,124577792.0,126277750.0,128035834.0,130686292.0,131543042.0,133147167.0,135573834.0,137977000.0,139563292.0,141994584.0,143750333.0,145669375.0,147371042.0,149428792.0,150948125.0,152698834.0,154956958.0,158760500.0,158808458.0,160603834.0,162235458.0,165329333.0,167418542.0,169826167.0,187584833.0,173431583.0,175019000.0,176025334.0,178360792.0,181902208.0,182337958.0,184685833.0,186468583.0,188742417.0,191423875.0,192654292.0,201314042.0],"unit":"ns","throughput":[{"per_iteration":15,"unit":"elements"}],"typical":{"estimate":974619.5497088813,"lower_bound":970999.1672063577,"upper_bound":979930.1300742803,"unit":"ns"},"mean":{"estimate":973557.3464800061,"lower_bound":971624.6545212195,"upper_bound":976088.2403060923,"unit":"ns"},"median":{"estimate":971403.2115384615,"lower_bound":970064.9509803922,"upper_bound":972836.3035714285,"unit":"ns"},"median_abs_dev":{"estimate":4918.336014605142,"lower_bound":3890.785302875953,"upper_bound":6178.142757623847,"unit":"ns"},"slope":{"estimate":974619.5497088813,"lower_bound":970999.1672063577,"upper_bound":979930.1300742803,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_20","report_directory":"","iteration_count":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100],"measured_values":[1327500.0,2475541.0,3675333.0,4993500.0,6213500.0,7332459.0,8693459.0,9731083.0,11022916.0,12359833.0,13593000.0,14835792.0,15875375.0,17059458.0,18249125.0,19344417.0,20591084.0,22045084.0,23413000.0,24297250.0,25848417.0,26713291.0,27830833.0,29185500.0,30300333.0,31356167.0,32594125.0,34272167.0,35035791.0,36320500.0,37536375.0,38643584.0,39862208.0,40979667.0,42232834.0,43952167.0,46435875.0,46636834.0,47961458.0,49008625.0,50218250.0,50880375.0,51852375.0,53271750.0,55005209.0,56148000.0,57108166.0,57955500.0,59046584.0,61260250.0,61947625.0,62755375.0,64021083.0,65167375.0,66259875.0,67972417.0,68973000.0,70443208.0,71826667.0,72830917.0,74733167.0,75630584.0,76024459.0,77091250.0,81410541.0,80931292.0,82439959.0,82189291.0,83694750.0,84598667.0,85694250.0,88414875.0,90567041.0,90917458.0,90906792.0,91822458.0,93535875.0,94676833.0,95738875.0,97654000.0,98325875.0,99553083.0,103732500.0,104270459.0,104977084.0,104486958.0,105793875.0,107283791.0,110834833.0,111774042.0,111266833.0,111907792.0,114402584.0,117268959.0,116852792.0,117286833.0,118019166.0,119157125.0,120546959.0,121394500.0],"unit":"ns","throughput":[{"per_iteration":20,"unit":"elements"}],"typical":{"estimate":1221299.71035614,"lower_bound":1218129.9084466272,"upper_bound":1224575.1017444748,"unit":"ns"},"mean":{"estimate":1221433.948687972,"lower_bound":1218474.8178936741,"upper_bound":1224839.2559310328,"unit":"ns"},"median":{"estimate":1216498.6884057971,"lower_bound":1214653.9308510637,"upper_bound":1220895.693979933,"unit":"ns"},"median_abs_dev":{"estimate":12289.425077433238,"lower_bound":8879.875806506556,"upper_bound":14712.619387689805,"unit":"ns"},"slope":{"estimate":1221299.71035614,"lower_bound":1218129.9084466272,"upper_bound":1224575.1017444748,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_25","report_directory":"","iteration_count":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100],"measured_values":[1462458.0,2944417.0,4400000.0,5841000.0,7295250.0,8858334.0,10232292.0,11856417.0,13267042.0,14721375.0,16131916.0,17469500.0,18979708.0,20464875.0,21917334.0,23351583.0,25040916.0,26348125.0,27615083.0,29213375.0,30612292.0,32047125.0,33511917.0,34832291.0,37446000.0,38127250.0,39754208.0,40878375.0,42058250.0,43468791.0,44863334.0,46586333.0,48126500.0,49799875.0,51225042.0,52270375.0,53825791.0,55148417.0,56482833.0,57966167.0,59566500.0,60783542.0,62177042.0,63696334.0,65068542.0,66994459.0,68106625.0,70032500.0,70852334.0,72777625.0,74767125.0,75685542.0,76944833.0,78279542.0,79870167.0,82066625.0,85079667.0,85738042.0,86680250.0,89951000.0,92835083.0,101757500.0,96441833.0,95056042.0,96812250.0,99032500.0,103805250.0,106348791.0,115210417.0,137630958.0,126226375.0,118900583.0,124562084.0,118938167.0,116368167.0,115212000.0,116841041.0,118115834.0,119727917.0,121128709.0,126255959.0,125285125.0,125890542.0,127285417.0,127965375.0,129304250.0,131865291.0,134217958.0,134679708.0,132159417.0,134018583.0,135995250.0,136570500.0,136300167.0,138267417.0,142323625.0,142056959.0,143138000.0,144430042.0,146458208.0],"unit":"ns","throughput":[{"per_iteration":25,"unit":"elements"}],"typical":{"estimate":1511340.4867090292,"lower_bound":1492886.7340210748,"upper_bound":1533488.834967713,"unit":"ns"},"mean":{"estimate":1492481.0753320022,"lower_bound":1479373.934509595,"upper_bound":1508183.0109396738,"unit":"ns"},"median":{"estimate":1465748.7526260503,"lower_bound":1461155.6,"upper_bound":1472470.0906593406,"unit":"ns"},"median_abs_dev":{"estimate":20573.552247782663,"lower_bound":13760.511936611461,"upper_bound":30986.543646908773,"unit":"ns"},"slope":{"estimate":1511340.4867090292,"lower_bound":1492886.7340210748,"upper_bound":1533488.834967713,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_30","report_directory":"","iteration_count":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100],"measured_values":[1683208.0,3358083.0,5209625.0,6842958.0,8605375.0,10377791.0,12104542.0,13759208.0,15346709.0,17280708.0,18909125.0,20395125.0,22215750.0,24049500.0,25671833.0,27206334.0,28860583.0,30537375.0,32171959.0,33897042.0,35761334.0,38296750.0,39733083.0,41736500.0,42731584.0,44149625.0,46042834.0,47780458.0,49121042.0,51243209.0,52460375.0,54209833.0,56163625.0,58243750.0,59350417.0,61327250.0,63111250.0,64418500.0,66059125.0,67490292.0,69341833.0,71260625.0,72940583.0,74991625.0,76072458.0,78291958.0,79538917.0,81528542.0,82727833.0,84492458.0,86235792.0,87837125.0,89502584.0,91486125.0,93192000.0,95184916.0,96371917.0,98647208.0,100268208.0,101282417.0,103287792.0,104532375.0,106266084.0,108461250.0,110292875.0,111945541.0,112994375.0,115505416.0,116690500.0,118430166.0,120128292.0,121549791.0,124087333.0,125461583.0,126934959.0,128775083.0,130615458.0,132393041.0,133377542.0,135403542.0,137439917.0,138976625.0,140154375.0,142416667.0,143667708.0,145490167.0,147450167.0,149387167.0,150932042.0,152223583.0,154640500.0,155597416.0,157100917.0,160096083.0,161104000.0,162681500.0,164410750.0,167515375.0,167041667.0,169593375.0],"unit":"ns","throughput":[{"per_iteration":30,"unit":"elements"}],"typical":{"estimate":1694836.1673828876,"lower_bound":1693542.6842918058,"upper_bound":1696263.9887036164,"unit":"ns"},"mean":{"estimate":1699390.22138437,"lower_bound":1697089.8625611172,"upper_bound":1701870.0470829573,"unit":"ns"},"median":{"estimate":1695848.6096984032,"lower_bound":1694707.03125,"upper_bound":1697346.6794871795,"unit":"ns"},"median_abs_dev":{"estimate":6869.132692510763,"lower_bound":5306.392612658779,"upper_bound":9731.62923850182,"unit":"ns"},"slope":{"estimate":1694836.1673828876,"lower_bound":1693542.6842918058,"upper_bound":1696263.9887036164,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_35","report_directory":"","iteration_count":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100],"measured_values":[2003750.0,3923250.0,5776708.0,7763542.0,9832750.0,11769125.0,13677458.0,15631666.0,17382292.0,19272666.0,21283292.0,23394667.0,25356500.0,27412208.0,29212416.0,30944833.0,32824625.0,34800083.0,36642417.0,39080750.0,40775583.0,42487875.0,44291875.0,46864833.0,48279500.0,50243666.0,52484125.0,54632667.0,56041500.0,58188125.0,60359833.0,61859166.0,63767291.0,65944041.0,67822375.0,69697458.0,71426375.0,73539250.0,75295875.0,77389208.0,79575000.0,81094000.0,83453958.0,85222250.0,87157667.0,88717834.0,91104917.0,93080750.0,94612792.0,96986542.0,98352500.0,101009333.0,102777875.0,104598375.0,106707750.0,108022084.0,110236541.0,112443625.0,114120584.0,115888750.0,117789500.0,121119000.0,122481458.0,124013291.0,125566709.0,127975708.0,129616625.0,131460625.0,133007542.0,136521875.0,137723667.0,138843875.0,141448959.0,142927292.0,145025000.0,146513875.0,149650833.0,151194000.0,152581333.0,155123583.0,156394416.0,158470875.0,160977542.0,163316959.0,164621042.0,166614625.0,168241916.0,170052500.0,172210792.0,174990250.0,176347042.0,178472666.0,180119792.0,181452500.0,184039708.0,186483584.0,187577417.0,189625458.0,191655458.0,193940458.0],"unit":"ns","throughput":[{"per_iteration":35,"unit":"elements"}],"typical":{"estimate":1936566.6582621546,"lower_bound":1935354.530311152,"upper_bound":1937821.0246771297,"unit":"ns"},"mean":{"estimate":1938739.7257038406,"lower_bound":1936811.6422326262,"upper_bound":1940956.883975652,"unit":"ns"},"median":{"estimate":1936853.1813131312,"lower_bound":1934710.6099050203,"upper_bound":1938404.0108225108,"unit":"ns"},"median_abs_dev":{"estimate":6634.218556158306,"lower_bound":5150.994229311434,"upper_bound":8940.788306568476,"unit":"ns"},"slope":{"estimate":1936566.6582621546,"lower_bound":1935354.530311152,"upper_bound":1937821.0246771297,"unit":"ns"},"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_40","report_directory":"","iteration_count":[23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23,23],"measured_values":[50263375.0,50053208.0,49997792.0,51398000.0,50362416.0,50172416.0,50438666.0,50054250.0,50295834.0,50050750.0,49924084.0,49948959.0,50051416.0,50140708.0,50611166.0,50085459.0,49889000.0,49945667.0,50009958.0,50027875.0,49874042.0,50001458.0,49999042.0,50003125.0,49878875.0,50245333.0,50355583.0,50110583.0,50210250.0,50223542.0,50033708.0,49942708.0,50153750.0,50206708.0,50036709.0,50649583.0,49886333.0,49965875.0,50586667.0,50193333.0,49895375.0,50017750.0,49836583.0,50036500.0,49975375.0,49968667.0,50394958.0,50054875.0,50342917.0,50151083.0,49939375.0,49924666.0,50096792.0,50172459.0,50030167.0,50501625.0,50086583.0,49914500.0,50144791.0,50112583.0,49954792.0,49905750.0,49966125.0,50184458.0,49975416.0,50020208.0,50437792.0,50121166.0,50250583.0,50205333.0,49977708.0,49864667.0,49975500.0,50376334.0,50158583.0,50144583.0,50392375.0,49918000.0,50127083.0,50115750.0,49876417.0,49904250.0,49971667.0,50177959.0,50019042.0,49955625.0,50352458.0,50048791.0,50281917.0,50178875.0,49897750.0,49912291.0,50037916.0,50252875.0,50029334.0,50185084.0,49942166.0,50040917.0,50057667.0,50108750.0],"unit":"ns","throughput":[{"per_iteration":40,"unit":"elements"}],"typical":{"estimate":2178773.8204347827,"lower_bound":2177022.353478261,"upper_bound":2180739.130945651,"unit":"ns"},"mean":{"estimate":2178773.8204347827,"lower_bound":2177022.353478261,"upper_bound":2180739.130945651,"unit":"ns"},"median":{"estimate":2176187.4782608696,"lower_bound":2175174.8260869565,"upper_bound":2178833.3260869565,"unit":"ns"},"median_abs_dev":{"estimate":6969.799167565752,"lower_bound":5515.465284689418,"upper_bound":8828.496078045862,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_45","report_directory":"","iteration_count":[21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21,21],"measured_values":[51115208.0,50740875.0,50855583.0,51016084.0,50820250.0,51069333.0,50928459.0,50760833.0,50638500.0,50920791.0,50616792.0,50944458.0,50704750.0,50592792.0,50795709.0,50867250.0,50805500.0,50650709.0,50718709.0,51586375.0,50666000.0,50738250.0,50864416.0,51069083.0,50863417.0,51051208.0,50748000.0,50668250.0,50758792.0,51072459.0,50795458.0,50985709.0,50610584.0,50620750.0,50652167.0,50944583.0,50637167.0,50661958.0,50738833.0,50860083.0,51383791.0,50705667.0,51117208.0,50899666.0,50915584.0,51088584.0,50734292.0,50665083.0,50877583.0,50934000.0,50860208.0,50865209.0,50694666.0,51031209.0,50830709.0,50888167.0,50626542.0,50802209.0,50698583.0,50903833.0,51232917.0,50704292.0,51137500.0,50912667.0,51176750.0,50731625.0,50739667.0,50615709.0,50936833.0,50875375.0,50891083.0,50785583.0,50715166.0,50785000.0,51207834.0,50870917.0,50706750.0,50650417.0,50851500.0,50646917.0,50808917.0,51474000.0,51034375.0,50675041.0,51058667.0,50655584.0,50635542.0,50619500.0,51055417.0,50737042.0,50940958.0,50628000.0,50603125.0,50624625.0,50857500.0,50724292.0,50627333.0,50759833.0,50890584.0,50679042.0],"unit":"ns","throughput":[{"per_iteration":45,"unit":"elements"}],"typical":{"estimate":2420925.1423809514,"lower_bound":2419163.255273809,"upper_bound":2422786.81,"unit":"ns"},"mean":{"estimate":2420925.1423809514,"lower_bound":2419163.255273809,"upper_bound":2422786.81,"unit":"ns"},"median":{"estimate":2419231.166666667,"lower_bound":2416232.1428571427,"upper_bound":2422067.476190476,"unit":"ns"},"median_abs_dev":{"estimate":8804.4199436903,"lower_bound":6751.124880143476,"upper_bound":11375.568645543755,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_50","report_directory":"","iteration_count":[19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19,19],"measured_values":[51062250.0,50847500.0,50399292.0,50440500.0,50631083.0,50469583.0,50476000.0,50365708.0,50808875.0,50619583.0,50632417.0,50746708.0,50466334.0,50991333.0,50849041.0,50468500.0,50725458.0,50459125.0,50440541.0,50391042.0,50765875.0,50586333.0,50406541.0,50431583.0,50610833.0,50430709.0,50438500.0,50538209.0,50941000.0,50629417.0,50702750.0,50677667.0,50441709.0,50372042.0,51185000.0,50566250.0,50637000.0,50487000.0,50516917.0,50354208.0,50497625.0,50561708.0,50401958.0,50475167.0,50617959.0,50453542.0,50408500.0,50440750.0,51052041.0,50347666.0,50816750.0,50480875.0,50447459.0,50303292.0,50925292.0,51050375.0,50670958.0,50389833.0,50416334.0,50407500.0,50675500.0,50540208.0,50660084.0,50390959.0,50736250.0,50384208.0,50415958.0,50572916.0,50881750.0,50447250.0,51117917.0,50847292.0,50653959.0,50515292.0,50810167.0,50746583.0,50867042.0,50398292.0,50327958.0,51748750.0,51382875.0,51144917.0,51104500.0,51557541.0,51041125.0,51050791.0,52445750.0,51603167.0,52066959.0,52760541.0,51821542.0,52324458.0,51900250.0,52680917.0,53386708.0,54509083.0,61238542.0,66557000.0,54708500.0,55117792.0],"unit":"ns","throughput":[{"per_iteration":50,"unit":"elements"}],"typical":{"estimate":2694377.522631579,"lower_bound":2676346.4928815775,"upper_bound":2718197.8797763158,"unit":"ns"},"mean":{"estimate":2694377.522631579,"lower_bound":2676346.4928815775,"upper_bound":2718197.8797763158,"unit":"ns"},"median":{"estimate":2664828.947368421,"lower_bound":2660010.947368421,"upper_bound":2670328.947368421,"unit":"ns"},"median_abs_dev":{"estimate":16827.197574941827,"lower_bound":10917.944237746582,"upper_bound":23386.688047960968,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_55","report_directory":"","iteration_count":[17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17,17],"measured_values":[49828833.0,49431833.0,49657417.0,49177125.0,49440583.0,49192042.0,49215791.0,49223541.0,49414041.0,49815917.0,49434500.0,49529833.0,50163083.0,50049541.0,49282542.0,49342709.0,49193958.0,49302000.0,49504083.0,49414250.0,49715292.0,49769292.0,49485042.0,49433583.0,49479750.0,49291500.0,49187125.0,49805875.0,49448250.0,49410958.0,49752334.0,49670791.0,49242250.0,49160375.0,49151750.0,49103709.0,49246583.0,49168709.0,49775792.0,49271625.0,49748041.0,49633458.0,49672917.0,49545750.0,49534583.0,49313167.0,49194333.0,49206542.0,49165583.0,49419666.0,49283000.0,49643375.0,49525375.0,49164375.0,49164750.0,49209167.0,49203250.0,49177750.0,49446500.0,49429709.0,49499834.0,49637542.0,49812000.0,49638667.0,49264083.0,49435875.0,49229750.0,49234333.0,49244000.0,49348375.0,49317000.0,49381292.0,49901792.0,49376084.0,49262792.0,49169167.0,49229750.0,49222208.0,49401667.0,49673459.0,49336458.0,49656042.0,49512958.0,49930833.0,49230916.0,49401833.0,49316333.0,49229500.0,49208042.0,49531834.0,49392000.0,49190541.0,49293959.0,49651792.0,49717625.0,49316208.0,49273291.0,49173625.0,49322916.0,49454125.0],"unit":"ns","throughput":[{"per_iteration":55,"unit":"elements"}],"typical":{"estimate":2907227.0582352937,"lower_bound":2904655.5266617653,"upper_bound":2909904.706485294,"unit":"ns"},"mean":{"estimate":2907227.0582352937,"lower_bound":2904655.5266617653,"upper_bound":2909904.706485294,"unit":"ns"},"median":{"estimate":2905096.823529412,"lower_bound":2900535.529411765,"upper_bound":2907857.8235294116,"unit":"ns"},"median_abs_dev":{"estimate":13683.176792369453,"lower_bound":9733.22522131854,"upper_bound":18130.976713404874,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_60","report_directory":"","iteration_count":[16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16],"measured_values":[53411041.0,53828875.0,53359417.0,53340708.0,53288417.0,53200667.0,53317750.0,53382500.0,53508167.0,53408083.0,53411333.0,53427291.0,52520167.0,53020041.0,52012917.0,52543125.0,51770375.0,52636125.0,50790125.0,50720750.0,50914750.0,50479958.0,50202792.0,50183792.0,50165750.0,50162583.0,50766708.0,50758875.0,50283125.0,50937917.0,50221917.0,50505667.0,50424250.0,50371917.0,50269792.0,50145958.0,50166084.0,50241834.0,50283416.0,50706041.0,50132625.0,50263917.0,50132667.0,50146958.0,50197917.0,50248875.0,50144334.0,51195125.0,50257208.0,50952500.0,50213541.0,50433792.0,50544209.0,50348291.0,50202750.0,50188917.0,50278041.0,50418917.0,50376583.0,50635125.0,50305375.0,50265875.0,50214458.0,50085958.0,50612042.0,50258458.0,50275792.0,50767000.0,50796292.0,51210000.0,50734500.0,50674458.0,50437625.0,50360916.0,50340292.0,50137667.0,50177084.0,50404459.0,50324292.0,50348125.0,51523375.0,50634625.0,50138000.0,50245459.0,50208417.0,50295833.0,50344625.0,50598208.0,50469875.0,50874500.0,50683291.0,50755166.0,50321042.0,50501833.0,50095459.0,50183541.0,50113708.0,50337083.0,50341917.0,50164500.0],"unit":"ns","throughput":[{"per_iteration":60,"unit":"elements"}],"typical":{"estimate":3181227.760625,"lower_bound":3168591.4074999997,"upper_bound":3194929.194515625,"unit":"ns"},"mean":{"estimate":3181227.760625,"lower_bound":3168591.4074999997,"upper_bound":3194929.194515625,"unit":"ns"},"median":{"estimate":3149407.5625,"lower_bound":3145268.25,"upper_bound":3159496.09375,"unit":"ns"},"median_abs_dev":{"estimate":20885.7564792037,"lower_bound":14075.016518868506,"upper_bound":30667.44146179408,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_65","report_directory":"","iteration_count":[15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15,15],"measured_values":[51251459.0,50718709.0,50686333.0,51037666.0,51019959.0,50690000.0,50680625.0,50673500.0,50633167.0,50676000.0,51151709.0,50761791.0,50966750.0,50897416.0,50961333.0,50965208.0,50792667.0,51059583.0,50978500.0,51115833.0,51164541.0,50729542.0,50620792.0,51175542.0,50663333.0,50818583.0,50584667.0,50769708.0,50720917.0,50703958.0,51326833.0,50742708.0,51205333.0,50680292.0,51200042.0,50734416.0,50786708.0,50795834.0,50873958.0,51267625.0,51008833.0,50840750.0,50789875.0,51255292.0,50679917.0,50626916.0,50684500.0,50613542.0,50624459.0,50654917.0,51240833.0,50815833.0,51075459.0,50742125.0,51130917.0,50673333.0,50832500.0,50869875.0,50923500.0,50902292.0,51714750.0,50804083.0,51216917.0,50924916.0,50665792.0,50701750.0,50589250.0,50650292.0,50654542.0,50713208.0,51222791.0,50893333.0,51046250.0,50899333.0,50963916.0,50657084.0,50985000.0,50701208.0,50837000.0,50613916.0,51037333.0,50969709.0,51106166.0,50757541.0,50785375.0,50588416.0,50602875.0,50621542.0,50584459.0,50675000.0,51012583.0,51253916.0,50803208.0,51155416.0,50701000.0,50841375.0,50904375.0,50820166.0,50741958.0,50647458.0],"unit":"ns","throughput":[{"per_iteration":65,"unit":"elements"}],"typical":{"estimate":3390889.606666667,"lower_bound":3388098.1339833327,"upper_bound":3393796.6396,"unit":"ns"},"mean":{"estimate":3390889.606666667,"lower_bound":3388098.1339833327,"upper_bound":3393796.6396,"unit":"ns"},"median":{"estimate":3386909.7,"lower_bound":3382847.2,"upper_bound":3391597.2,"unit":"ns"},"median_abs_dev":{"estimate":14177.312828302913,"lower_bound":9900.506104230604,"upper_bound":17980.626540779976,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_70","report_directory":"","iteration_count":[14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14,14],"measured_values":[51150583.0,50748500.0,51130708.0,51175709.0,50928333.0,50916375.0,50978917.0,50700750.0,50634459.0,51880042.0,51487875.0,50937125.0,51323416.0,51028666.0,51599584.0,50781958.0,50795417.0,50903333.0,50777167.0,50780667.0,51423250.0,51027666.0,50850458.0,51262834.0,50770000.0,50689125.0,51302583.0,50679041.0,50656042.0,50698083.0,51335250.0,50918042.0,51432250.0,51179833.0,51319667.0,50777291.0,50894625.0,50808708.0,50936708.0,50736167.0,51038458.0,50876708.0,50920958.0,51097833.0,50753625.0,50697542.0,50728625.0,51052834.0,50601167.0,50625791.0,51114250.0,51023833.0,51090667.0,50985042.0,51245166.0,50822542.0,51002541.0,50774125.0,50845250.0,50708541.0,51033666.0,50995000.0,51034500.0,50764208.0,50704625.0,50661041.0,50635416.0,51225708.0,50786875.0,50728417.0,50893042.0,51302625.0,50916958.0,51100333.0,50815542.0,50850250.0,50847500.0,50720167.0,50848166.0,50720708.0,50757583.0,51157583.0,51323500.0,50653833.0,50790000.0,52176208.0,50938125.0,50678125.0,51232458.0,50625500.0,50757167.0,51197750.0,50840750.0,51132625.0,50625334.0,50792125.0,50765792.0,50825458.0,50757584.0,50678625.0],"unit":"ns","throughput":[{"per_iteration":70,"unit":"elements"}],"typical":{"estimate":3639038.2692857143,"lower_bound":3635316.304999999,"upper_bound":3643024.0208571446,"unit":"ns"},"mean":{"estimate":3639038.2692857143,"lower_bound":3635316.304999999,"upper_bound":3643024.0208571446,"unit":"ns"},"median":{"estimate":3633113.0714285714,"lower_bound":3628962.8214285714,"upper_bound":3638336.285714286,"unit":"ns"},"median_abs_dev":{"estimate":16902.063299928344,"lower_bound":11492.267795971176,"upper_bound":21149.076824529027,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_75","report_directory":"","iteration_count":[13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13],"measured_values":[50834000.0,51041542.0,50284834.0,50139291.0,50136667.0,50280708.0,50233500.0,50457875.0,50544458.0,51144208.0,50467208.0,50531583.0,50703834.0,50312167.0,50398959.0,50185083.0,50106750.0,51050875.0,50654125.0,50201750.0,50227750.0,50721917.0,50108167.0,50201834.0,50086542.0,50092750.0,50131000.0,50501208.0,50363542.0,50907667.0,50320625.0,50662000.0,50564291.0,50478250.0,50220042.0,54021958.0,50138667.0,50290000.0,50451875.0,50224083.0,50276125.0,50633583.0,50248375.0,50127625.0,50208000.0,50115584.0,50216292.0,50615958.0,50382834.0,50705000.0,50746875.0,50742334.0,50610625.0,50401917.0,50270292.0,50232667.0,50189916.0,50332291.0,50811625.0,50351750.0,50085542.0,50677750.0,50352541.0,50205083.0,50207334.0,50217875.0,50185292.0,50853209.0,50289000.0,50785667.0,50148500.0,51103458.0,50438625.0,50426917.0,50470625.0,50179583.0,50259542.0,50328542.0,50390250.0,50124334.0,50181958.0,50557333.0,50465458.0,50473375.0,50187250.0,50168750.0,50288458.0,50636791.0,50206250.0,51469125.0,51612167.0,52052625.0,50577791.0,50721167.0,50324667.0,50165500.0,50107542.0,50403875.0,50579042.0,50214333.0],"unit":"ns","throughput":[{"per_iteration":75,"unit":"elements"}],"typical":{"estimate":3882897.214615384,"lower_bound":3876270.6207115413,"upper_bound":3891051.8071346134,"unit":"ns"},"mean":{"estimate":3882897.214615384,"lower_bound":3876270.6207115413,"upper_bound":3891051.8071346134,"unit":"ns"},"median":{"estimate":3872463.1153846155,"lower_bound":3867729.1923076925,"upper_bound":3879894.230769231,"unit":"ns"},"median_abs_dev":{"estimate":18204.616984495537,"lower_bound":13407.55072350688,"upper_bound":26002.812007396267,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_80","report_directory":"","iteration_count":[13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13,13],"measured_values":[53914542.0,53343666.0,53788708.0,53425959.0,53658250.0,53410625.0,53429833.0,53445333.0,53543625.0,53731583.0,53512416.0,53300042.0,53235375.0,53295292.0,53265459.0,53262750.0,53351334.0,53811417.0,53583708.0,53776416.0,53336250.0,53733042.0,53398083.0,53433292.0,53319541.0,53312125.0,53746417.0,53632500.0,53379208.0,53749208.0,53415667.0,53267333.0,53228000.0,53277042.0,53246500.0,53477917.0,53758834.0,53699291.0,53769250.0,53634958.0,53748833.0,53440625.0,53411208.0,53289667.0,53250750.0,53480291.0,53958125.0,53327125.0,53796084.0,53436500.0,53308417.0,53399292.0,53226042.0,53226709.0,53433875.0,53816833.0,53867209.0,53519042.0,53927125.0,53681583.0,53588334.0,53327750.0,53393750.0,53427625.0,53623583.0,54060625.0,53286583.0,53903500.0,53322875.0,53287542.0,53328875.0,53340083.0,53253875.0,53640916.0,53644083.0,53778625.0,53480750.0,53892209.0,53548917.0,53543084.0,53376000.0,53345667.0,53290084.0,53752833.0,53397584.0,53833125.0,54287666.0,53567375.0,53341792.0,53426208.0,53314666.0,53261208.0,53901000.0,53404750.0,53642042.0,53374000.0,53921708.0,53432333.0,53592666.0,53350917.0],"unit":"ns","throughput":[{"per_iteration":80,"unit":"elements"}],"typical":{"estimate":4116871.7953846157,"lower_bound":4113499.404500001,"upper_bound":4120375.4933653823,"unit":"ns"},"mean":{"estimate":4116871.7953846157,"lower_bound":4113499.404500001,"upper_bound":4120375.4933653823,"unit":"ns"},"median":{"estimate":4110275.653846154,"lower_bound":4108057.6923076925,"upper_bound":4118698.769230769,"unit":"ns"},"median_abs_dev":{"estimate":16710.156211027865,"lower_bound":12706.737120564274,"upper_bound":23232.113495239977,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_85","report_directory":"","iteration_count":[12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12,12],"measured_values":[52843458.0,52371209.0,52347875.0,52428875.0,52297083.0,52079792.0,52030125.0,52021125.0,52288083.0,52185000.0,52541167.0,52124250.0,52041584.0,52062542.0,52139708.0,52412250.0,52323041.0,52419125.0,52371250.0,52601708.0,52630958.0,52272791.0,52543000.0,52383875.0,52290625.0,52157667.0,52139875.0,52389000.0,52457041.0,52280375.0,52471458.0,52051667.0,52137625.0,52084208.0,52183625.0,52524208.0,52476208.0,52515000.0,52631250.0,52463917.0,52341916.0,52726458.0,52219750.0,52332250.0,52162250.0,52189875.0,52186000.0,52238166.0,52145916.0,52351083.0,52104167.0,52060083.0,52069208.0,52030000.0,52072000.0,52666916.0,52562917.0,52276541.0,52362166.0,52304791.0,52390041.0,52133209.0,52268792.0,52168375.0,52080500.0,52204000.0,52409125.0,52327209.0,52264375.0,52272167.0,51995208.0,52032292.0,52003083.0,52005459.0,52155166.0,52404542.0,52597000.0,52346000.0,52397792.0,52368833.0,52121166.0,52282500.0,52089917.0,52064959.0,52054041.0,52228291.0,52108250.0,52030959.0,52395208.0,52221916.0,52000500.0,52050792.0,52046917.0,52064208.0,52475459.0,52225000.0,52656125.0,52158417.0,52510667.0,52118500.0],"unit":"ns","throughput":[{"per_iteration":85,"unit":"elements"}],"typical":{"estimate":4355952.53,"lower_bound":4352853.457874999,"upper_bound":4359160.341687501,"unit":"ns"},"mean":{"estimate":4355952.53,"lower_bound":4352853.457874999,"upper_bound":4359160.341687501,"unit":"ns"},"median":{"estimate":4355548.625,"lower_bound":4348734.375,"upper_bound":4359171.833333334,"unit":"ns"},"median_abs_dev":{"estimate":17788.666909187803,"lower_bound":14445.033318549855,"upper_bound":22308.56134269387,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_90","report_directory":"","iteration_count":[11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11],"measured_values":[51026792.0,51075333.0,50941250.0,50898375.0,50487042.0,50410125.0,50783792.0,50430416.0,50433833.0,50377792.0,50860333.0,50602875.0,50489584.0,50370709.0,50355708.0,51074125.0,50766958.0,50828166.0,50564583.0,50465541.0,51035291.0,50553125.0,51013625.0,50483250.0,50827042.0,50471292.0,50592500.0,50707417.0,50432792.0,50537917.0,50815875.0,50758958.0,50379375.0,50383542.0,50396083.0,50759959.0,50648750.0,50959417.0,50468959.0,50583250.0,50716958.0,50953375.0,50928333.0,50849875.0,50336625.0,50592875.0,50451000.0,50429875.0,50338541.0,50388375.0,50495375.0,51013750.0,50436000.0,50479292.0,50399292.0,50808500.0,50646542.0,50813000.0,50411500.0,50706750.0,50442875.0,50975500.0,50626666.0,50721667.0,50491083.0,50624500.0,50646833.0,50399667.0,50438000.0,50357542.0,50424709.0,50906333.0,50519458.0,50513916.0,50380166.0,50926750.0,50651000.0,50902708.0,50513542.0,50634125.0,50444333.0,50834542.0,50625750.0,50414250.0,51174541.0,51093584.0,50369042.0,50541500.0,50490417.0,50635042.0,50382542.0,50881542.0,50617667.0,50536459.0,50527083.0,50878791.0,50760833.0,50655625.0,50424333.0,50618083.0],"unit":"ns","throughput":[{"per_iteration":90,"unit":"elements"}],"typical":{"estimate":4602955.075454545,"lower_bound":4599138.893590908,"upper_bound":4606892.808227274,"unit":"ns"},"mean":{"estimate":4602955.075454545,"lower_bound":4599138.893590908,"upper_bound":4606892.808227274,"unit":"ns"},"median":{"estimate":4599335.227272727,"lower_bound":4592174.181818182,"upper_bound":4604231.090909091,"unit":"ns"},"median_abs_dev":{"estimate":22592.80187162536,"lower_bound":14974.866252324924,"upper_bound":27576.898637685095,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_95","report_directory":"","iteration_count":[11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11,11],"measured_values":[53368542.0,53341041.0,53676291.0,54600583.0,53461334.0,53498250.0,53264708.0,52990167.0,53033458.0,53323084.0,53411125.0,53501292.0,53236167.0,53436750.0,53167833.0,53241916.0,53188667.0,53260917.0,53025959.0,53421750.0,53317500.0,53644791.0,53031250.0,53061334.0,53043375.0,53350584.0,53026375.0,53007167.0,52973917.0,53659542.0,53556000.0,53495875.0,53199916.0,53209416.0,53153416.0,53132791.0,53160458.0,53137000.0,53221625.0,53779000.0,53168541.0,53042584.0,53169917.0,53106791.0,53157041.0,53893291.0,53129375.0,53077750.0,53442541.0,53346375.0,53601083.0,53121959.0,53323042.0,53099000.0,53233667.0,53250000.0,53042875.0,53141291.0,53602041.0,53103916.0,53138833.0,53112959.0,53124834.0,53083708.0,53307250.0,53012084.0,53054000.0,53479958.0,53386208.0,53611250.0,53164708.0,53314667.0,53014333.0,53336042.0,53081625.0,53047000.0,53080125.0,53674833.0,53046792.0,53041708.0,53279458.0,53149041.0,52992000.0,53021750.0,53838417.0,53077375.0,53473458.0,53595333.0,53466667.0,53151541.0,53301833.0,53109000.0,53259583.0,53098708.0,53101417.0,53255500.0,53625458.0,53316458.0,53100167.0,53108542.0],"unit":"ns","throughput":[{"per_iteration":95,"unit":"elements"}],"typical":{"estimate":4842817.153636363,"lower_bound":4838518.587136362,"upper_bound":4847471.988022727,"unit":"ns"},"mean":{"estimate":4842817.153636363,"lower_bound":4838518.587136362,"upper_bound":4847471.988022727,"unit":"ns"},"median":{"estimate":4835844.681818182,"lower_bound":4831492.363636363,"upper_bound":4842013.227272727,"unit":"ns"},"median_abs_dev":{"estimate":17639.57014138017,"lower_bound":13009.275641766964,"upper_bound":24277.574568986893,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_100","report_directory":"","iteration_count":[10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10],"measured_values":[50857208.0,51474083.0,50835625.0,50714250.0,50893625.0,50670250.0,50731917.0,50889000.0,50734083.0,50667875.0,51306541.0,50688083.0,50621125.0,50618834.0,50803041.0,50613666.0,50567916.0,51253959.0,50910167.0,51163000.0,51259542.0,51300834.0,50864625.0,50839792.0,50724958.0,51203834.0,50640583.0,50948083.0,50659667.0,50651166.0,51345208.0,50839875.0,50621708.0,50626042.0,50696000.0,50729000.0,50712125.0,51270833.0,51008584.0,50791542.0,51212167.0,50944667.0,51246625.0,50948333.0,50676542.0,50699625.0,50833958.0,51189500.0,50733709.0,50734375.0,51206333.0,51284459.0,50921292.0,50666875.0,50721833.0,50742250.0,50964458.0,51027084.0,51285875.0,50744584.0,50938375.0,50958375.0,51234250.0,51333083.0,50724208.0,50676667.0,50876875.0,51102375.0,50622458.0,50694750.0,50836167.0,51224583.0,50747208.0,50939000.0,50593042.0,50659209.0,51145666.0,51073916.0,51187292.0,50698375.0,50978459.0,51175041.0,50877750.0,51240750.0,50735542.0,50689958.0,50929958.0,50822417.0,51975667.0,51352334.0,50982834.0,51178375.0,50772541.0,50687334.0,50673333.0,50739709.0,51388709.0,51139792.0,50787583.0,50767333.0],"unit":"ns","throughput":[{"per_iteration":100,"unit":"elements"}],"typical":{"estimate":5091694.085999999,"lower_bound":5086725.414774999,"upper_bound":5096918.769499999,"unit":"ns"},"mean":{"estimate":5091694.085999999,"lower_bound":5086725.414774999,"upper_bound":5096918.769499999,"unit":"ns"},"median":{"estimate":5083983.35,"lower_bound":5075987.449999999,"upper_bound":5092995.8,"unit":"ns"},"median_abs_dev":{"estimate":22554.052099584842,"lower_bound":16200.444042384348,"upper_bound":34575.565726161556,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_105","report_directory":"","iteration_count":[10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10,10],"measured_values":[53618792.0,53028292.0,53270125.0,53638417.0,53475250.0,53194250.0,54739917.0,53063166.0,53281542.0,53071459.0,53519666.0,53208417.0,53548417.0,53088916.0,52986959.0,53695042.0,53031750.0,53174583.0,52977542.0,53360542.0,53420583.0,53428625.0,53566875.0,53430042.0,53273083.0,53202375.0,53096625.0,53321125.0,53283417.0,53448875.0,53417750.0,53402625.0,53209750.0,53068500.0,53671583.0,53302833.0,53079208.0,53428458.0,53149167.0,53794333.0,53452125.0,53567750.0,53276708.0,53290625.0,53204083.0,53169000.0,53185750.0,53097542.0,53337583.0,53626125.0,53291750.0,53091250.0,53018292.0,52968542.0,53659208.0,53038833.0,53222834.0,53006875.0,53511584.0,53697791.0,53438750.0,53257750.0,53261583.0,53139291.0,53211167.0,53130083.0,53065375.0,53326667.0,53662667.0,53474166.0,53310833.0,53098334.0,53058416.0,53005333.0,53833417.0,53178834.0,52952833.0,53393708.0,53319833.0,53339459.0,53095542.0,53283666.0,53130042.0,53315875.0,53152791.0,53079917.0,53371500.0,53651584.0,56563000.0,66575625.0,56259750.0,54155250.0,53694625.0,53744750.0,53645125.0,54315500.0,55115541.0,54109709.0,53140334.0,54358042.0],"unit":"ns","throughput":[{"per_iteration":105,"unit":"elements"}],"typical":{"estimate":5356899.7930000005,"lower_bound":5335904.669699999,"upper_bound":5389750.551925,"unit":"ns"},"mean":{"estimate":5356899.7930000005,"lower_bound":5335904.669699999,"upper_bound":5389750.551925,"unit":"ns"},"median":{"estimate":5329729.15,"lower_bound":5324220.85,"upper_bound":5337712.5,"unit":"ns"},"median_abs_dev":{"estimate":29556.149385273595,"lower_bound":21300.06904184846,"upper_bound":36240.30681310663,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_110","report_directory":"","iteration_count":[9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9],"measured_values":[50284291.0,50649833.0,50048916.0,50255084.0,50126500.0,49933708.0,49984292.0,50207375.0,50276833.0,49987667.0,50438333.0,50156583.0,50041750.0,49935542.0,49846583.0,49984583.0,50032583.0,50088375.0,50352416.0,50399042.0,50171916.0,50868541.0,49918334.0,50161417.0,50112333.0,49961375.0,49927250.0,50404000.0,50090000.0,49902000.0,50167584.0,50318291.0,49964042.0,49861375.0,49894583.0,49938000.0,49959917.0,50212375.0,50311666.0,50476500.0,50008500.0,50554625.0,50536917.0,50209542.0,50025667.0,49861000.0,49974125.0,50218708.0,50300500.0,49862875.0,49877167.0,50311084.0,50051708.0,49910375.0,49856791.0,49938875.0,49953792.0,50253583.0,50787458.0,50269083.0,49859791.0,50552958.0,50041250.0,50620583.0,50282333.0,49927708.0,49928833.0,50149708.0,50044083.0,49957458.0,49925625.0,50193666.0,50425834.0,49919209.0,49901250.0,49990125.0,49933459.0,50135083.0,50545125.0,50270833.0,49946958.0,50246125.0,50564041.0,50135375.0,50567417.0,49996291.0,49857167.0,50163625.0,50152959.0,49943958.0,49970208.0,50132875.0,50540166.0,49959416.0,50027625.0,49925875.0,49928333.0,49988958.0,50424625.0,50435917.0],"unit":"ns","throughput":[{"per_iteration":110,"unit":"elements"}],"typical":{"estimate":5571029.989999997,"lower_bound":5566013.495527777,"upper_bound":5576255.05536111,"unit":"ns"},"mean":{"estimate":5571029.989999997,"lower_bound":5566013.495527777,"upper_bound":5576255.05536111,"unit":"ns"},"median":{"estimate":5563337.944444444,"lower_bound":5555143.444444444,"upper_bound":5572962.944444444,"unit":"ns"},"median_abs_dev":{"estimate":23447.0714837315,"lower_bound":16761.44545825877,"upper_bound":32713.32131922199,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_115","report_directory":"","iteration_count":[9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9],"measured_values":[52422042.0,52389708.0,52113417.0,52117500.0,52622167.0,52251375.0,52209958.0,54049333.0,52165541.0,52312083.0,52512000.0,52123250.0,52132625.0,52236458.0,54868542.0,56641208.0,53507500.0,52617416.0,52171250.0,52883791.0,52284208.0,52076541.0,52062125.0,52483792.0,52634417.0,53256958.0,52278375.0,52318500.0,52193083.0,52637875.0,52050458.0,52196167.0,52091875.0,52682084.0,52513375.0,52374125.0,52267791.0,52163833.0,52296458.0,52443333.0,52061625.0,52074666.0,52392875.0,52398750.0,52175083.0,52074209.0,52170917.0,52802042.0,52547375.0,52325125.0,52208542.0,52146583.0,52476417.0,52754334.0,52228833.0,52312541.0,52207000.0,52266709.0,52493375.0,52055375.0,52122875.0,52290584.0,52511250.0,52128750.0,52037084.0,52090125.0,52076459.0,52863416.0,52719125.0,52090291.0,52150542.0,52393667.0,52783375.0,52502625.0,52326416.0,52269333.0,52261417.0,52321792.0,52505209.0,52109041.0,52403917.0,52521459.0,52141042.0,52100125.0,52072833.0,52092500.0,52028542.0,52915542.0,52363792.0,52142417.0,52201083.0,52712458.0,52839541.0,52200791.0,52339042.0,52097584.0,52464708.0,52460250.0,52302625.0,52336833.0],"unit":"ns","throughput":[{"per_iteration":115,"unit":"elements"}],"typical":{"estimate":5825683.747777779,"lower_bound":5814565.427000001,"upper_bound":5839962.609583337,"unit":"ns"},"mean":{"estimate":5825683.747777779,"lower_bound":5814565.427000001,"upper_bound":5839962.609583337,"unit":"ns"},"median":{"estimate":5810391.222222222,"lower_bound":5803203.666666667,"upper_bound":5816701.388888889,"unit":"ns"},"median_abs_dev":{"estimate":28155.644266803753,"lower_bound":20481.789169709195,"upper_bound":33303.64320874214,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_120","report_directory":"","iteration_count":[9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9,9],"measured_values":[54886084.0,54338916.0,54384917.0,54243667.0,55204333.0,54759375.0,54424416.0,54551000.0,54307959.0,54700083.0,54315042.0,54270084.0,54555250.0,54722000.0,54334583.0,54279375.0,54301375.0,54194750.0,54672875.0,54561583.0,54428125.0,54294209.0,54581000.0,55168834.0,54353042.0,54644209.0,54301959.0,54571334.0,54979459.0,54293584.0,54345167.0,54887500.0,54333584.0,54377417.0,54338583.0,54248417.0,54320917.0,54728500.0,54458625.0,54257875.0,54388000.0,54923209.0,54921959.0,54603542.0,54426709.0,54467458.0,54478375.0,54817416.0,54322708.0,54699959.0,54612250.0,54330417.0,54303250.0,54265542.0,54653000.0,54281917.0,54697042.0,54194250.0,54234500.0,54819833.0,54796625.0,54557583.0,54791791.0,54375875.0,54425125.0,54371500.0,54684250.0,54359583.0,54737416.0,54958500.0,54856584.0,54213625.0,54911000.0,54620666.0,54233333.0,54845417.0,54523083.0,54450625.0,55030625.0,54377042.0,54954208.0,54814292.0,54565750.0,54318750.0,54542458.0,54680958.0,54883792.0,54519834.0,54273750.0,54268834.0,54295167.0,54512459.0,54768625.0,54287292.0,54811959.0,54387417.0,54856167.0,54579709.0,54412458.0,54433583.0],"unit":"ns","throughput":[{"per_iteration":120,"unit":"elements"}],"typical":{"estimate":6059610.05888889,"lower_bound":6054357.877583333,"upper_bound":6065012.164388889,"unit":"ns"},"mean":{"estimate":6059610.05888889,"lower_bound":6054357.877583333,"upper_bound":6065012.164388889,"unit":"ns"},"median":{"estimate":6052546.277777778,"lower_bound":6043111.111111111,"upper_bound":6062939.833333334,"unit":"ns"},"median_abs_dev":{"estimate":29360.17437875317,"lower_bound":20756.399631499553,"upper_bound":38952.68681178468,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_125","report_directory":"","iteration_count":[8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8],"measured_values":[50514209.0,50445666.0,50149208.0,50130500.0,50230958.0,50154292.0,50167667.0,50774291.0,50576833.0,50253125.0,50844750.0,50180459.0,50497500.0,50380958.0,50112125.0,50622917.0,50531333.0,50277250.0,50089792.0,50118291.0,50781458.0,50211125.0,50131583.0,50112875.0,50122834.0,50194583.0,50102833.0,50867416.0,50459959.0,50239917.0,50673666.0,50373416.0,50688166.0,50190208.0,50166833.0,50135917.0,50820625.0,50280000.0,50100458.0,50182916.0,50608041.0,50386791.0,50141166.0,50126583.0,50201375.0,50177083.0,50319292.0,50792667.0,50372291.0,50286166.0,50714416.0,50414583.0,50658834.0,50278625.0,50295125.0,50147000.0,50404333.0,50895667.0,50203084.0,50200584.0,50450750.0,50673209.0,50355333.0,50184125.0,50184208.0,50159250.0,50338334.0,50774959.0,50381166.0,50327875.0,50336583.0,50677041.0,50633334.0,50202208.0,50232292.0,50124292.0,50534833.0,50269709.0,50761792.0,50204917.0,50171458.0,50676208.0,50353500.0,50197334.0,50158625.0,50271542.0,50376625.0,50879792.0,50329208.0,50304500.0,50266542.0,50305625.0,50755208.0,50196125.0,50125500.0,53652583.0,51524917.0,50301083.0,50602458.0,50500708.0],"unit":"ns","throughput":[{"per_iteration":125,"unit":"elements"}],"typical":{"estimate":6301585.46125,"lower_bound":6292832.028125,"upper_bound":6312952.2313125,"unit":"ns"},"mean":{"estimate":6301585.46125,"lower_bound":6292832.028125,"upper_bound":6312952.2313125,"unit":"ns"},"median":{"estimate":6287848.9375,"lower_bound":6281640.625,"upper_bound":6296536.375,"unit":"ns"},"median_abs_dev":{"estimate":26659.742076694965,"lower_bound":18891.566852107644,"upper_bound":36624.85479134135,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_130","report_directory":"","iteration_count":[8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8],"measured_values":[52493209.0,52459167.0,52151834.0,52191625.0,52218958.0,52321583.0,52128500.0,52088541.0,52053667.0,52074750.0,52696375.0,52576042.0,52130333.0,52026292.0,52176583.0,52640250.0,52366625.0,52310750.0,52195833.0,52127500.0,52589500.0,52206500.0,52141583.0,52090042.0,52347459.0,52138167.0,52121667.0,52069959.0,52059708.0,52239084.0,52531375.0,52573500.0,52031166.0,52122041.0,52567666.0,52506875.0,52182584.0,52234375.0,52060334.0,52332292.0,52384958.0,52179958.0,52094042.0,52559708.0,52380875.0,52046167.0,52063375.0,52140750.0,52098542.0,52776625.0,52037250.0,52636208.0,52073084.0,52463542.0,52648125.0,52125542.0,52312125.0,52121708.0,52400166.0,52696292.0,52129667.0,52090292.0,52517000.0,52548709.0,52088167.0,52018500.0,52091625.0,52210958.0,52295917.0,52708083.0,52078542.0,52493875.0,52554625.0,52675166.0,52172083.0,52362750.0,52121500.0,52264667.0,52341500.0,52668000.0,52146041.0,52323708.0,52554958.0,52110667.0,52144458.0,52144166.0,52107083.0,52116833.0,52427167.0,52606000.0,52099041.0,52797583.0,52809375.0,52173459.0,52339709.0,52281583.0,52246167.0,52347500.0,52435542.0,52357875.0],"unit":"ns","throughput":[{"per_iteration":130,"unit":"elements"}],"typical":{"estimate":6537267.34625,"lower_bound":6532014.7914375,"upper_bound":6542644.347531251,"unit":"ns"},"mean":{"estimate":6537267.34625,"lower_bound":6532014.7914375,"upper_bound":6542644.347531251,"unit":"ns"},"median":{"estimate":6528333.3125,"lower_bound":6520330.8125,"upper_bound":6540739.625,"unit":"ns"},"median_abs_dev":{"estimate":25246.63897678256,"lower_bound":16686.65575406516,"upper_bound":37960.767713561654,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_135","report_directory":"","iteration_count":[8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8],"measured_values":[54326416.0,54514917.0,54085292.0,54205708.0,53963917.0,54024000.0,54155250.0,54118334.0,54027250.0,54466208.0,54295375.0,54202291.0,54129500.0,54055958.0,54049500.0,54056541.0,54293667.0,54451041.0,54179000.0,54523708.0,54515709.0,54437250.0,54226375.0,54046458.0,53989334.0,54360292.0,54005250.0,54020250.0,54316875.0,54418292.0,54072750.0,53971000.0,53946334.0,54040708.0,54061375.0,54614584.0,54400500.0,54417084.0,54555125.0,54781084.0,54450208.0,54120584.0,53968000.0,54175583.0,54129416.0,54083541.0,53982042.0,54647416.0,54079750.0,54149334.0,54001750.0,54050791.0,54042875.0,54663000.0,55585250.0,54063916.0,54372417.0,54406708.0,54395917.0,54512791.0,53976459.0,54010250.0,54284250.0,53989875.0,54030459.0,54065834.0,54622042.0,54049375.0,54045166.0,54004708.0,54022750.0,54189125.0,54591417.0,54219917.0,54325875.0,55124500.0,55347333.0,54809417.0,54536708.0,54019792.0,54368167.0,54106667.0,54039584.0,53945875.0,54085833.0,54494750.0,54088417.0,54112500.0,53994666.0,53966000.0,54503250.0,54408459.0,54284583.0,54183084.0,54497375.0,54861750.0,54075291.0,54596625.0,54101583.0,54325542.0],"unit":"ns","throughput":[{"per_iteration":135,"unit":"elements"}],"typical":{"estimate":6783133.805,"lower_bound":6776101.12821875,"upper_bound":6790677.08159375,"unit":"ns"},"mean":{"estimate":6783133.805,"lower_bound":6776101.12821875,"upper_bound":6790677.08159375,"unit":"ns"},"median":{"estimate":6770677.0625,"lower_bound":6761052.125,"upper_bound":6785572.875,"unit":"ns"},"median_abs_dev":{"estimate":29679.05692309141,"lower_bound":17694.733390230685,"upper_bound":40532.0593804121,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_140","report_directory":"","iteration_count":[8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8],"measured_values":[56455708.0,56121583.0,56387000.0,56150375.0,56465875.0,55902958.0,55977542.0,55957958.0,56348417.0,56473875.0,56132541.0,55999458.0,56563459.0,56064791.0,56266375.0,55995792.0,56074500.0,56087125.0,56264750.0,56203042.0,56785000.0,58894000.0,71135750.0,59312250.0,57178459.0,56603791.0,56352833.0,56897875.0,56070333.0,56283917.0,56235458.0,56173834.0,56195667.0,56086542.0,56207000.0,56039666.0,56617875.0,56361792.0,56379208.0,55989667.0,58022291.0,56677583.0,58272209.0,56010000.0,55999250.0,56520125.0,56114292.0,56301000.0,56439833.0,57234792.0,57454291.0,56672542.0,56066125.0,55955167.0,56552166.0,56310250.0,56218042.0,56023625.0,55982542.0,56625875.0,55977875.0,55987000.0,55960834.0,56313375.0,56173000.0,56552500.0,55971542.0,56239042.0,56026833.0,56114375.0,56107292.0,55903666.0,56320250.0,56607625.0,56129958.0,55965541.0,55895583.0,55865583.0,56681041.0,55938042.0,56022250.0,56468042.0,56265792.0,56503750.0,56129333.0,56430541.0,55977959.0,56141333.0,56114125.0,55915125.0,56201166.0,56650166.0,56045125.0,55921834.0,55927625.0,55942917.0,56094458.0,56767959.0,56099834.0,55980250.0],"unit":"ns","throughput":[{"per_iteration":140,"unit":"elements"}],"typical":{"estimate":7061845.7275,"lower_bound":7033229.725374999,"upper_bound":7106534.92125,"unit":"ns"},"mean":{"estimate":7061845.7275,"lower_bound":7033229.725374999,"upper_bound":7106534.92125,"unit":"ns"},"median":{"estimate":7023093.8125,"lower_bound":7014281.25,"upper_bound":7034356.8125,"unit":"ns"},"median_abs_dev":{"estimate":37061.20017953217,"lower_bound":25995.629950985312,"upper_bound":48265.578830614686,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_145","report_directory":"","iteration_count":[7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7],"measured_values":[50875334.0,51291000.0,51287667.0,50821959.0,51102417.0,50985375.0,50672917.0,50961250.0,50702084.0,50806083.0,51037000.0,51077125.0,50711416.0,50709166.0,50646375.0,50642916.0,50745958.0,50595834.0,50677500.0,50728584.0,51188708.0,50669125.0,51670833.0,50651625.0,51164875.0,50744958.0,50749500.0,50831334.0,50637959.0,50831375.0,51137667.0,51109667.0,50693292.0,50739459.0,50703542.0,50561125.0,50785208.0,50655916.0,50571791.0,50889000.0,51190041.0,50755750.0,51333209.0,51110167.0,51133250.0,50714041.0,50733917.0,50893625.0,50640083.0,51132875.0,51328292.0,50954458.0,50727042.0,50642583.0,50610667.0,50582209.0,50632917.0,50529875.0,50568250.0,50849667.0,51120208.0,51014875.0,50820625.0,51487750.0,50857750.0,50651375.0,50825083.0,50725292.0,50606334.0,50815125.0,51172584.0,50965208.0,50557584.0,50683416.0,50651583.0,50696250.0,50623417.0,50646334.0,50652083.0,50637916.0,51404042.0,51292750.0,50752791.0,51183083.0,51119375.0,50634750.0,51051125.0,50630208.0,50642916.0,50732708.0,51188625.0,50799875.0,50691750.0,50599333.0,50703583.0,50616791.0,50633458.0,50637209.0,50603875.0,50613208.0],"unit":"ns","throughput":[{"per_iteration":145,"unit":"elements"}],"typical":{"estimate":7262671.548571428,"lower_bound":7255934.035714285,"upper_bound":7269731.623928573,"unit":"ns"},"mean":{"estimate":7262671.548571428,"lower_bound":7255934.035714285,"upper_bound":7269731.623928573,"unit":"ns"},"median":{"estimate":7248098.285714285,"lower_bound":7243261.928571429,"upper_bound":7259303.571428572,"unit":"ns"},"median_abs_dev":{"estimate":22265.580504706337,"lower_bound":16643.765262013185,"upper_bound":37245.87653875351,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_150","report_directory":"","iteration_count":[7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7],"measured_values":[52971792.0,52665500.0,52792667.0,52379125.0,52471583.0,52467250.0,52655500.0,52640000.0,52506750.0,52476750.0,52732084.0,52456708.0,52357875.0,52348500.0,52332917.0,52286875.0,52276958.0,52337583.0,52381291.0,52956708.0,52770250.0,52895041.0,52390958.0,52497959.0,52566750.0,52423708.0,52790417.0,52790375.0,52515958.0,52707333.0,52480542.0,52404083.0,52301625.0,52313750.0,52352125.0,52263875.0,52302541.0,52532458.0,52881583.0,52759041.0,52810834.0,52426416.0,52445583.0,52547792.0,52351042.0,52411417.0,52980542.0,52730625.0,52401542.0,52967791.0,52442083.0,52501791.0,52379625.0,52342125.0,52400709.0,52329125.0,52405542.0,52900000.0,52577209.0,53098500.0,52474291.0,52429416.0,52493292.0,52355750.0,52293541.0,52431042.0,52864250.0,52342917.0,52783875.0,52513500.0,52341709.0,52270500.0,52300042.0,52287083.0,52229625.0,52333584.0,52898333.0,52484792.0,52967417.0,52602500.0,52345667.0,52823833.0,52372041.0,52374000.0,52398750.0,52539416.0,52890458.0,52430708.0,53019375.0,52443916.0,52288917.0,52252667.0,52332209.0,52431542.0,52382208.0,52815875.0,52549875.0,52921458.0,52846250.0,52404542.0],"unit":"ns","throughput":[{"per_iteration":150,"unit":"elements"}],"typical":{"estimate":7504923.21,"lower_bound":7498769.564571429,"upper_bound":7511283.982928569,"unit":"ns"},"mean":{"estimate":7504923.21,"lower_bound":7498769.564571429,"upper_bound":7511283.982928569,"unit":"ns"},"median":{"estimate":7493020.785714285,"lower_bound":7487345.285714285,"upper_bound":7500255.857142857,"unit":"ns"},"median_abs_dev":{"estimate":24268.78486914284,"lower_bound":17711.66878555522,"upper_bound":36244.28759403549,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_155","report_directory":"","iteration_count":[7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7],"measured_values":[54463042.0,54459416.0,54071583.0,54057708.0,54010292.0,54073042.0,54669000.0,54092792.0,53958875.0,54400791.0,54371208.0,54583834.0,54188000.0,54571458.0,54013833.0,54192083.0,54077542.0,53987125.0,54126375.0,54852833.0,54102958.0,54001458.0,53930792.0,54006833.0,54024083.0,54105500.0,54701791.0,54110000.0,54393209.0,54741292.0,54195792.0,54408792.0,53985375.0,54126125.0,54375041.0,54103250.0,54107541.0,54246958.0,54454875.0,54003416.0,53999333.0,54016250.0,54028417.0,54042542.0,54079042.0,54670792.0,54010250.0,54703625.0,54469583.0,54203000.0,54517792.0,53993542.0,54290583.0,54190917.0,53993709.0,54055417.0,54428625.0,54669750.0,54563083.0,54049792.0,54068167.0,53981333.0,53970083.0,53916125.0,54212667.0,54455500.0,54619542.0,54089167.0,54515959.0,54635166.0,54195125.0,54232875.0,53986208.0,54061417.0,54030333.0,54591125.0,54235250.0,54109250.0,54137916.0,53976916.0,54832917.0,54075833.0,54001792.0,54120083.0,55083834.0,54480667.0,54322583.0,54336333.0,54038792.0,54244541.0,54068625.0,54030083.0,54077375.0,54348791.0,54579291.0,54027750.0,54051375.0,53972042.0,53998042.0,54076542.0],"unit":"ns","throughput":[{"per_iteration":155,"unit":"elements"}],"typical":{"estimate":7747299.102857143,"lower_bound":7740359.828107147,"upper_bound":7754590.24517857,"unit":"ns"},"mean":{"estimate":7747299.102857143,"lower_bound":7740359.828107147,"upper_bound":7754590.24517857,"unit":"ns"},"median":{"estimate":7729946.428571429,"lower_bound":7725291.714285715,"upper_bound":7742505.928571429,"unit":"ns"},"median_abs_dev":{"estimate":24092.143672279737,"lower_bound":17296.9643929173,"upper_bound":40881.95297420025,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_160","report_directory":"","iteration_count":[7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7],"measured_values":[56355459.0,56315083.0,56247417.0,55825000.0,55695709.0,55782709.0,55806458.0,55716666.0,55766833.0,56353458.0,55796042.0,55678584.0,55716500.0,55790500.0,55729375.0,56050208.0,55964333.0,56090084.0,56090917.0,56033959.0,56687750.0,55742417.0,56012917.0,55830709.0,55880834.0,55807500.0,55740958.0,56121416.0,56037000.0,55800125.0,55644584.0,55733750.0,55717625.0,56158584.0,55992542.0,56206083.0,56213292.0,55949958.0,56554541.0,55776166.0,55695000.0,55878958.0,55957375.0,55705833.0,55751000.0,55908709.0,56078167.0,55719750.0,55722042.0,55716000.0,55678084.0,56121500.0,55951709.0,56288708.0,55860708.0,56123459.0,56124917.0,56281750.0,55806000.0,55879375.0,56187167.0,55714291.0,55689500.0,55668542.0,56306292.0,55738125.0,55758625.0,55723042.0,55648250.0,56280292.0,55888208.0,56371791.0,55895458.0,56258292.0,56251292.0,55694834.0,56214250.0,55863417.0,55913583.0,55720667.0,55736875.0,55699958.0,56155209.0,55937417.0,55767917.0,55807583.0,55769833.0,56241459.0,55789000.0,56246166.0,55730000.0,56216041.0,56089208.0,55786958.0,55858333.0,56229958.0,55835875.0,55776125.0,55717791.0,55665500.0],"unit":"ns","throughput":[{"per_iteration":160,"unit":"elements"}],"typical":{"estimate":7990723.161428571,"lower_bound":7984314.889785718,"upper_bound":7997383.94892857,"unit":"ns"},"mean":{"estimate":7990723.161428571,"lower_bound":7984314.889785718,"upper_bound":7997383.94892857,"unit":"ns"},"median":{"estimate":7979931.5,"lower_bound":7970863.142857143,"upper_bound":7991059.571428572,"unit":"ns"},"median_abs_dev":{"estimate":30274.16196252599,"lower_bound":18748.747467144083,"upper_bound":42236.519850150304,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_165","report_directory":"","iteration_count":[7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7,7],"measured_values":[57975875.0,57635375.0,57959625.0,57545458.0,57415125.0,57460500.0,57633458.0,57482291.0,57602292.0,58248625.0,57451458.0,57380666.0,57350750.0,57376458.0,57583500.0,58015459.0,58485292.0,58894334.0,57706542.0,57906292.0,58179042.0,57323292.0,57369334.0,57977958.0,57477666.0,57447875.0,57771542.0,57966750.0,57486000.0,57349042.0,57378042.0,57405458.0,57766375.0,57797958.0,58003416.0,58053375.0,57665792.0,57945541.0,57414583.0,57366041.0,57578291.0,57547542.0,57374334.0,57382333.0,57985417.0,57853041.0,57702375.0,57455000.0,57376084.0,57548500.0,58028000.0,57921750.0,57574833.0,57998250.0,57839542.0,57734583.0,57376042.0,57428042.0,57785542.0,57391042.0,57394208.0,57550584.0,58119875.0,57959750.0,57331709.0,57342042.0,57338667.0,57889333.0,57561958.0,57803416.0,57590917.0,57999458.0,57840500.0,57436292.0,57363083.0,57607292.0,57438792.0,57379917.0,57353375.0,57571375.0,57853959.0,58054459.0,57394834.0,57386708.0,57770875.0,57734250.0,57993750.0,57500833.0,57595459.0,58183291.0,57548209.0,57480458.0,57535083.0,57901459.0,57449125.0,57445083.0,57310250.0,57549708.0,57820166.0,57808167.0],"unit":"ns","throughput":[{"per_iteration":165,"unit":"elements"}],"typical":{"estimate":8237605.241428568,"lower_bound":8229653.082214287,"upper_bound":8246015.73103572,"unit":"ns"},"mean":{"estimate":8237605.241428568,"lower_bound":8229653.082214287,"upper_bound":8246015.73103572,"unit":"ns"},"median":{"estimate":8225223.142857143,"lower_bound":8217592.214285715,"upper_bound":8243196.428571428,"unit":"ns"},"median_abs_dev":{"estimate":41847.97275704987,"lower_bound":30944.71280812021,"upper_bound":53691.299046789194,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_170","report_directory":"","iteration_count":[6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6],"measured_values":[51224666.0,51455958.0,50592916.0,50605458.0,50583834.0,50727833.0,50615834.0,50648125.0,50677709.0,50727042.0,51206958.0,51131792.0,51059750.0,50863541.0,50683709.0,50804625.0,50839625.0,50626792.0,50667958.0,50769125.0,51232292.0,50734584.0,50636625.0,50603958.0,50631416.0,50598583.0,50666542.0,50746000.0,50700000.0,50706416.0,51212125.0,50687750.0,51161792.0,51167541.0,50712208.0,50927625.0,50685666.0,50694916.0,50694333.0,50798583.0,51239084.0,50652250.0,50753917.0,50616584.0,54135667.0,55180500.0,52324542.0,50718250.0,50752125.0,51104041.0,51078958.0,50901916.0,51031500.0,51091333.0,52231500.0,51477834.0,81850209.0,52132333.0,51523959.0,51467208.0,51562792.0,51354000.0,51362375.0,50949209.0,50839167.0,50764750.0,50751083.0,50804250.0,51056250.0,50974208.0,50878334.0,50937958.0,50955500.0,50891959.0,51269875.0,50832250.0,51695375.0,52429792.0,52090250.0,53033208.0,53237667.0,52648875.0,52764500.0,52301083.0,51552708.0,52816417.0,51012875.0,51143541.0,51119833.0,52324667.0,51124250.0,50849917.0,50995459.0,51085625.0,51145042.0,51828500.0,51360250.0,51104166.0,51134416.0,51177333.0],"unit":"ns","throughput":[{"per_iteration":170,"unit":"elements"}],"typical":{"estimate":8588226.040000001,"lower_bound":8518389.016,"upper_bound":8705770.382166667,"unit":"ns"},"mean":{"estimate":8588226.040000001,"lower_bound":8518389.016,"upper_bound":8705770.382166667,"unit":"ns"},"median":{"estimate":8500694.5,"lower_bound":8477256.833333334,"upper_bound":8519972.166666666,"unit":"ns"},"median_abs_dev":{"estimate":69563.83786499547,"lower_bound":50953.99589538574,"upper_bound":94551.82492136909,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_175","report_directory":"","iteration_count":[6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6],"measured_values":[52716292.0,52707666.0,52937583.0,52639042.0,52705083.0,52753584.0,52530792.0,52512709.0,52423917.0,52631875.0,52495000.0,52480375.0,52641500.0,52594167.0,52429166.0,52420541.0,52549667.0,52496000.0,52561875.0,52509541.0,52622166.0,52829125.0,52606583.0,52422542.0,52713042.0,52701875.0,52558417.0,52528584.0,52384292.0,52541167.0,52505167.0,52455167.0,52474541.0,52497000.0,52309625.0,52515125.0,52382375.0,52398000.0,52492500.0,52582542.0,52722083.0,52585708.0,52675959.0,52846000.0,52605667.0,52450292.0,52574375.0,52421125.0,52397083.0,52486750.0,52449292.0,52423291.0,52581208.0,52505458.0,52430167.0,52555833.0,52341167.0,52662792.0,52489542.0,52753792.0,54101042.0,62034708.0,55030250.0,53517459.0,53800250.0,54327084.0,53772208.0,53864750.0,54483042.0,53766250.0,54036750.0,54085500.0,52845833.0,52574958.0,52116667.0,52103541.0,52066208.0,52089584.0,52289000.0,52056583.0,52652959.0,52177417.0,52031333.0,52018583.0,52778542.0,52097083.0,52222250.0,52128084.0,52091417.0,52204542.0,52074417.0,52192208.0,52555750.0,52050083.0,52087459.0,52132042.0,52141875.0,52445584.0,52298625.0,52151584.0],"unit":"ns","throughput":[{"per_iteration":175,"unit":"elements"}],"typical":{"estimate":8787845.496666668,"lower_bound":8758492.993333332,"upper_bound":8828772.485374998,"unit":"ns"},"mean":{"estimate":8787845.496666668,"lower_bound":8758492.993333332,"upper_bound":8828772.485374998,"unit":"ns"},"median":{"estimate":8751854.166666666,"lower_bound":8746729.166666666,"upper_bound":8761066.0,"unit":"ns"},"median_abs_dev":{"estimate":39103.69785577151,"lower_bound":26068.92598718681,"upper_bound":54907.59582519715,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_180","report_directory":"","iteration_count":[6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6],"measured_values":[54446541.0,55412292.0,55000667.0,53695500.0,53549208.0,53476458.0,53431833.0,53642958.0,53566834.0,53679042.0,53555167.0,54449208.0,54089292.0,54205916.0,53544625.0,53474375.0,53416417.0,53438042.0,53497250.0,53656833.0,53637584.0,53448708.0,53532125.0,53567458.0,54083166.0,53564166.0,53531791.0,53455000.0,54048250.0,53652416.0,53896875.0,53485333.0,54066125.0,53481000.0,53751625.0,53537875.0,53489791.0,53509125.0,53880208.0,53599958.0,53437958.0,53779541.0,53518208.0,53458500.0,53453416.0,53500292.0,55212041.0,54230708.0,54144375.0,54203917.0,53617583.0,54698792.0,55527084.0,54560625.0,59470250.0,56344750.0,56784250.0,94364959.0,56038583.0,53536667.0,55758792.0,53654750.0,53780417.0,53946292.0,55168833.0,56071750.0,55194833.0,55073000.0,55365292.0,55600125.0,55924000.0,55498625.0,53841791.0,54998167.0,54532750.0,54123667.0,53529500.0,53765458.0,53518000.0,53476416.0,53474875.0,53457750.0,53438250.0,53611750.0,55151042.0,54120584.0,55252292.0,55908792.0,55368833.0,54577791.0,55514750.0,54003958.0,53462000.0,53828333.0,53445625.0,53545458.0,53478083.0,53440875.0,53416958.0,53464292.0],"unit":"ns","throughput":[{"per_iteration":180,"unit":"elements"}],"typical":{"estimate":9098523.808333335,"lower_bound":9007826.02004167,"upper_bound":9251764.828416668,"unit":"ns"},"mean":{"estimate":9098523.808333335,"lower_bound":9007826.02004167,"upper_bound":9251764.828416668,"unit":"ns"},"median":{"estimate":8953927.083333332,"lower_bound":8931600.666666666,"upper_bound":9000659.666666666,"unit":"ns"},"median_abs_dev":{"estimate":67334.87235456567,"lower_bound":35963.3040115228,"upper_bound":129409.26996752594,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_185","report_directory":"","iteration_count":[6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6],"measured_values":[55714042.0,54896000.0,55015291.0,55095708.0,55018917.0,55020875.0,54908792.0,54936167.0,54970125.0,55080583.0,54948167.0,55095292.0,55393000.0,55279833.0,55010792.0,55016583.0,54984125.0,55012709.0,54935584.0,55172292.0,55543042.0,55029459.0,54946959.0,54942792.0,55126791.0,54935041.0,54917625.0,55026000.0,55069875.0,54997583.0,55205584.0,55421917.0,55346708.0,55232042.0,54969000.0,55059833.0,54971417.0,54848042.0,54877625.0,55566125.0,55103250.0,55003375.0,55409333.0,55449291.0,55392042.0,55404458.0,55311291.0,55321000.0,55670375.0,55461542.0,56462000.0,55392166.0,55321667.0,55147667.0,55463375.0,55245041.0,55172375.0,55742042.0,55565375.0,55189459.0,55113834.0,55475458.0,55111875.0,55436625.0,55212625.0,57094375.0,55924416.0,55196708.0,55400375.0,55524083.0,55094292.0,55128000.0,55102417.0,55113458.0,55236000.0,55517125.0,55054375.0,55751417.0,55315250.0,55158167.0,55152750.0,55191750.0,55219542.0,55443542.0,55162000.0,55597833.0,55071792.0,55375708.0,54958292.0,55178334.0,55170542.0,55368625.0,55534709.0,55088875.0,54997542.0,55102125.0,55702292.0,55162875.0,55101666.0,55190791.0],"unit":"ns","throughput":[{"per_iteration":185,"unit":"elements"}],"typical":{"estimate":9207496.534999998,"lower_bound":9197684.215625,"upper_bound":9218661.255291665,"unit":"ns"},"mean":{"estimate":9207496.534999998,"lower_bound":9197684.215625,"upper_bound":9218661.255291665,"unit":"ns"},"median":{"estimate":9194451.416666668,"lower_bound":9184725.666666668,"upper_bound":9202104.166666666,"unit":"ns"},"median_abs_dev":{"estimate":39443.33679974263,"lower_bound":27824.44790601684,"upper_bound":51020.96308294587,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_190","report_directory":"","iteration_count":[6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6],"measured_values":[57077542.0,56851500.0,57018583.0,56813250.0,56926584.0,57041167.0,56497625.0,56643208.0,56921625.0,56655500.0,56659708.0,56791666.0,56761333.0,56852583.0,56490583.0,56646209.0,56857417.0,56793708.0,56544709.0,57413875.0,57080625.0,56769375.0,56917750.0,56650583.0,56925250.0,56648250.0,56467458.0,56610042.0,57026834.0,56431084.0,56417917.0,56650125.0,56446834.0,56368417.0,56393042.0,56416125.0,56408834.0,57134500.0,56738541.0,56401500.0,56622209.0,56421750.0,56367125.0,57772875.0,56516375.0,56390083.0,56498416.0,56482542.0,56584875.0,56347750.0,56429000.0,56576916.0,56376417.0,56404333.0,56621000.0,56770750.0,57085917.0,56455417.0,56533416.0,56363209.0,56572416.0,56413542.0,56428959.0,56632500.0,56376875.0,56402750.0,56371208.0,56374875.0,56439000.0,56430292.0,56374625.0,56548250.0,56414458.0,56853666.0,56798375.0,56641875.0,56549083.0,56597125.0,56498292.0,56348959.0,56575625.0,56373875.0,56434666.0,56498916.0,56383125.0,56583292.0,56549375.0,56334458.0,56568250.0,56680417.0,56743209.0,56712917.0,56397625.0,57310000.0,56480708.0,56634000.0,56736458.0,56550542.0,56515375.0,56615416.0],"unit":"ns","throughput":[{"per_iteration":190,"unit":"elements"}],"typical":{"estimate":9437545.35,"lower_bound":9429535.613666669,"upper_bound":9446134.492124997,"unit":"ns"},"mean":{"estimate":9437545.35,"lower_bound":9429535.613666669,"upper_bound":9446134.492124997,"unit":"ns"},"median":{"estimate":9428388.833333332,"lower_bound":9416486.0,"upper_bound":9437451.333333332,"unit":"ns"},"median_abs_dev":{"estimate":38310.75396984677,"lower_bound":27268.47291588737,"upper_bound":46727.59757041885,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"benchmark-complete","id":"heavy_linear_data/heavy_linear_data_195","report_directory":"","iteration_count":[6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6],"measured_values":[57754250.0,57607958.0,57583167.0,57624083.0,57630666.0,57630958.0,57657750.0,57624833.0,57745500.0,57611916.0,57601375.0,57619083.0,58621000.0,61597792.0,58071666.0,58331541.0,58337875.0,57682500.0,57619084.0,58932792.0,57607667.0,57612708.0,57728416.0,57774166.0,57658458.0,57577250.0,57744958.0,57645584.0,57608833.0,57588250.0,57647458.0,58584625.0,58205416.0,57677500.0,57600875.0,57582000.0,57632250.0,57618084.0,57571292.0,57579125.0,57674875.0,57631750.0,57770041.0,57749208.0,57630792.0,57638334.0,57580083.0,57645666.0,57685041.0,58183291.0,58154167.0,57612000.0,57831709.0,57664584.0,57700917.0,57651875.0,57602709.0,57842208.0,58066125.0,57589375.0,57769083.0,57612500.0,57685875.0,57624875.0,57603625.0,57865833.0,57731084.0,57597417.0,58373709.0,57765375.0,57661500.0,57623667.0,57625625.0,57596625.0,57625541.0,57732333.0,57599292.0,57615542.0,57787917.0,59978791.0,57647333.0,57576458.0,58583667.0,58738709.0,57706167.0,57842750.0,57913833.0,58346959.0,57692792.0,57606833.0,57660958.0,57630708.0,57817166.0,57663583.0,57634500.0,57678333.0,57646125.0,57647083.0,57633833.0,57625959.0],"unit":"ns","throughput":[{"per_iteration":195,"unit":"elements"}],"typical":{"estimate":9638009.020000005,"lower_bound":9623148.483083338,"upper_bound":9656712.016041663,"unit":"ns"},"mean":{"estimate":9638009.020000005,"lower_bound":9623148.483083338,"upper_bound":9656712.016041663,"unit":"ns"},"median":{"estimate":9608277.75,"lower_bound":9605375.0,"upper_bound":9613333.333333332,"unit":"ns"},"median_abs_dev":{"estimate":11768.013741076453,"lower_bound":7855.429321788251,"upper_bound":18326.54182463838,"unit":"ns"},"slope":null,"change":null}
-        {"reason":"group-complete","group_name":"heavy_linear_data","benchmarks":["heavy_linear_data/heavy_linear_data_0","heavy_linear_data/heavy_linear_data_5","heavy_linear_data/heavy_linear_data_10","heavy_linear_data/heavy_linear_data_15","heavy_linear_data/heavy_linear_data_20","heavy_linear_data/heavy_linear_data_25","heavy_linear_data/heavy_linear_data_30","heavy_linear_data/heavy_linear_data_35","heavy_linear_data/heavy_linear_data_40","heavy_linear_data/heavy_linear_data_45","heavy_linear_data/heavy_linear_data_50","heavy_linear_data/heavy_linear_data_55","heavy_linear_data/heavy_linear_data_60","heavy_linear_data/heavy_linear_data_65","heavy_linear_data/heavy_linear_data_70","heavy_linear_data/heavy_linear_data_75","heavy_linear_data/heavy_linear_data_80","heavy_linear_data/heavy_linear_data_85","heavy_linear_data/heavy_linear_data_90","heavy_linear_data/heavy_linear_data_95","heavy_linear_data/heavy_linear_data_100","heavy_linear_data/heavy_linear_data_105","heavy_linear_data/heavy_linear_data_110","heavy_linear_data/heavy_linear_data_115","heavy_linear_data/heavy_linear_data_120","heavy_linear_data/heavy_linear_data_125","heavy_linear_data/heavy_linear_data_130","heavy_linear_data/heavy_linear_data_135","heavy_linear_data/heavy_linear_data_140","heavy_linear_data/heavy_linear_data_145","heavy_linear_data/heavy_linear_data_150","heavy_linear_data/heavy_linear_data_155","heavy_linear_data/heavy_linear_data_160","heavy_linear_data/heavy_linear_data_165","heavy_linear_data/heavy_linear_data_170","heavy_linear_data/heavy_linear_data_175","heavy_linear_data/heavy_linear_data_180","heavy_linear_data/heavy_linear_data_185","heavy_linear_data/heavy_linear_data_190","heavy_linear_data/heavy_linear_data_195"],"report_directory":""}
-        "#;
-
-        let mut state = State {
-            all: false,
-            baseline: "baseline/baseline".into(),
-            ids: Default::default(),
-            throughput: Default::default(),
-            groups: Default::default(),
-        };
-        for line in input.lines() {
-            extract_state(line, &mut state, false);
-        }
-
-        let (_, costs) = state.into_dependent_costs();
-        let linear_data_cost = costs
-            .0
-            .get("heavy_linear_data")
-            .expect("Expected linear_data cost");
-
-        // f(x) = 2x + 10
-        let expected_cost = DependentCost::HeavyOperation {
-            base: 10,
-            gas_per_unit: 2,
-        };
-        assert!(matches!(linear_data_cost, Cost::Dependent(c) if *c == expected_cost));
-    }
-}
