+++ benches/src/lib.rs
@@ -1,9 +1,5 @@
-pub mod collect;
 pub mod import;
-pub mod model;
-pub mod regression;
 
-pub use collect::collect;
 use fuel_core::database::vm_database::VmDatabase;
 pub use fuel_core::database::Database;
 use fuel_core_types::{
@@ -32,7 +28,6 @@ use fuel_core_types::{
         *,
     },
 };
-pub use model::*;
 
 pub use rand::Rng;
 use std::iter;
