--- benches/src/regression.rs
@@ -1,5 +1,0 @@
-mod linear;
-mod quadratic;
-
-pub use linear::linear_regression;
-pub use quadratic::quadratic_regression;
